% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_ll.R
\name{add_column_style}
\alias{add_column_style}
\title{Add inline CSS styles to columns in an HTML table object}
\usage{
add_column_style(html_tbl, columns = NULL, property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{columns}{an optional vector of column index
numbers or column names (available in the
\code{html_tbl}'s \code{column} and
\code{column_name} variables, respectively) that
are to be targeted for transformation. If nothing
is provided here then the style will be applied
to all columns.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Modify column styles for a table within an HTML
table object.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- gt(tbl = iris)

# With `add_column_style()` function,
# we can manually apply a CSS style to
# one or more columns
html_table_transformed <-
  html_table \%>\%
  add_column_style(
    columns = 1,
    property = "color",
    values = "blue")
}
