% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_multi.R
\name{add_gt_tbls}
\alias{add_gt_tbls}
\title{Add one or more \strong{gt} tables to a \code{gt_multi} container object}
\usage{
add_gt_tbls(.data, ..., .list = list2(...), .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A \code{gt_multi} container object, typically generated through use
of the \code{\link[=gt_multi]{gt_multi()}} function along with one or more \code{gt_tbl} objects.}

\item{...}{One or more \strong{gt} table (\code{gt_tbl}) objects, typically generated
via the \code{\link[=gt]{gt()}} function.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}

\item{.before, .after}{A single index for either \code{.before} or \code{.after},
specifying where the supplied \code{gt_tbl} objects should be placed amongst the
existing collection of \strong{gt} tables. If nothing is provided for either
argument the incoming \code{gt_tbl} objects will be appended.}
}
\value{
An object of class \code{gt_multi}.
}
\description{
Should you have a \code{gt_multi} object, created through use of the \code{\link[=gt_multi]{gt_multi()}}
function, you might want to add more \strong{gt} tables to that container. While
it's common to generate a \code{gt_multi} object with a collection of \code{gt_tbl}
objects, one can also create an 'empty' \code{gt_multi} object. Whatever your
workflow might be, the \code{add_gt_tbls()} function makes it possible to flexibly
add one or more new
\strong{gt} tables, returning a refreshed \code{gt_multi}
object.
}
\section{Function ID}{

14-3
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other multiple-table functions: 
\code{\link{gt_multi}()},
\code{\link{opts_gt_tbls}()},
\code{\link{pull_gt_tbls}()},
\code{\link{replace_gt_tbls}()},
\code{\link{rm_gt_tbls}()}
}
\concept{multiple-table functions}
