% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_ll.R
\name{add_header_style}
\alias{add_header_style}
\title{Add inline CSS styles to the header of an HTML table object}
\usage{
add_header_style(html_tbl, property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Modify the style of the header for a table
within an HTML table object. This is part of
an intermediate set of step in the
\code{gt()} -> ... -> \code{emit_html()} pattern.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- gt(tbl = iris)

# With `add_header_style()` function,
# we can manually apply a CSS style
# to the <thead> element
html_table_transformed <-
  html_table \%>\%
  add_header_style(
    property = "color",
    values = "blue")
}
