% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_source_note.R
\name{add_source_note}
\alias{add_source_note}
\title{Add a source note citation}
\usage{
add_source_note(html_tbl, source_note, lead_in = "Source: ")
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{build_html_table()}
function.}

\item{source_note}{text to be used in the
source note.}

\item{lead_in}{a lead in word or phrase
indicating the citation as the source of the
tabulated data.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a source note citation to the source note
part of the table.
}
