% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{add_stub}
\alias{add_stub}
\title{Add a stub to a table}
\usage{
add_stub(html_tbl, column = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{column}{an option to use an existing
column as the stub.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a stub to a table
}
