% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{add_stubhead_caption}
\alias{add_stubhead_caption}
\title{Add caption text to the stubhead}
\usage{
add_stubhead_caption(html_tbl, caption, alignment = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{caption}{the text to be used as the
stubhead caption.}

\item{alignment}{an option to modify the
alignment of the stubhead caption. By
default this is \code{left}. Other options
are \code{center} and \code{right}.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a caption to the stubhead of a table. If
a stub does not exist, no change will be made
but a message to this effect will be emitted.
}
