% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_columns.R
\name{add_style_to_columns}
\alias{add_style_to_columns}
\title{Add inline CSS styles to specified table columns}
\usage{
add_style_to_columns(html_tbl, columns = NULL, include_col_n = TRUE,
  property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{columns}{the columns that are to have the
style values added.}

\item{include_col_n}{an option for whether to
include the rightmost column when not providing
explicit column names for \code{columns}.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Modify the style of table columns.
}
