% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_parts.R
\name{add_style_to_part}
\alias{add_style_to_part}
\title{Add inline CSS styles to a table part}
\usage{
add_style_to_part(html_tbl, part, property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{part}{a table part which could be any
of \code{stub}, \code{boxhead}, or \code{field}.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Modify the style of a specific table part.
}
