% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_rows.R
\name{add_style_to_row_n}
\alias{add_style_to_row_n}
\title{Add inline CSS styles to `row n` (boxhead and stubhead)}
\usage{
add_style_to_row_n(html_tbl, property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add inline CSS style to the boxhead and stubhead
table parts that are part of `row n`.
}
