% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_rows.R
\name{add_style_to_rows}
\alias{add_style_to_rows}
\title{Add inline CSS styles to specified table rows}
\usage{
add_style_to_rows(html_tbl, rows = NULL, parts = c("stub", "field"),
  property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{rows}{the rows that are to have the
style values added.}

\item{parts}{the table parts where rows are to
have styles applied. Options are either or both
of \code{stub} and \code{field}.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Modify the style of specified table rows in
both the stub and field table parts.
}
