% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tbl_style_rows.R
\name{add_style_to_summary_caption_rows}
\alias{add_style_to_summary_caption_rows}
\title{Add inline CSS styles to \code{summary_caption} rows.}
\usage{
add_style_to_summary_caption_rows(html_tbl, extend_to_field = FALSE, property,
  values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{extend_to_field}{an option for applying
property values to adjacent cells in the field.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add inline CSS styles to \code{summary_caption} rows.
}
