% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{add_table}
\alias{add_table}
\title{Helper function for adding a data table}
\usage{
add_table(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
a character object with an HTML
fragment that can be placed inside the
message body wherever the table should
appear.
}
\description{
Add a data table inside the body of the
email with this helper function. Simply
provide a \code{data.frame} or tibble
object and the function will sensibly
style the table during insertion.
}
\details{
If more control over the table styling is
required, then an alternative is to use the
\code{build_html_table()} ->
\code{add_column_style()} -> \code{emit_html()}
pattern for building and inserting a table.
}
