% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_constructors.R
\name{add_tbl}
\alias{add_tbl}
\title{Build on a \code{tbl_gt} by combining with another table}
\usage{
add_tbl(
  x,
  y,
  below = NULL,
  above = NULL,
  by = NULL,
  right_of = NULL,
  left_of = NULL,
  mode = c("insert", "replace")
)
}
\description{
Build on a \code{tbl_gt} by combining with another table
}
