% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{adjust_luminance}
\alias{adjust_luminance}
\title{Adjust the luminance for a palette of colors}
\usage{
adjust_luminance(colors, steps)
}
\arguments{
\item{colors}{a vector of colors that will undergo an adjustment in
luminance.}

\item{steps}{a positive or negative factor by which the luminance will be
adjusted. Must be a number between \code{-2.0} and \code{2.0}.}
}
\description{
Adjust the luminance for a palette of colors
}
\seealso{
Other colorizing helper functions: \code{\link{color_bin}},
  \code{\link{color_factor}}, \code{\link{color_numeric}},
  \code{\link{color_quantile}}
}
\concept{colorizing helper functions}
