% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{apply_alignment_center}
\alias{apply_alignment_center}
\title{Apply center alignment to column data}
\usage{
apply_alignment_center(html_tbl, columns = NULL, types = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{an optional vector of column names
for which the center alignment should be applied.}

\item{types}{an optional vector of column types
for which the center alignment should be applied.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply the center alignment rule to data cells within
table columns. Options exist to selectively
apply the alignment to specified column names or
to columns of certain types.
}
