% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{apply_font}
\alias{apply_font}
\title{Apply fonts to column data}
\usage{
apply_font(html_tbl, columns = NULL, types = NULL, font)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{columns}{an optional vector of column names
for which the font should be applied.}

\item{types}{an optional vector of column types
for which the font should be applied.}

\item{font}{the name of the font to use for
the specified columns. This could be provided
as a vector of fonts where subsequent font names
provide fallbacks in the case that fonts are
not available.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply fonts to column data
}
