% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transform.R
\name{apply_spanner_headings}
\alias{apply_spanner_headings}
\alias{apply_spanner_headings}
\title{Apply spanner headings}
\usage{
apply_spanner_headings(html_tbl)

apply_spanner_headings(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}
}
\value{
an object of class \code{html_table}.

an object of class \code{html_table}.
}
\description{
Apply spanner headings

Apply spanner headings
}
