% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_styles}
\alias{apply_styles}
\title{Helper for defining custom styles for table cells}
\usage{
apply_styles(background_color = NULL, text_color = NULL, text_font = NULL,
  text_size = NULL, text_align = NULL, text_indent = NULL,
  text_style = NULL, text_decorate = NULL, text_transform = NULL)
}
\value{
a character vector containing formatted styles.
}
\description{
Helper for defining custom styles for table cells
}
\seealso{
Other helper functions: \code{\link{col_labels}},
  \code{\link{columns_with}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{not_in_group}},
  \code{\link{rownames_with}}, \code{\link{rows_where}},
  \code{\link{target_cells}}, \code{\link{target_cell}},
  \code{\link{vars}}
}
