% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{apply_theme_classical}
\alias{apply_theme_classical}
\title{Apply line styles that give a table a classic look}
\usage{
apply_theme_classical(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply horizontal and vertical line styles that
visually separate the columns and rows. This
adds reasonably thick horizontal and vertical
rules.
}
