% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{apply_theme_minimal}
\alias{apply_theme_minimal}
\title{Apply a minimal set of line styles to the table}
\usage{
apply_theme_minimal(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply line styles with a minimal set of
horizontal rules for demarcation of the
boxhead and the field. This adds reasonably
thick horizontal rules enclosing the boxhead
and stubhead and a similar horizontal rule
below the field and stub.
}
