% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{apply_theme_striped}
\alias{apply_theme_striped}
\title{Apply the striped theme to the table}
\usage{
apply_theme_striped(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply the striped theme to the table. This
adds striped rows, reasonably thick horizontal
rules enclosing the boxhead and stubhead and
a similar horizontal rule below the field and
stub.
}
