% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.gt_tbl}
\alias{as.data.frame.gt_tbl}
\title{Transform a \strong{gt} object to a data frame}
\usage{
\method{as.data.frame}{gt_tbl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gt_tbl}.}

\item{...}{Any additional parameters.}
}
\description{
This facilitates conversion of the \code{gt_tbl} object to a data frame.
}
\keyword{internal}
