% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.tags.gt_tbl}
\alias{as.tags.gt_tbl}
\title{Convert a gt table to ab htmltools tagList}
\usage{
\method{as.tags}{gt_tbl}(x, ...)
}
\arguments{
\item{x}{Object to be converted.}

\item{...}{Any additional parameters.}
}
\description{
This converts a gt table object to an \pkg{htmltools} \code{\link[htmltools:tagList]{htmltools::tagList()}}
object. The returned object is of the \code{shiny.tag.list} class and using
\code{as.character()} with that will render the HTML, resulting in a length 1
character vector that contains the HTML table.
}
