% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_gtable}
\alias{as_gtable}
\title{Transform a \strong{gt} table to a \code{gtable} object}
\usage{
as_gtable(data, plot = FALSE, text_grob = grid::textGrob)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{plot}{A \code{logical(1)} whether the gtable should be rendered on the
graphics device.}

\item{text_grob}{A \code{function} used to draw text. Defaults to
\code{grid::textGrob()}, but can be swapped to \code{gridtext::richtext_grob()} to
better render html content.}
}
\value{
A \code{gtable} object.
}
\description{
Take a \code{gt_tbl} table and transform it to a \code{gtable}.
}
\section{Function ID}{

13-6
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other table export functions: 
\code{\link{as_latex}()},
\code{\link{as_raw_html}()},
\code{\link{as_rtf}()},
\code{\link{as_word}()},
\code{\link{extract_body}()},
\code{\link{extract_cells}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{table export functions}
