% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{as_latex}
\alias{as_latex}
\title{Output a gt table object as LateX}
\usage{
as_latex(data)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}
}
\description{
Take a \code{gt_tbl} table object and emit LateX.
}
\examples{
\dontrun{
# Create a `gt` table using the `sleep`
# dataset and output as LaTeX
gt(data = sleep) \%>\%
  as_latex()
}
}
\seealso{
Other table export functions: \code{\link{as_rtf}},
  \code{\link{extract_summary}}
}
\concept{table export functions}
