% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_rtf}
\alias{as_rtf}
\title{Output a \strong{gt} object as RTF}
\usage{
as_rtf(
  data,
  page_numbering.active = FALSE,
  page_numbering.location = c("header", "footer")
)
}
\arguments{
\item{data}{A table object that is created using the \code{gt()} function.}

\item{page_numbering.active}{An option to include page numbering in the RTF
document. The location for page numbering text is either in the document
header or footer (governed by the \code{page_numbering.location} option). By
default, page numbering is not active (\code{FALSE}).}

\item{page_numbering.location}{If \code{page_numbering.active} is \code{TRUE} then this
option determines where the page numbering text will be placed. It will
either be in the RTF document \code{"header"} (default option) or the
\code{"footer"}.}
}
\description{
Get the RTF content from a \code{gt_tbl} object as as a single-element character
vector. This object can be used with \code{writeLines()} to generate a valid .rtf
file that can be opened by RTF readers.
}
\section{Function ID}{

13-4
}

\examples{
# Use `gtcars` to create a gt table;
# add a header and then export as
# RTF code
tab_rtf <-
  gtcars \%>\%
  dplyr::select(mfr, model) \%>\%
  dplyr::slice(1:2) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_rtf()

}
\seealso{
Other Export Functions: 
\code{\link{as_latex}()},
\code{\link{as_raw_html}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{Export Functions}
