% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_gt.R
\name{as_tbl_gt}
\alias{as_tbl_gt}
\title{Create a \code{tbl_gt} object from an existing table}
\usage{
as_tbl_gt(tbl, add_index_col = FALSE)
}
\arguments{
\item{tbl}{A table object, which could be of the \code{tbl_df} class or a data
frame.}
}
\description{
Create a \code{tbl_gt} object from an existing table
}
