% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_typst}
\alias{as_typst}
\title{Output a \strong{gt} object as Typst}
\usage{
as_typst(data)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}
}
\value{
An object of class \code{typst_text}.
}
\description{
Get the Typst content from a \code{gt_tbl} object as a single-element character
vector. This object can be used with \code{writeLines()} to generate a valid .typ
file, or it will be automatically rendered in Quarto documents targeting
Typst output.
}
\section{Function ID}{

13-10
}

\section{Function Introduced}{

\code{v0.13.0} (2026-03-26)
}

\seealso{
Other table export functions: 
\code{\link{as_gtable}()},
\code{\link{as_latex}()},
\code{\link{as_raw_html}()},
\code{\link{as_rtf}()},
\code{\link{as_word}()},
\code{\link{extract_body}()},
\code{\link{extract_cells}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{table export functions}
