% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_word}
\alias{as_word}
\title{Output a \strong{gt} object as Word}
\usage{
as_word(data)
}
\arguments{
\item{data}{A table object that is created using the \code{gt()} function.}
}
\description{
Get the Open Office XML table tag content from a \code{gt_tbl} object as as a
single-element character vector.
}
\section{Function ID}{

13-5
}

\examples{
# Use `gtcars` to create a gt table;
# add a header and then export as
# OOXML code for Word
tab_rtf <-
  gtcars \%>\%
  dplyr::select(mfr, model) \%>\%
  dplyr::slice(1:2) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_word()

}
\seealso{
Other Export Functions: 
\code{\link{as_latex}()},
\code{\link{as_raw_html}()},
\code{\link{as_rtf}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{Export Functions}
