% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scale.R
\name{bg_color_fill_manual}
\alias{bg_color_fill_manual}
\title{Set background data cell colors using provided colors}
\usage{
bg_color_fill_manual(data, column, values, colors, na_color = "darkgrey")
}
\description{
Set background data cell colors using provided colors
}
