% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorizing.R
\name{bg_color_gradient_n}
\alias{bg_color_gradient_n}
\title{Set background data cell colors using provided colors and breaks}
\usage{
bg_color_gradient_n(data, column, breaks, colors, alpha = 1,
  na_color = "darkgrey")
}
\description{
Set background data cell colors using provided colors and breaks
}
