% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_rows.R
\name{blocks_arrange}
\alias{blocks_arrange}
\title{Modify the ordering of the stub block groups}
\usage{
blocks_arrange(data, groups)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{groups}{a vector of group names, or, a numeric
vector of indices corresponding to the new ordering. Either vector must
correspond to assigned group names or the index positions. Also, either
vector isn't required to have all group names or index positions within it;
any omitted values will be added to the end while preserving the original
ordering.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Modify the ordering of the stub block groups
}
\concept{row modification functions}
