% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cells.R
\name{cell}
\alias{cell}
\title{Define a single table cell}
\usage{
cell(row, column)
}
\arguments{
\item{row}{an integer representing the
table row to target (as part of the
\code{row}-\code{column} dyad).}

\item{column}{an integer representing the
table column to target (as part of the
\code{row}-\code{column} dyad).}
}
\description{
Define a single table cell
}
