% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cell_fill}
\alias{cell_fill}
\title{Helper for defining custom fills for table cells}
\usage{
cell_fill(color = NULL, alpha = NULL)
}
\arguments{
\item{color}{The fill color. If nothing is provided, then \code{"#D3D3D3"} (light
gray) will be used as a default.}

\item{alpha}{The alpha transparency value for the \code{color} as single value in
the range of \code{0} (fully transparent) to \code{1} (fully opaque). If not provided
the fill color will be fully opaque.}
}
\description{
The \code{cell_fill()} helper function is to be used with the \code{\link[=tab_style]{tab_style()}}
function, which itself allows for the setting of custom styles to one or more
cells. Specifically, the call to \code{cell_fill()} should be bound to the
\code{styles} argument of \code{\link[=tab_style]{tab_style()}}.
}
\seealso{
Other helper functions: \code{\link{cell_borders}},
  \code{\link{cell_text}}, \code{\link{cells_styles}},
  \code{\link{escape_latex}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}},
  \code{\link{random_id}}
}
\concept{helper functions}
