% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cells.R
\name{cells}
\alias{cells}
\title{Define one or more table cells}
\usage{
cells(row, column)
}
\arguments{
\item{row}{an integer representing the
table row to target (as part of the
\code{row}-\code{column} dyad).}

\item{column}{an integer representing the
table column to target (as part of the
\code{row}-\code{column} dyad).}
}
\value{
a list object of class \code{helper_cells}
}
\description{
Define one or more table cells
}
