% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cells_data}
\alias{cells_data}
\title{Location helper for targeting data cells in the table body (deprecated)}
\usage{
cells_data(columns = TRUE, rows = TRUE)
}
\arguments{
\item{columns}{The names of the columns that are to be targeted.}

\item{rows}{The names of the rows that are to be targeted.}
}
\description{
Location helper for targeting data cells in the table body (deprecated)
}
