% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cells_group}
\alias{cells_group}
\title{Location helper for targeting row groups (deprecated)}
\usage{
cells_group(groups = everything())
}
\arguments{
\item{groups}{The names of the row groups that are to be targeted.}
}
\description{
Location helper for targeting row groups (deprecated)
}
\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\keyword{internal}
