% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{color_factor}
\alias{color_factor}
\title{Colorizing helper function that colors according to factor levels}
\usage{
color_factor(palette, domain = NULL, levels = NULL, ordered = FALSE,
  na_color = "#808080")
}
\arguments{
\item{palette}{the colors or color function that values will be mapped to.}

\item{domain}{the possible values that should be mapped.}

\item{levels}{provides a means to specify levels differently. If this is
provided then the \code{domain} argument is ignored.}

\item{ordered}{if set to \code{TRUE} and the \code{domain} needs to be
coerced to a factor, treat it as already in the correct order. The default
value for this is \code{FALSE}.}

\item{na_color}{the default color for cells where data isn't mapped.}
}
\value{
a function that maps cell data to colors in a palette (for use in
  \code{\link{cols_color_scale}()}).
}
\description{
Colorizing helper function that colors according to factor levels
}
\seealso{
Other colorizing helper functions: \code{\link{adjust_luminance}},
  \code{\link{color_bin}}, \code{\link{color_numeric}},
  \code{\link{color_quantile}}
}
\concept{colorizing helper functions}
