% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{color_numeric}
\alias{color_numeric}
\title{Colorizing helper function that linearly colors continuous data}
\usage{
color_numeric(palette, domain = NULL, na_color = "#808080")
}
\arguments{
\item{palette}{the colors or color function that values will be mapped to.}

\item{domain}{the possible values that should be mapped.}

\item{na_color}{the default color for cells where data isn't mapped.}
}
\description{
Colorizing helper function that linearly colors continuous data
}
