% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_add}
\alias{cols_add}
\title{Add one or more columns to a \strong{gt} table}
\usage{
cols_add(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} --- \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}

\item{...}{\emph{Cell data assignments}

\verb{<multiple expressions>} --- (or, use \code{.list})

Expressions for the assignment of cell values to the new columns.
Name-value pairs, in the form of \verb{<column> = <value vector>} will work, so
long as any \verb{<column>} value does not already exist in the table. The
\verb{<value vector>} may be an expression that uses one or more column names in
the table to generate a vector of values. Single values in \verb{<value vector>}
will be repeated down the new column. A vector where the length is exactly
the number of rows in the table can also be used.}

\item{.before, .after}{\emph{Column used as anchor}

\verb{<column-targeting expression>} --- \emph{default:} \code{NULL} (\code{optional})

A single column-resolving expression or column index an be given to either
\code{.before} or \code{.after}. The column specifies where the new columns should be
positioned among the existing columns in the input data table. While select
helper functions such as \code{\link[=starts_with]{starts_with()}} and \code{\link[=ends_with]{ends_with()}} can be used for
column targeting, it's recommended that a single column name or index be
used. This is to ensure that exactly one column is provided to either of
these arguments (otherwise, the function will be stopped). If nothing is
provided for either argument then any new column will be placed at the end
of the column series.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
We can add new columns to a table with the \code{cols_add()} function and it works
quite a bit like the \strong{dplyr} \code{mutate()} function. The idea is that you
supply name-value pairs where the name is the new column name and the value
part describes the data that will go into the column. The latter can:
(1) be a vector where the length of the number of rows in the data table,
(2) be a single value (which will be repeated all the way down), or (3)
involve other columns in the table (as they represent vectors of the correct
length). The new columns are added to the end of the column series by default
but can instead be added internally by using either the \code{.before} or \code{.after}
arguments. If entirely empty (i.e., all \code{NA}) columns need to be added, you
can use any of the \code{NA} types (e.g., \code{NA}, \code{NA_character_}, \code{NA_real_}, etc.)
for such columns.
}
\section{Targeting the column for insertion with \code{.before} or \code{.after}}{


The targeting of a column for insertion is done through the \code{.before} or
\code{.after} arguments (only one of these options should be be used). While
\strong{tidyselect}-style expressions or indices can used to target a column, it's
advised that a single column name be used. This is to avoid the possibility
of inadvertently resolving multiple columns (since the requirement is for a
single column).
}

\section{Examples}{


Let's take a subset of the \code{\link{exibble}} dataset and make a simple \strong{gt} table
with it (using the \code{row} column for labels in the stub). We'll add a single
column to the right of all the existing columns and call it \code{country}. This
new column needs eight values and these will be supplied when using
\code{cols_add()}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  dplyr::select(num, char, datetime, currency, group) |>
  gt(rowname_col = "row") |>
  cols_add(
    country = c("TL", "PY", "GL", "PA", "MO", "EE", "CO", "AU")
  )
}\if{html}{\out{</div>}}
}

\section{Function ID}{

5-7
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other column modification functions: 
\code{\link{cols_align_decimal}()},
\code{\link{cols_align}()},
\code{\link{cols_hide}()},
\code{\link{cols_label_with}()},
\code{\link{cols_label}()},
\code{\link{cols_merge_n_pct}()},
\code{\link{cols_merge_range}()},
\code{\link{cols_merge_uncert}()},
\code{\link{cols_merge}()},
\code{\link{cols_move_to_end}()},
\code{\link{cols_move_to_start}()},
\code{\link{cols_move}()},
\code{\link{cols_unhide}()},
\code{\link{cols_units}()},
\code{\link{cols_width}()}
}
\concept{column modification functions}
