% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align}
\alias{cols_align}
\title{Set the alignment of columns}
\usage{
cols_align(data, align = c("auto", "left", "center", "right"),
  columns = TRUE)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{align}{the alignment type. This can be any of \code{"center"},
\code{"left"}, or \code{"right"} for center-, left-, or center-alignment.
Alternatively, the \code{"auto"} option (the default), will automatically
align values in columns according to the data type (i.e., column class). In
this way, left-alignment is applied to columns of class \code{character},
\code{Date}, or \code{POSIXct}; center-alignment is for columns of class
\code{logical}, \code{factor}, or \code{list}; and right-aligned is for the
\code{numeric} and \code{integer} columns.}

\item{columns}{an optional vector of column names for which the alignment
should be applied. If nothing is supplied, or if \code{columns} is
\code{TRUE}), then the alignment affects all columns.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set the alignment of columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and align the `mpg`
# column to the left
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    cols_align(
      align = "left",
      columns = vars(mpg))
}
\seealso{
Other column modification functions: \code{\link{cols_hide}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}},
  \code{\link{data_color}}
}
\concept{column modification functions}
