% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align_center}
\alias{cols_align_center}
\title{Set columns to be aligned to the center}
\usage{
cols_align_center(data, columns = NULL, types = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an optional vector of column names
for which the center alignment should be applied.}

\item{types}{an optional vector of column types
for which the center alignment should be applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set columns to be aligned to the center
}
\examples{
# Create a table object using the
# `mtcars` dataset and center align
# all columns
gt(data = mtcars \%>\% tibble::rownames_to_column()) \%>\%
  cols_align_center()
}
