% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{cols_color_gradient_n}
\alias{cols_color_gradient_n}
\title{Set data cell colors as a gradient with provided colors and breaks}
\usage{
cols_color_gradient_n(data, column, colors, breaks, alpha = 1,
  na_color = "darkgrey")
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{colors}{a vector of colors interpolate between according to the value
provided in \code{breaks}. Each color value provided must either be a color
name (in the set of colors provided by \code{grDevices::colors()}) or
hexadecimal strings in the form of "#RRGGBB" or "#RRGGBBAA".}

\item{breaks}{numeric breaks that represent the transition points between the
colors provided in \code{colors}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set data cell colors as a gradient with provided colors and breaks
}
