% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{cols_color_scale}
\alias{cols_color_scale}
\title{Set data cell colors using a palette or a colorizing helper function}
\usage{
cols_color_scale(data, columns, colors, alpha = NULL,
  apply_to = "bkgd", autocolor_text = TRUE)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns wherein changes to cell data colors should occur.}

\item{colors}{either a colorizing helper function or a vector of colors to
use for each distinct value in each of the provided \code{columns}. The
colorizing helper functions are: \code{\link{color_quantile}()},
\code{\link{color_bin}()}, \code{\link{color_numeric}()}, and
\code{\link{color_factor}()}. Details for the helper functions are provided
in their respective help files. If providing a vector of colors as a
palette, each color value provided must either be a color name (in the set
of colors provided by \code{grDevices::colors()}) or hexadecimal strings in
the form of "#RRGGBB" or "#RRGGBBAA".}

\item{alpha}{an optional, fixed alpha transparency value that will be applied
to all of the \code{colors} provided if they are provided as a vector of
colors. If using a colorizing helper function for \code{colors} then this
option is ignored (each of the colorizing helper functions has its own
\code{alpha} argument).}

\item{apply_to}{which style element should the colors be applied to? Options
include the cell background (the default, given as \code{bkgd}) or the cell
text (\code{text}).}

\item{autocolor_text}{an option to let gt modify the coloring of text within
cells undergoing background coloring. This will in some cases yield more
optimal text-to-background color contrast. By default, this is set to
\code{TRUE}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set data cell colors using a palette or a colorizing helper function
}
\seealso{
Other column modification functions: \code{\link{cols_align}},
  \code{\link{cols_color_gradient_n}},
  \code{\link{cols_color_manual}}, \code{\link{cols_hide}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
