% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_hide}
\alias{cols_hide}
\title{Hide one or more columns}
\usage{
cols_hide(data, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the column names to hide from the output display table. The
order of the remaining columns will be preserved. Values provided that do
not correspond to column names will be disregarded.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Hide one or more columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and hide the `mpg`,
# `cyl`, and `disp` columns
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    cols_hide(
      columns = vars(mpg, cyl, disp))
}
\seealso{
Other column modification functions: \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}},
  \code{\link{data_color}}
}
\concept{column modification functions}
