% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_label_with}
\alias{cols_label_with}
\title{Relabel columns with a function}
\usage{
cols_label_with(.data, fn, columns = everything(), ...)
}
\arguments{
\item{.data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{fn}{The function to apply to existing column labels.}

\item{columns}{The column names to apply \code{fn} to. Applies to all columns by default.}

\item{...}{arguments to pass to \code{fn}}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Relabel columns with a function
}
\section{Examples}{


Use \code{\link{countrypops}} to create a \strong{gt} table. Relabel all the table's columns
with the \code{cols_label_with()} function to improve its presentation. It applies
to existing labels, rather than just names, so we can make more difficult
changes directly.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{countrypops \%>\%
  dplyr::select(-contains("code")) \%>\%
  dplyr::filter(country_name == "Mongolia") \%>\%
  tail(5) \%>\%
  gt() \%>\%
  cols_label(
    country_name = "Name"
  ) \%>\%
  cols_label_with(
    tools::toTitleCase
  )
}\if{html}{\out{</div>}}
}

\section{Function ID}{

4-14
}

\seealso{
Other column modification functions: 
\code{\link{cols_align_decimal}()},
\code{\link{cols_align}()},
\code{\link{cols_hide}()},
\code{\link{cols_label}()},
\code{\link{cols_merge_n_pct}()},
\code{\link{cols_merge_range}()},
\code{\link{cols_merge_uncert}()},
\code{\link{cols_merge}()},
\code{\link{cols_move_to_end}()},
\code{\link{cols_move_to_start}()},
\code{\link{cols_move}()},
\code{\link{cols_unhide}()},
\code{\link{cols_width}()}
}
\concept{column modification functions}
