% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{cols_merge}
\alias{cols_merge}
\title{Merge two columns to a single column with a formatter}
\usage{
cols_merge(data, col_1, col_2, format = "{1} {2}")
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{col_1}{a column that contains values for the start of the range.}

\item{col_2}{a column that contains values for the end of the range.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Merge two columns to a single column with a formatter
}
