% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_merge_n_pct}
\alias{cols_merge_n_pct}
\title{Merge two columns to combine counts and percentages}
\usage{
cols_merge_n_pct(data, col_n, col_pct, autohide = TRUE)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{col_n}{A column that contains values for the count component.}

\item{col_pct}{A column that contains fractional values for the percentage
component.}

\item{autohide}{An option to automatically hide the column specified as
\code{col_pct}. Any columns with their state changed to hidden will behave
the same as before, they just won't be displayed in the finalized table.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{cols_merge_n_pct()} function is a specialized variant of the
\code{\link[=cols_merge]{cols_merge()}} function. It operates by taking two columns that constitute
both a count (\code{col_n}) and a fraction of the total population (\code{col_pct}) and
merges them into a single column. What results is a column containing both
counts and their associated percentages (e.g., \verb{12 (23.2\%)}). The column
specified in \code{col_pct} is dropped from the output table.
}
\details{
This function could be somewhat replicated using \code{\link[=cols_merge]{cols_merge()}}, however,
\code{cols_merge_n_pct()} employs the following specialized semantics for \code{NA}
and zero-value handling:

\enumerate{
\item \code{NA}s in \code{col_n} result in missing values for the merged
column (e.g., \code{NA} + \verb{10.2\%} = \code{NA})
\item \code{NA}s in \code{col_pct} (but not \code{col_n}) result in
base values only for the merged column (e.g., \code{13} + \code{NA} = \code{13})
\item \code{NA}s both \code{col_n} and \code{col_pct} result in
missing values for the merged column (e.g., \code{NA} + \code{NA} = \code{NA})
\item If a zero (\code{0}) value is in \code{col_n} then the formatted output will be
\code{"0"} (i.e., no percentage will be shown)
}

Any resulting \code{NA} values in the \code{col_n} column following the merge
operation can be easily formatted using the \code{\link[=fmt_missing]{fmt_missing()}} function.
Separate calls of \code{\link[=fmt_missing]{fmt_missing()}} can be used for the \code{col_n} and
\code{col_pct} columns for finer control of the replacement values. It is the
responsibility of the user to ensure that values are correct in both the
\code{col_n} and \code{col_pct} columns (this function neither generates nor
recalculates values in either). Formatting of each column can be done
independently in separate \code{\link[=fmt_number]{fmt_number()}} and \code{\link[=fmt_percent]{fmt_percent()}} calls.

This function is part of a set of four column-merging functions. The other
two are the general \code{\link[=cols_merge]{cols_merge()}} function and the specialized
\code{\link[=cols_merge_uncert]{cols_merge_uncert()}} and \code{\link[=cols_merge_range]{cols_merge_range()}} functions. These functions
operate similarly, where the non-target columns can be optionally hidden from
the output table through the \code{hide_columns} or \code{autohide} options.
}
\section{Function ID}{

4-10
}

\seealso{
Other Modify Columns: 
\code{\link{cols_align}()},
\code{\link{cols_hide}()},
\code{\link{cols_label}()},
\code{\link{cols_merge_range}()},
\code{\link{cols_merge_uncert}()},
\code{\link{cols_merge}()},
\code{\link{cols_move_to_end}()},
\code{\link{cols_move_to_start}()},
\code{\link{cols_move}()},
\code{\link{cols_width}()}
}
\concept{Modify Columns}
