% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{cols_merge_uncert}
\alias{cols_merge_uncert}
\title{Merge two columns to a value + uncertainty column}
\usage{
cols_merge_uncert(data, col_val, col_uncert)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{col_val}{a single column name that contains the base values.}

\item{col_uncert}{a single column name that contains the uncertainty values.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Merge two columns to a value + uncertainty column
}
