% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{cols_move}
\alias{cols_move}
\title{Move one or more columns}
\usage{
cols_move(html_tbl, columns, after)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{the column names to move to
as a group to a different position. The
order of the remaining columns will be
preserved. Values provided that do not
correspond to column names will be
disregarded.}

\item{after}{a column name used to anchor the
insertion of the moved columns. All of the
moved columns will be placed to the right of
this column.}
}
\value{
an object of class \code{html_table}.
}
\description{
Move one or more columns
}
