% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{cols_rename}
\alias{cols_rename}
\title{Rename one or more columns}
\usage{
cols_rename(html_tbl, ...)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{...}{a series of named vectors for
specifying the mappings between new column
names and existing column names.}
}
\value{
an object of class \code{html_table}.
}
\description{
Rename one or more columns
}
\examples{
# Create a table based on `airquality` where
# some of the columns are renamed
tab_create(tbl = airquality) \%>\%
  theme_striped() \%>\%
  cols_rename(
    "Solar Radiation" = "Solar.R",
    "Wind (mph)" = "Wind",
    "Ozone (ppbV)" = "Ozone",
    "T (degrees F)" = "Temp"
    )
}
