% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{constants}
\alias{constants}
\title{Physical constants}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 354 rows and 4 columns.
}
\usage{
constants
}
\description{
A dataset with physical constants.
}
\section{Examples}{


Here is a glimpse at the data available in \code{constants}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dplyr::glimpse(constants)
#> Rows: 354
#> Columns: 4
#> $ name   <chr> "alpha particle-electron mass ratio", "alpha particle mass", "a~
#> $ value  <dbl> 7.294300e+03, 6.644657e-27, 5.971920e-10, 3.727379e+03, 4.00150~
#> $ uncert <dbl> 2.4e-07, 2.0e-36, 1.8e-19, 1.1e-06, 6.3e-11, 1.2e-12, 2.2e-10, ~
#> $ units  <chr> NA, "kg", "J", "MeV", "u", "kg mol^-1", NA, NA, "m", "kg", "J",~
}\if{html}{\out{</div>}}
}

\section{Dataset ID and Badge}{

DATA-9
}

\section{Dataset Introduced}{

\emph{In Development}
}

\seealso{
Other datasets: 
\code{\link{countrypops}},
\code{\link{exibble}},
\code{\link{gtcars}},
\code{\link{illness}},
\code{\link{metro}},
\code{\link{pizzaplace}},
\code{\link{rx_addv}},
\code{\link{rx_adsl}},
\code{\link{sp500}},
\code{\link{sza}},
\code{\link{towny}}
}
\concept{datasets}
\keyword{datasets}
