% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{countrypops}
\alias{countrypops}
\title{Yearly populations of countries from 1960 to 2017 (Wide Format)}
\format{A tibble with 12470 rows and 5 variables:
\describe{
  \item{country_name}{Name of the country}
  \item{country_code_2}{The 2-letter ISO 3166-1 country code}
  \item{country_code_3}{The 3-letter ISO 3166-1 country code}
  \item{year}{the year for the population estimate}
  \item{population}{the population estimate, midway through the year}
}}
\source{
\url{https://data.worldbank.org/indicator/SP.POP.TOTL}
}
\usage{
countrypops
}
\description{
A dataset that presents yearly populations of countries. The
country identifiers include the English country names, and the 2- and
3-letter ISO 3166-1 country codes. Each row contains a \code{population}
value for a given \code{year} (from 1960 to 2017). Any \code{NA} values
for \code{populations} indicate the non-existence of the country during
that year.
}
\keyword{datasets}
