% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{create_html}
\alias{create_html}
\title{Transform an HTML table object to an HTML fragment}
\usage{
create_html(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}
}
\value{
a character object with an HTML fragment.
}
\description{
Take a suitably styled HTML table object and
transform it to an HTML fragment. This is the
final step in the \code{gt()} -> ... ->
\code{emit_html()} pattern.
}
