% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_lines.R
\name{create_summary_lines}
\alias{create_summary_lines}
\title{Create lines with summarized data}
\usage{
create_summary_lines(html_tbl, group = NULL, columns, fcns, summary_labels)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{group}{the group to use when summarizing.}

\item{columns}{the columns to summarize.}

\item{fcns}{the names of functions to apply per
group per column.}

\item{summary_labels}{the labels to use in the stub
for the created summary rows.}
}
\description{
Create lines with summarized data
}
