% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{css_file}
\alias{css_file}
\title{Specify CSS rule sets using a CSS file}
\usage{
css_file(file)
}
\arguments{
\item{file}{Path to a CSS stylesheet.}
}
\description{
Specify CSS rule sets using a CSS file
}
\section{Function ID}{

7-18
}

\seealso{
Other Helper Functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{css_dec}()},
\code{\link{css_list}()},
\code{\link{css_sel}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()}
}
\concept{Helper Functions}
