% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{currency}
\alias{currency}
\title{Supply a context-aware currency symbol}
\usage{
currency(..., .list = list2(...))
}
\arguments{
\item{...}{One or more named arguments using output contexts as the names and
currency symbol text as the values.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}
}
\description{
Supply a context-aware currency symbol
}
