% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{darken_palette}
\alias{darken_palette}
\title{Darken a palette of colors by a fixed factor}
\usage{
darken_palette(colors, factor = 1.5)
}
\arguments{
\item{colors}{a vector of colors that is to be darkened.}

\item{factor}{the factor to which the colors will be darkened.}
}
\description{
Darken a palette of colors by a fixed factor
}
