% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{data_cells}
\alias{data_cells}
\title{Helper for targeting multiple data cells}
\usage{
data_cells(columns = NULL, rows = NULL)
}
\arguments{
\item{columns, rows}{either a vector of column/row names, a vector of
column/rows indices, column names provided in \code{\link{vars}()}, row
captions provided \code{c()}, or a helper function focused on selections.
The select helper functions are: \code{\link{starts_with}()},
\code{\link{ends_with}()}, \code{\link{contains}()},
\code{\link{matches}()}, \code{\link{one_of}()}, and
\code{\link{everything}()}.}
}
\value{
a list object of class \code{data_cells}.
}
\description{
This function is to be used to target individual data cells through the
  intersections of columns and rows. Thus, we provide targeting directives
  separately for columns and for rows.
}
