% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{define_units}
\alias{define_units}
\title{Define measurement units with \strong{gt}'s units notation}
\usage{
define_units(units_notation)
}
\arguments{
\item{units_notation}{\emph{Text in specialized units notation}

\verb{scalar<character>} --- \strong{required}

A single string that defines the units (e.g., \code{"m/s"}) to be used.}
}
\value{
An object of class \code{units_definition}.
}
\description{
The \code{define_units()} function is available for working with text in \strong{gt}'s
units notation.
}
\section{Function ID}{

8-7
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_footnotes}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_source_notes}()},
\code{\link{cells_stub_grand_summary}()},
\code{\link{cells_stub_summary}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()},
\code{\link{system_fonts}()}
}
\concept{helper functions}
