% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{emit_html}
\alias{emit_html}
\title{Transform a HTML table object to an HTML fragment}
\usage{
emit_html(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{build_html_table()}
function.}
}
\value{
a character object with an HTML fragment.
}
\description{
Take a suitably styled HTML table object and
transform it to an HTML fragment. This is the
final step in the \code{build_html_table()} ->
\code{add_..._style()} -> \code{emit_html()}
pattern.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- build_html_table(iris)

# The resulting object can be used
# in transformations
html_table_transformed <-
  html_table \%>\%
  add_column_style(
    columns = 1,
    property = "color",
    values = "blue")

# We can extract the HTML from the
# HTML table object using `emit_html()`
html_table_transformed \%>\%
  emit_html()
}
