% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_latex.R
\name{enlandscape_latex}
\alias{enlandscape_latex}
\title{Place LaTeX table in landscape orientation}
\usage{
enlandscape_latex(tab_latex)
}
\arguments{
\item{tab_latex}{The LaTeX table as a length-one character vector.}
}
\description{
Place LaTeX table in landscape orientation (requires pdflscape package)
}
