% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_latex.R
\name{entable_latex}
\alias{entable_latex}
\title{Place LaTeX table in a table environment}
\usage{
entable_latex(tab_latex)
}
\arguments{
\item{tab_latex}{The table as a length-one LaTeX character string.}
}
\description{
Place LaTeX table in a table environment by enclosing it in
\code{\\begin\{table\}} and \code{\\end\{table\}} tags.
}
