% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{fa_icon}
\alias{fa_icon}
\title{Helper function for adding a FontAwesome icon}
\usage{
fa_icon(name, height = 30, fill = NULL)
}
\arguments{
\item{name}{the name of the FontAwesome icon.}

\item{height}{the absolute height (px) of the icon in the table cell.}

\item{fill}{an option to change the fill color of the icon.}
}
\value{
a character object with an HTML fragment that can be placed inside of
a cell.
}
\description{
Add a FontAwesome icon inside a table with this helper function.
}
\examples{
# Create an HTML fragment that
# contains an FA icon
fa_icon_html <-
  fa_icon(
    name = "r-project")
}
