% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt}
\alias{fmt}
\title{Set a column format with a formatter function}
\usage{
fmt(data, columns, rows = TRUE, fcns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()} function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{fcns}{a single formatting function or a named list of functions.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a column format with a formatter function
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_missing}}, \code{\link{fmt_number}},
  \code{\link{fmt_passthrough}}, \code{\link{fmt_percent}},
  \code{\link{fmt_scientific}}, \code{\link{fmt_time}}
}
\concept{data formatting functions}
