% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_duration}
\alias{fmt_duration}
\title{Format values as time durations}
\usage{
fmt_duration(
  data,
  columns,
  rows = everything(),
  input_values = c("days", "hours", "minutes", "seconds"),
  output_units = "auto",
  duration_style = 1,
  pattern = "{x}"
)
}
\description{
Format input values to time durations whether those input values are numbers
or of the \code{difftime} class. We can specify which time units any numeric input
values have (as days, hours, minutes, or seconds) and the output can be
customized with a duration style (\code{1}-\code{3}, representing short, medium, and
long forms) and a choice of output units from weeks to milliseconds.
}
\section{Function ID}{

3-11
}

\seealso{
Other Format Data: 
\code{\link{data_color}()},
\code{\link{fmt_bytes}()},
\code{\link{fmt_currency}()},
\code{\link{fmt_datetime}()},
\code{\link{fmt_date}()},
\code{\link{fmt_engineering}()},
\code{\link{fmt_fraction}()},
\code{\link{fmt_integer}()},
\code{\link{fmt_markdown}()},
\code{\link{fmt_missing}()},
\code{\link{fmt_number}()},
\code{\link{fmt_passthrough}()},
\code{\link{fmt_percent}()},
\code{\link{fmt_scientific}()},
\code{\link{fmt_time}()},
\code{\link{fmt}()},
\code{\link{text_transform}()}
}
\concept{Format Data}
