% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{fmt_html}
\alias{fmt_html}
\title{Format to use HTML}
\usage{
fmt_html(html_tbl, columns = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{a vector of column names
for which the HTML formatting should be applied.}
}
\value{
an object of class \code{html_table}.
}
\description{
Format to use HTML
}
