% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_missing}
\alias{fmt_missing}
\title{Add a standardized mark for missing data}
\usage{
fmt_missing(data, columns, rows = NULL, missing_text = "---")
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{missing_text}{the text to be used in place of \code{NA} values in the
rendered table.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Wherever there is missing data (i.e., \code{NA} values) a customizable mark
may present better than the standard `NA` text that would otherwise appear.
}
