% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_per}
\alias{fmt_per}
\title{Format values as parts per some fixed quantity}
\usage{
fmt_per(
  data,
  columns,
  rows = everything(),
  to_units = c("ppm", "ppb", "ppt", "ppq", "per-mille", "per-myriad"),
  symbol = "auto",
  decimals = 2,
  drop_trailing_zeros = FALSE,
  drop_trailing_dec_mark = TRUE,
  scale_values = TRUE,
  use_seps = TRUE,
  pattern = "{x}",
  sep_mark = ",",
  dec_mark = ".",
  force_sign = FALSE,
  incl_space = "auto",
  placement = "right",
  locale = NULL
)
}
\description{
With numeric values in a \strong{gt} table we can format the values so that they
are rendered as ppm, ppb, etc., quantities.
}
\section{Function ID}{

3-6
}

\seealso{
Other Format Data: 
\code{\link{data_color}()},
\code{\link{fmt_bytes}()},
\code{\link{fmt_currency}()},
\code{\link{fmt_datetime}()},
\code{\link{fmt_date}()},
\code{\link{fmt_engineering}()},
\code{\link{fmt_fraction}()},
\code{\link{fmt_integer}()},
\code{\link{fmt_markdown}()},
\code{\link{fmt_missing}()},
\code{\link{fmt_number}()},
\code{\link{fmt_passthrough}()},
\code{\link{fmt_percent}()},
\code{\link{fmt_scientific}()},
\code{\link{fmt_time}()},
\code{\link{fmt}()},
\code{\link{text_transform}()}
}
\concept{Format Data}
