% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{fmt_range}
\alias{fmt_range}
\title{Combine two columns to a single, value range column}
\usage{
fmt_range(data, col_begin, col_end)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{col_begin}{a column that contains values for the start of the range.}

\item{col_end}{a column that contains values for the end of the range.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Combine two columns to a single, value range column
}
