% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{fmt_uncertainty}
\alias{fmt_uncertainty}
\title{Combine columns to a single value + uncertainty column}
\usage{
fmt_uncertainty(data, col_val, col_uncert)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{col_val}{a single column name that contains the base
values.}

\item{col_uncert}{a single column name that contains the
uncertainty values.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Combine columns to a single value + uncertainty column
}
