% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_options.R
\name{footnote_marks}
\alias{footnote_marks}
\title{Footnote marks composed of symbols}
\usage{
footnote_marks(type)
}
\arguments{
\item{type}{An option for a preset type of footnote marks. This can be a
traditional symbol set with a \code{"standard"} set (with four symbols), or, the
\code{"extended"} set (adds two more symbols, making six). We can use lowercase
\code{"letters"} or uppercase \code{"LETTERS"}. Finally, there is an option for
using \code{"numbers"}.}
}
\description{
Using \code{footnote_marks()} allows us to quickly define a set of footnote marks
for any footnotes that may be in the table. The use of this function pairs
well with the \code{\link[=opt_footnote_marks]{opt_footnote_marks()}} function, used for customizing the
footnote marks (and provides easier access to the \code{footnote.marks} argument
of \code{\link[=tab_options]{tab_options()}}).
}
\seealso{
The \code{\link[=opt_footnote_marks]{opt_footnote_marks()}} function for easily setting footnote
marks.
}
