% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_options.R
\name{footnote_symbols}
\alias{footnote_symbols}
\title{Footnote marks composed of symbols}
\usage{
footnote_symbols(type = c("standard", "extended"))
}
\arguments{
\item{type}{An option for whether the symbol set should be the \code{"standard"}
set (four symbols), or, the \code{"extended"} set (adds two more symbols, making
six).}
}
\description{
Using \code{footnote_symbols()} allows us to quickly define a set of footnote
marks for any footnotes that may be in the table. The use of this function
pairs well with the \code{\link[=tab_footnote_marks]{tab_footnote_marks()}} function, used for customizing the
footnote marks (and provides easier access to the \code{footnote.glyph} argument
of \code{\link[=tab_options]{tab_options()}}).
}
