% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formats.R
\name{format_as_currency}
\alias{format_as_currency}
\title{Format values as currency values}
\usage{
format_as_currency(html_tbl, columns = NULL, currency, use_subunits = TRUE,
  decimals = NULL, use_big_mark = FALSE, negative_style = "signed")
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{columns}{an option specify which columns
are to be formatted.}

\item{currency}{the currency to use for the
numeric value.}

\item{use_subunits}{an option for whether the
subunits portion of a currency value should be
displayed.}

\item{decimals}{an option to specify exactly
the number of decimal places to use.}

\item{use_big_mark}{a logical value that
allows for inclusion of a number grouping
separator for larger numbers.}

\item{negative_style}{the formatting to use
for negative numbers. With \code{signed},
negative numbers will be shown with a negative
sign. Using \code{parens} will show the
negative value in parentheses. The \code{red}
option will display the number in red.
Finally, \code{parens-red} will display
negative numbers as red and enclosed in
parentheses.}
}
\value{
an object of class \code{html_table}.
}
\description{
Format values as currency values
}
