% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formats.R
\name{format_as_date}
\alias{format_as_date}
\title{Format values as dates}
\usage{
format_as_date(html_tbl, columns = NULL, date_style = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{an option specify which columns
are to be formatted.}

\item{date_style}{the formatting style to use
for dates.}
}
\value{
an object of class \code{html_table}.
}
\description{
Format values as dates
}
