% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formats.R
\name{format_as_time}
\alias{format_as_time}
\title{Format values as times}
\usage{
format_as_time(html_tbl, columns = NULL, time_style = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{columns}{an option specify which columns
are to be formatted.}

\item{time_style}{the formatting style to use
for times.}
}
\value{
an object of class \code{html_table}.
}
\description{
Format values as times
}
