% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_pattern}
\alias{format_pattern}
\title{Supply a custom format pattern as a contextual list}
\usage{
format_pattern(..., .list = list2(...))
}
\arguments{
\item{...}{One or more named arguments using output contexts as the names and
formatting patterns as the values.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}
}
\value{
A list object of class \code{gt_format_pattern}.
}
\description{
Supply a custom format pattern as a contextual list
}
