% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{from_column}
\alias{from_column}
\title{Reference a column of values for formatting purposes}
\usage{
from_column(column, na_value = NULL, fn = NULL)
}
\arguments{
\item{column}{\emph{Column name}

\verb{scalar<character>} // \strong{required}

A single column name in quotation marks. Values will be extracted from this
column and provided to compatible arguments.}

\item{na_value}{\emph{Default replacement for \code{NA} values}

\verb{scalar<character|numeric|logical>} // \emph{default:} \code{NULL} (\code{optional})

A single value to replace any \code{NA} values in the \code{column}. Take care to
provide a value that is of the same type as the \code{column} values to avoid
any undesirable coercion.}

\item{fn}{\emph{Function to apply}

\verb{function|formula} // \emph{default:} \code{NULL} (\code{optional})

If a function is provided here, any values extracted from the table
\code{column} (except \code{NA} values) can be mutated.}
}
\value{
A list object of class \code{gt_column}.
}
\description{
For some functions, it is useful to get parameter values from a column in the
\strong{gt} table.
}
\section{Function ID}{

8-5
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_footnotes}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_source_notes}()},
\code{\link{cells_stub_grand_summary}()},
\code{\link{cells_stub_summary}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{define_units}()},
\code{\link{escape_latex}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()},
\code{\link{system_fonts}()}
}
\concept{helper functions}
