% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{funs}
\alias{funs}
\title{Helper for creating a list of functions}
\usage{
funs(...)
}
\arguments{
\item{...}{one or more functions that can be either a function name (e.g.,
\code{"mean"}), the function itself (e.g., \code{mean}), or calls to
functions with \code{.} as a dummy argument (e.g.,
\code{mean(., na.rm = TRUE)}).}
}
\description{
See \code{dplyr::\link[dplyr]{funs}} for details.
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}},
  \code{\link{vars}}
}
\concept{helper functions}
