% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_source_notes.R
\name{generate_source_notes}
\alias{generate_source_notes}
\title{Generate a HTML fragment for the source note part}
\usage{
generate_source_notes(source_note_tbl, span_amount, font_size = "115\%",
  align = "center", padding = "3px")
}
\arguments{
\item{source_note_tbl}{the table that contains
the source note data.}

\item{font_size}{the font size to use for the
source notes.}

\item{align}{the alignment of text to use for the
source notes.}

\item{padding}{the amount of padding to use for
the source notes.}
}
\description{
Generate a HTML fragment for the source note part
}
