% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table_caption.R
\name{generate_table_caption}
\alias{generate_table_caption}
\title{Generate a HTML fragment for the table heading part}
\usage{
generate_table_caption(title, headnote = NULL, border_top = c("solid",
  "2px", "#A8A8A8"), border_bottom = c("solid", "2px", "#A8A8A8"),
  title_font_size = "115\%", title_align = "center",
  title_padding = "3px", headnote_font_size = "85\%",
  headnote_align = "center", headnote_padding = "3px")
}
\arguments{
\item{title}{the text to be used in the table title.}

\item{headnote}{optional text to be used in the
table headnote.}

\item{border_top}{values to be used for the
\code{border-top} style property. This allows for
a horizontal rule at the top of the table.}

\item{border_bottom}{values to be used for the
\code{border-bottom} style property. This allows
for a horizontal rule below the table title or
the table headnote.}

\item{title_font_size}{the font size to use for
the table title.}

\item{title_padding}{the amount of padding to
use for the table title.}

\item{headnote_font_size}{the font size to use for
the table headnote.}

\item{headnote_padding}{the amount of padding to
use for the table headnote.}
}
\value{
a character object with an HTML fragment.
}
\description{
Generate a HTML fragment for the table heading part
}
