% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html.R
\name{get_html}
\alias{get_html}
\title{Return the table as an HTML fragment}
\usage{
get_html(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\description{
Return the table as an HTML fragment
}
