% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_latex_dependencies}
\alias{get_latex_dependencies}
\title{Get the LaTeX dependencies required for a gt table}
\usage{
get_latex_dependencies()
}
\description{
Get the LaTeX dependencies required for a gt table
}
\seealso{
Other helper functions: \code{\link{cells_styles}},
  \code{\link{escape_latex}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}}
}
\concept{helper functions}
