% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{ggplot_image}
\alias{ggplot_image}
\title{Helper function for adding a ggplot}
\usage{
ggplot_image(plot_object, height = 30, x = 5, y = 5)
}
\arguments{
\item{plot_object}{the ggplot plot object.}

\item{height}{the absolute height (px) of the image in the table cell.}

\item{x}{the x component of the plot's final aspect ratio.}

\item{y}{the y component of the plot's final aspect ratio.}
}
\value{
a character object with an HTML fragment that can be placed inside of
a cell.
}
\description{
Add a ggplot plot inside a table with this helper function.
}
\examples{
\dontrun{
library(ggplot2)

# Create a ggplot plot
plot <-
  ggplot(
    data = mtcars,
    aes(x = disp, y = hp,
        color = wt, size = mpg)) +
  geom_point()

# Create an HTML fragment that
# contains an the ggplot as an
# embedded plot
plot_html <-
  ggplot_image(
    plot_object = plot)
}
}
