% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{google_fonts}
\alias{google_fonts}
\title{Helper function for specifying a font from the Google Fonts service}
\usage{
google_fonts(name)
}
\arguments{
\item{name}{The complete name of a font available in Google Fonts.}
}
\value{
An object of class \code{google_fonts}.
}
\description{
The \code{google_fonts()} helper function can be used wherever a font name should
be specified. There are two instances where this helper can be used: the
\code{name} argument in \code{\link[=opt_table_font]{opt_table_font()}} (for setting a table font) and in that
of \code{\link[=cell_text]{cell_text()}} (used with \code{\link[=tab_style]{tab_style()}}).
}
\section{Function ID}{

7-22
}

\seealso{
Other Helper Functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()}
}
\concept{Helper Functions}
