% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_group.R
\name{grp_pull}
\alias{grp_pull}
\title{Pull out a \strong{gt} table from a \code{gt_group} container object}
\usage{
grp_pull(data, which)
}
\arguments{
\item{data}{A \code{gt_group} container object, typically generated through use of
the \code{\link[=gt_group]{gt_group()}} function along with one or more \code{gt_tbl} objects.}

\item{which}{An index value denoting which \code{gt_tbl} table should be obtained
from the \code{gt_group} object.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Should you have a \code{gt_group} object, created through use of the \code{\link[=gt_group]{gt_group()}}
function, you may have a need to extract a \strong{gt} table from that container.
The \code{grp_pull()} function makes this possible, returning a \code{gt_tbl} object.
The only thing you need to provide is the index value for the \strong{gt} table
within the \code{gt_group} object.
}
\section{Function ID}{

14-3
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other table group functions: 
\code{\link{grp_add}()},
\code{\link{grp_clone}()},
\code{\link{grp_options}()},
\code{\link{grp_replace}()},
\code{\link{grp_rm}()},
\code{\link{gt_group}()},
\code{\link{gt_split}()}
}
\concept{table group functions}
