% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gt-options}
\alias{gt-options}
\title{\strong{gt} package options}
\description{
\strong{gt} package options

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\section{Package options}{


\strong{gt} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{gt.row_group.sep}: a separator between groups for the row group
label.
}
}

