% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-package.R
\docType{package}
\name{gt-package}
\alias{gt}
\alias{gt-package}
\title{A package for making display tables}
\description{
Use a high-level, declarative interface for making display tables.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/gt}
  \item Report bugs at \url{https://github.com/rstudio/gt/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{rich@rstudio.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})

Authors:
\itemize{
  \item Joe Cheng \email{joe@rstudio.com}
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
