% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pdf.R
\name{gt_jpg}
\alias{gt_jpg}
\title{Export the gt table as a JPG file}
\usage{
gt_jpg(data, file_name = NULL, quality = 85, trim_lr = TRUE)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{file_name}{the output file name of the JPG file.}
}
\description{
Export the gt table as a JPG file
}
\examples{
\dontrun{
# Use `countrypops` to create a gt table;
# export the table as a JPG file
countrypops \%>\%
  dplyr::select(-contains("code")) \%>\%
  dplyr::filter(country_name == "Mongolia") \%>\%
  tail(5) \%>\%
  gt() \%>\%
  cols_align(
    align = "left",
    columns = vars(population)
  ) \%>\%
  gt_jpg(file_name = "countrypops.jpg")
}
}
\seealso{
Other table export functions: \code{\link{as_latex}},
  \code{\link{as_raw_html}}, \code{\link{as_rtf}},
  \code{\link{extract_summary}}, \code{\link{gt_pdf}},
  \code{\link{gt_png}}
}
\concept{table export functions}
