% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_multi.R
\name{gt_multi}
\alias{gt_multi}
\title{Create a \code{gt_multi} container for multiple \strong{gt} table objects}
\usage{
gt_multi(..., .list = list2(...), .use_parent_opts = FALSE)
}
\arguments{
\item{...}{One or more gt table (\code{gt_tbl}) objects, typically generated via
the \code{\link[=gt]{gt()}} function.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}

\item{.use_parent_opts}{Should options specified in the \code{gt_multi} object be
applied to all contained \strong{gt} tables? By default this is \code{FALSE}.}
}
\value{
An object of class \code{gt_multi}.
}
\description{
The \code{gt_multi()} function creates a container for storage of multiple \strong{gt}
tables. This type of object allows for flexibility in printing multiple
tables in different output formats. For example, if printing multiple tables
in a paginated output environment (e.g., RTF, Word, etc.), each \strong{gt} table
can be printed independently and table separation (usually a page break)
occurs between each of those.
}
