% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gt_output}
\alias{gt_output}
\title{Create a gt display table output element for Shiny}
\usage{
gt_output(outputId)
}
\value{
A gt table output element that can be included in a panel.
}
\description{
Create a gt display table output element for Shiny
}
\seealso{
render_gt
}
