% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_preview.R
\name{gt_preview}
\alias{gt_preview}
\title{Preview a gt table object}
\usage{
gt_preview(data, rownames_to_stub = FALSE, top_n = 5, bottom_n = 1)
}
\arguments{
\item{data}{a \code{data.frame} object or a tibble.}

\item{rownames_to_stub}{an option to take rownames from the input \code{data}
table as row captions in the stub.}

\item{top_n}{this value will be used as the number of rows from the top of
the table to display.}

\item{bottom_n}{the value will be used as the number of rows from the bottom
of the table to display.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Create a preview version of table. Here we can specify the first and last
\code{n} rows for the preview table.
}
\examples{
# Create a preview of the `mtcars` dataset; it
# will be displayed in the Viewer
gt_preview(
  mtcars,
  rownames_to_stub = TRUE)
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt}},
  \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_options}}, \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}},
  \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
