% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{gt_vctr}
\alias{gt_vctr}
\title{Attach a spec to a vector}
\usage{
gt_vctr(x, spec)
}
\arguments{
\item{x}{A numeric vector}

\item{spec}{Created by \code{\link[=gt_spec_number]{gt_spec_number()}} and other functions.}
}
\description{
Use \code{gt_vctr()} early in your analysis to define the formatting of numbers
and other data.
The formatting is then picked up automatically by \code{gt()}.
}
