% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_tbl_parts.R
\name{highlight_tbl_parts}
\alias{highlight_tbl_parts}
\title{Highlight Table Parts and Preview}
\usage{
highlight_tbl_parts(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}
}
\description{
Highlights the parts of a table in a table object
with color fills.
}
