% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{html}
\alias{html}
\title{Interpret input text as HTML-formatted text}
\usage{
html(text)
}
\arguments{
\item{text}{the text that is understood to be HTML text, which is to be
preserved.}
}
\value{
a character object that is tagged as an HTML fragment that is not to
be sanitized.

a character object of class \code{preserve_html}.
# Create a table object using the
# `mtcars` dataset and add a caption
# that is to interpreted as HTML
gt(mtcars, rownames_to_stub = TRUE) %>%
  tab_stubhead_caption(
    caption = html(
      "car <em>make</em> and <em>model</em>"))
}
\description{
Interpret input text as HTML-formatted text
}
