% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{id_random}
\alias{id_random}
\title{Helper for creating a random \code{id} for a \pkg{gt} table}
\usage{
id_random()
}
\description{
This helper function is to be used with \code{id} argument of the \code{\link[=gt]{gt()}} function.
The \code{id} option in \code{\link[=gt]{gt()}} uses \code{id_random()} by default however we can
optionally supply a custom \code{id} value, or, use \code{NULL} for no ID at all.
}
\seealso{
Other helper functions: \code{\link{cells_styles}},
  \code{\link{escape_latex}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}}
}
\concept{helper functions}
