% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_tables.R
\name{info_flags}
\alias{info_flags}
\title{View a table with all available flags for \code{fmt_flag()}}
\usage{
info_flags()
}
\value{
An object of class \code{gt_tbl}.
}
\description{
\code{\link[=fmt_flag]{fmt_flag()}} can be used to render flag icons within body cells that have
2-letter country codes. There are a lot of countries, so, calling
\code{info_flags()} can be helpful in showing all of the valid and supported
country codes along with their flag icons.
}
\section{Examples}{


Get a table of info on all the available flag icons.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{info_flags()
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_info_flags_1.png" alt="This image of a table was generated from the first code example in the `info_flags()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

11-8
}

\section{Function Introduced}{

\code{v0.10.0} (October 7, 2023)
}

\seealso{
Other information functions: 
\code{\link{info_currencies}()},
\code{\link{info_date_style}()},
\code{\link{info_google_fonts}()},
\code{\link{info_icons}()},
\code{\link{info_locales}()},
\code{\link{info_paletteer}()},
\code{\link{info_tf_style}()},
\code{\link{info_time_style}()},
\code{\link{info_unit_conversions}()}
}
\concept{information functions}
