% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_tables.R
\name{info_paletteer}
\alias{info_paletteer}
\title{View a table with info on many different color palettes}
\usage{
info_paletteer(color_packages = c("awtools", "dichromat", "dutchmasters",
  "ggsci", "ggpomological", "ggthemes", "ghibli", "grDevices", "jcolors",
  "LaCroixColoR", "NineteenEightyR", "nord", "ochRe", "palettetown",
  "pals", "Polychrome", "quickpalette", "rcartocolor", "RColorBrewer",
  "Redmonder", "wesanderson", "yarrr"))
}
\arguments{
\item{color_packages}{a vector of color packages to determine which
sets of palettes should be displayed in the information table.}
}
\description{
View a table with info on many different color palettes
}
\seealso{
Other information functions: \code{\link{info_date_style}},
  \code{\link{info_time_style}}
}
\concept{information functions}
