% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_tables.R
\name{info_unit_conversions}
\alias{info_unit_conversions}
\title{View a table with all units that can be converted by \code{unit_conversion()}}
\usage{
info_unit_conversions()
}
\value{
An object of class \code{gt_tbl}.
}
\description{
\code{\link[=unit_conversion]{unit_conversion()}} can be used to yield conversion factors across compatible
pairs of units. This is useful for expressing values in different units and
the conversion can be performed via the \code{scale_by} argument available in
several formatting functions. When calling \code{\link[=unit_conversion]{unit_conversion()}}, one must
supply two string-based keywords to specify the value's current units and the
desired units. All of these keywords are provided in the table shown by
calling \code{info_unit_conversions()}.
}
\section{Examples}{


Get a table of info on all the available keywords for unit conversions.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{info_unit_conversions()
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_info_unit_conversions_1.png" alt="This image of a table was generated from the first code example in the `info_unit_conversions()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

11-9
}

\section{Function Introduced}{

\code{v0.11.0} (July 9, 2024)
}

\seealso{
Other information functions: 
\code{\link{info_currencies}()},
\code{\link{info_date_style}()},
\code{\link{info_flags}()},
\code{\link{info_google_fonts}()},
\code{\link{info_icons}()},
\code{\link{info_locales}()},
\code{\link{info_paletteer}()},
\code{\link{info_tf_style}()},
\code{\link{info_time_style}()}
}
\concept{information functions}
