% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_latex.R
\name{interp_citation_latex}
\alias{interp_citation_latex}
\title{Interpolate in-text citation}
\usage{
interp_citation_latex(tab_latex)
}
\arguments{
\item{tab_latex}{The LaTeX table as a length-one character vector.}
}
\description{
Interpolate in-text citation by replacing the text \code{@cite_key} with
\code{\\cite{cite_key}}.
}
