% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_latex.R
\name{interp_degree_sym_latex}
\alias{interp_degree_sym_latex}
\title{Interpolate degree symbol}
\usage{
interp_degree_sym_latex(tab_latex)
}
\arguments{
\item{tab_latex}{The LaTeX table as a length-one character vector.}
}
\description{
Interpolate degree symbol by replacing the text \code{@circ@} with
\code{$^\\circ$}.
}
