% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.gt_tbl}
\alias{knit_print.gt_tbl}
\title{Knit print the table}
\usage{
knit_print.gt_tbl(x, ...)
}
\arguments{
\item{x}{an object of class \code{gt_tbl}.}
}
\description{
This facilitates printing of the HTML table within a knitr
code chunk.
}
\keyword{internal}
