% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.html_table}
\alias{knit_print.html_table}
\title{Knit print the table}
\usage{
\method{knit_print}{html_table}(x, ...)
}
\arguments{
\item{x}{an object of class \code{html_table}.}
}
\description{
This facilitates printing of the
HTML table within a knitr code chunk.
}
\keyword{internal}
