% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_latex.R
\name{latex_shrink}
\alias{latex_shrink}
\title{shrink latex}
\usage{
latex_shrink(latex_code, prop_shrink, align = c("l", "r", "c"))
}
\arguments{
\item{latex_code}{gt object}

\item{prop_shrink}{proportion to shrink}

\item{align}{align direction to align}
}
\description{
shrink latex
}
