% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaa.R
\name{latex_tab.options}
\alias{latex_tab.options}
\title{Additional formatting options for as_latex()}
\usage{
latex_tab.options(
  shrink = NULL,
  line.breaks = TRUE,
  column.sep = NULL,
  footnotes.align = "l",
  sourcenotes.align = "l"
)
}
\arguments{
\item{shrink}{The percentage to shrink the width by. Accepts a value < 1. Default is to stretch the table to full width of the page.}

\item{line.breaks}{Allow/do not allow line breaks in the table. Default is to allow for line breaks under two circumstances: 1.) When the text of the table at the smallest point size (5pt) cannot fit to the width of the page. 2.) A minimal number of line breaks are required to make point size larger. In the first condition, setting this parameter to FALSE will not affect the output. In the second, the font size will be scaled down to remove the minimal breaks.}

\item{column.sep}{set the}
}
\description{
Additional formatting options for as_latex()
}
