% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{lighten_palette}
\alias{lighten_palette}
\title{Lighten a palette of colors by a fixed factor}
\usage{
lighten_palette(colors, factor = 1.5)
}
\arguments{
\item{colors}{a vector of colors that is to be lightened.}

\item{factor}{the factor to which the colors will be lightened.}
}
\description{
Lighten a palette of colors by a fixed factor
}
