% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{local_image}
\alias{local_image}
\title{Helper function for adding a local image}
\usage{
local_image(file, height = 30)
}
\arguments{
\item{file}{a path to an image file.}

\item{height}{the absolute height (px) of the image in the table cell.}
}
\value{
a character object with an HTML fragment that can be placed inside of
a cell.
}
\description{
Add a local image inside a table with this helper function.
}
\examples{
\dontrun{
# Create an HTML fragment that
# contains an image
img_file_html <-
  local_image(
    file = "path/to/image.png")
}
}
