% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{local_svg}
\alias{local_svg}
\title{Helper function for adding a local image}
\usage{
local_svg(file, height = 30)
}
\arguments{
\item{file}{a path to an SVG file.}

\item{height}{the absolute height (px) of the image in the table cell.}
}
\value{
a character object with an HTML fragment that can be placed inside of
a cell.
}
\description{
Add an image from a local SVG file inside a table with this helper function.
}
\examples{
\dontrun{
# Create an HTML fragment that
# contains the SVG
img_file_html <-
  local_svg(
    file = "path/to/file.svg")
}
}
