% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{location_cells}
\alias{location_cells}
\alias{data_cells}
\alias{stub_cells}
\alias{boxhead_cells}
\alias{group_cells}
\alias{title_cells}
\title{Helpers for targeting multiple cells in different locations}
\usage{
data_cells(columns = NULL, rows = NULL)

stub_cells(rows)

boxhead_cells(columns, groups)

group_cells(groups)

title_cells(groups = c("title", "headnote"))
}
\arguments{
\item{columns, rows, groups}{either a vector of names, a vector of
indices, values provided by \code{\link{vars}()}, values provided by
\code{c()}, or a select helper function.}
}
\value{
a list object of class \code{location_cells}.
}
\description{
These helper functions are used to target individual cells in different
  locations. The locations and the specific functions used to target cells
  therein are:
}
\details{
- \code{data_cells()}: targets data cells in the table field using
  intersections of \code{columns} and \code{rows}
- \code{stub_cells()}: targets row captions in the table stub using the
  \code{rows} argument
- \code{boxhead_cells()}: targets captions for the column headers (the
  \code{columns} argument) or group spanners (the \code{groups} argument) in
  the table boxhead
- \code{group_cells()}: targets the group headings in the stub blocks using
  the \code{groups} argument
- \code{title_cells()}: targets the table title or the table headnote
  depending on the value given to the \code{groups} argument
  (\code{"title"} or \code{"headnote"})

The select helper functions are: \code{\link{starts_with}()},
  \code{\link{ends_with}()}, \code{\link{contains}()},
  \code{\link{matches}()}, \code{\link{one_of}()}, and
  \code{\link{everything}()}.
}
