% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{matches}
\alias{matches}
\title{Helper for resolving colnames or rownames that match a regular expression}
\usage{
matches(match, ignore.case = TRUE)
}
\arguments{
\item{match}{a string.}

\item{ignore.case}{if \code{TRUE}, the default, case is ignored when
matching names.}
}
\description{
Helper for resolving colnames or rownames that match a regular expression
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{contains}},
  \code{\link{ends_with}}, \code{\link{everything}},
  \code{\link{funs}}, \code{\link{html}}, \code{\link{md}},
  \code{\link{one_of}}, \code{\link{pct}},
  \code{\link{px}}, \code{\link{starts_with}},
  \code{\link{vars}}
}
\concept{helper functions}
