% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_formatting.R
\name{md}
\alias{md}
\title{Interpret input text as Markdown-formatted text}
\usage{
md(text)
}
\arguments{
\item{text}{the text that is understood to contain
Markdown formatting.}
}
\value{
a character object that is tagged for
a Markdown to HTML transformation.
}
\description{
Interpret input text as Markdown-formatted text
}
