% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{merge_columns}
\alias{merge_columns}
\title{Merge columns according to transform directives}
\usage{
merge_columns(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that
is created using the \code{gt()} function.}
}
\description{
Merge columns according to transform directives
}
