% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{move_columns}
\alias{move_columns}
\title{Move one or more columns to the end}
\usage{
move_columns(html_tbl, columns, after)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{the column names to move to
as a group to a different position. The
order of the remaining columns will be
preserved. Values provided that do not
correspond to column names will be
disregarded.}
}
\value{
an object of class \code{html_table}.
}
\description{
Move one or more columns to the end
}
