% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{not_in_group}
\alias{not_in_group}
\title{Helper for collecting ungrouped rows or columns}
\usage{
not_in_group()
}
\value{
a character object of class \code{not_in_group}.
}
\description{
Helper for collecting ungrouped rows or columns
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{columns_with}},
  \code{\link{data_cells}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{rownames_with}},
  \code{\link{rows_where}}, \code{\link{vars}}
}
