% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{one_of}
\alias{one_of}
\title{Helper for resolving colnames or rownames that are part of a set}
\usage{
one_of(..., ignore.case = TRUE)
}
\arguments{
\item{...}{one or more character vectors.}
}
\description{
See \code{tidyselect::\link[tidyselect]{select_helpers}} for details.
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{contains}},
  \code{\link{ends_with}}, \code{\link{everything}},
  \code{\link{funs}}, \code{\link{html}},
  \code{\link{matches}}, \code{\link{md}},
  \code{\link{pct}}, \code{\link{px}},
  \code{\link{starts_with}}, \code{\link{vars}}
}
\concept{helper functions}
