% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_add_table_outline}
\alias{opt_add_table_outline}
\title{Option to wrap an outline around the entire table}
\usage{
opt_add_table_outline(data, style = "solid", width = px(3),
  color = "#D3D3D3")
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{style, width, color}{The style, width, and color properties for the table
border.}
}
\description{
Option to wrap an outline around the entire table
}
\seealso{
Other table option functions: \code{\link{opt_add_row_striping}},
  \code{\link{opt_align_table_header}},
  \code{\link{opt_all_caps}},
  \code{\link{opt_footnote_marks}},
  \code{\link{opt_remove_row_striping}},
  \code{\link{opt_remove_table_outline}},
  \code{\link{opt_remove_top_bottom_borders}}
}
\concept{table option functions}
