% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_align_table_header}
\alias{opt_align_table_header}
\title{Option to align the table header}
\usage{
opt_align_table_header(data, align = "center")
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}
}
\description{
Option to align the table header
}
\seealso{
Other table option functions: \code{\link{opt_add_row_striping}},
  \code{\link{opt_add_table_outline}},
  \code{\link{opt_all_caps}},
  \code{\link{opt_footnote_marks}},
  \code{\link{opt_remove_row_striping}},
  \code{\link{opt_remove_table_outline}},
  \code{\link{opt_remove_top_bottom_borders}}
}
\concept{table option functions}
