% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_all_caps}
\alias{opt_all_caps}
\title{Option to use all caps in select table locations}
\usage{
opt_all_caps(data, locations = c("column_labels", "stub", "row_group"))
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{locations}{Which locations should undergo this text transformation? By
default it includes all of the \code{"column_labels"}, the \code{"stub"}, and the
\code{"row_group"} locations. However, we could just choose one or two of those.}
}
\description{
Option to use all caps in select table locations
}
\seealso{
Other table option functions: \code{\link{opt_add_row_striping}},
  \code{\link{opt_add_table_outline}},
  \code{\link{opt_align_table_header}},
  \code{\link{opt_footnote_marks}},
  \code{\link{opt_remove_row_striping}},
  \code{\link{opt_remove_table_outline}},
  \code{\link{opt_remove_top_bottom_borders}}
}
\concept{table option functions}
