% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_css}
\alias{opt_css}
\title{Option to add custom CSS for the table}
\usage{
opt_css(data, css, add = TRUE, allow_duplicates = FALSE)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{css}{The CSS to include as part of the rendered table's \verb{<style>}
element. A \code{css_list} that's generated though the \code{\link[=css_list]{css_list()}} function
(along with the \code{\link[=css_sel]{css_sel()}} and \code{\link[=css_dec]{css_dec()}} functions) is accepted here.
Another option is to read in a CSS file by use of the \code{\link[=css_file]{css_file()}}
function. Should these methods not be appealing, a character vector of CSS
lines representing CSS rulesets can be provided.}

\item{add}{If \code{TRUE}, the default, the CSS is added to any already-defined
CSS (typically from previous calls of \code{\link[=opt_table_font]{opt_table_font()}}, \code{opt_css()}, or,
directly setting CSS the \code{table.additional_css} value in \code{\link[=tab_options]{tab_options()}}).
If this is set to \code{FALSE}, the CSS provided here will replace any
previously-stored CSS.}

\item{allow_duplicates}{When this is \code{FALSE} (the default), the CSS provided
here won't be added (provided that \code{add = TRUE}) if it is seen in the
already-defined CSS.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{opt_css()} function makes it possible to add CSS to a \strong{gt} table. This
CSS will be added after the compiled CSS that \strong{gt} generates automatically
when the object is transformed to an HTML output table. You can supply \code{css}
as a vector of lines

We have the option to supply either a system font for the \code{font_name}, or, a
font available at the Google Fonts service by use of the \code{\link[=google_font]{google_font()}}
helper function.
}
\section{Function ID}{

9-8
}

\seealso{
Other Table Option Functions: 
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_row_striping}()},
\code{\link{opt_table_font}()},
\code{\link{opt_table_lines}()},
\code{\link{opt_table_outline}()}
}
\concept{Table Option Functions}
