% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_footnote_spec}
\alias{opt_footnote_spec}
\title{Option to specify the formatting of footnote marks}
\usage{
opt_footnote_spec(data, spec_ref = NULL, spec_ftr = NULL)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{spec_ref, spec_ftr}{Specification of the footnote marks when behaving as
footnote references and as marks in the footer section of the table. This
is a string containing spec characters. The default is the spec string
\code{"^i"}, which is superscript text set in italics.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Modify the way footnote marks are formatted. This can be performed for those
footnote marks that alight to the targeted text in cells in various locations
in the table or the footnote marks that appear in the table footer. A simple
specification string can be provided for either or both types of marks in
\code{opt_footnote_spec()} . This function serves as a shortcut for using either
of \code{tab_options(footnotes.spec_ref = {spec})} or
\code{tab_options(footnotes.spec_ftr = {spec})}.
}
\section{Specification rules for the formatting of footnote marks}{


Footnote marks can have the following types of formatting:
\itemize{
\item as superscript text (use \code{"^"} character in spec) or regular-sized text
residing on the baseline
\item bold text (with \code{"b"}), italicized text (with \code{"i"}), or unstyled text
(don't use \code{"b"} or \code{"i"})
\item enclosure in parentheses (use \code{"("} character in spec) or square brackets
(use \code{"["})
}
}

\section{Examples}{


Use \code{\link{sza}} to create a \strong{gt} table, adding three footnotes. Call
\code{opt_footnote_marks()} to specify which footnote marks to use.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sza |>
  dplyr::filter(latitude == 30) |>
  dplyr::group_by(tst) |>
  dplyr::summarize(
    SZA.Max = if (
      all(is.na(sza))) \{
      NA
    \} else \{
      max(sza, na.rm = TRUE)
    \},
    SZA.Min = if (
      all(is.na(sza))) \{
      NA
    \} else \{
      min(sza, na.rm = TRUE)
    \},
    .groups = "drop"
  ) |>
  gt(rowname_col = "tst") |>
  tab_spanner_delim(delim = ".") |>
  sub_missing(
    columns = everything(),
    missing_text = "90+"
  ) |>
  tab_stubhead(label = "TST") |>
  tab_footnote(
    footnote = "True solar time.",
    locations = cells_stubhead()
  ) |>
  tab_footnote(
    footnote = "Solar zenith angle.",
    locations = cells_column_spanners(
      spanners = "spanner-SZA.Max"
    )
  ) |>
  tab_footnote(
    footnote = "The Lowest SZA.",
    locations = cells_stub(rows = "1200")
  ) |>
  opt_footnote_marks(marks = "standard")
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_opt_footnote_spec_1.png" alt="This image of a table was generated from the first code example in the `opt_footnote_spec()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

10-4
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other table option functions: 
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_css}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_horizontal_padding}()},
\code{\link{opt_interactive}()},
\code{\link{opt_row_striping}()},
\code{\link{opt_stylize}()},
\code{\link{opt_table_font}()},
\code{\link{opt_table_lines}()},
\code{\link{opt_table_outline}()},
\code{\link{opt_vertical_padding}()}
}
\concept{table option functions}
