% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_fully_lined}
\alias{opt_fully_lined}
\title{Option to create lines everywhere in a table}
\usage{
opt_fully_lined(data)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{opt_fully_lined()} function yields a table where all possible lines are
visible. This is great if you want to start off with lots of lines and
subtract just a few of them with \code{\link[=tab_options]{tab_options()}} or \code{\link[=tab_style]{tab_style()}}. Or, use it
in those cases where you absolutely need to have boundaries between pieces of
information.
}
\section{Function ID}{

9-6
}

\examples{
# Use `exibble` to create a gt table with
# a number of table parts added; then, use
# the `opt_fully_lined()` function to
# haves lines everywhere there can possibly
# be lines
tab_1 <-
  exibble \%>\%
  gt(rowname_col = "row", groupname_col = "group") \%>\%
  summary_rows(
    groups = "grp_a",
    columns = vars(num, currency),
    fns = list(
      min = ~min(., na.rm = TRUE),
      max = ~max(., na.rm = TRUE)
    )) \%>\%
  grand_summary_rows(
    columns = vars(currency),
    fns = list(
      total = ~sum(., na.rm = TRUE)
    )) \%>\%
  tab_source_note(source_note = "This is a source note.") \%>\%
  tab_footnote(
    footnote = "This is a footnote.",
    locations = cells_body(columns = 1, rows = 1)
  ) \%>\%
  tab_header(
    title = "The title of the table",
    subtitle = "The table's subtitle"
  ) \%>\%
  opt_fully_lined()

}
\seealso{
Other Table Option Functions: 
\code{\link{opt_add_row_striping}()},
\code{\link{opt_add_table_outline}()},
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_lineless}()},
\code{\link{opt_remove_table_outline}()},
\code{\link{opt_remove_top_bottom_borders}()}
}
\concept{Table Option Functions}
