% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_horizontal_padding}
\alias{opt_horizontal_padding}
\title{Option to expand or contract horizontal padding}
\usage{
opt_horizontal_padding(data, steps = 0)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{steps}{A positive or negative factor by which the horizontal padding
will be adjusted. Must be a number between \code{-1.0} and \code{2.0}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
This function can increase or decrease the horizontal padding throughout all
locations of a gt table by an arbitrary number of steps, which is defined by
a real number between -1.0 and 2.0.
}
\section{Function ID}{

9-5
}

\seealso{
Other Table Option Functions: 
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_css}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_row_striping}()},
\code{\link{opt_table_font}()},
\code{\link{opt_table_lines}()},
\code{\link{opt_table_outline}()},
\code{\link{opt_vertical_padding}()}
}
\concept{Table Option Functions}
