% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_remove_table_outline}
\alias{opt_remove_table_outline}
\title{Option to completely remove any table outlines}
\usage{
opt_remove_table_outline(data)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}
}
\description{
Option to completely remove any table outlines
}
\seealso{
Other table option functions: \code{\link{opt_add_row_striping}},
  \code{\link{opt_add_table_outline}},
  \code{\link{opt_align_table_header}},
  \code{\link{opt_all_caps}},
  \code{\link{opt_footnote_marks}},
  \code{\link{opt_remove_row_striping}},
  \code{\link{opt_remove_top_bottom_borders}}
}
\concept{table option functions}
