% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_remove_top_bottom_borders}
\alias{opt_remove_top_bottom_borders}
\title{Option to remove the top and bottom borders}
\usage{
opt_remove_top_bottom_borders(data)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}
}
\description{
Option to remove the top and bottom borders
}
\seealso{
Other table option functions: \code{\link{opt_footnote_marks}},
  \code{\link{opt_table_outline}}
}
\concept{table option functions}
