% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_stylize}
\alias{opt_stylize}
\title{Stylize your table with a colorful look}
\usage{
opt_stylize(data, style = 1, color = "blue")
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{style}{Six numbered styles are available. Simply provide a number from
\code{1} (the default) to \code{6} to choose a distinct look.}

\item{color}{There are seven color variations: \code{"blue"} (the default),
\code{"orange"}, \code{"aqua"}, \code{"pink"}, \code{"green"}, \code{"red"}, and \code{"gray"}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
This function applies a set of styles related to background colors,
line colors, and line styles. There are six \code{style}s to choose from, they
largely vary in the extent of coloring applied to different locations. Some
have table borders applied, some apply darker colors to the table stub and
summary sections, and, some even have vertical lines. In addition to \code{style},
there are seven \code{color} variations that each use a range of five color tints.
Each of the color tints have been fine-tuned to maximize the contrast between
text and its background. In total, There are 42 combinations of \code{style} and
\code{color}. We can use the \code{\link[=info_preset_colorized]{info_preset_colorized()}} function to get an
information table which has the names and previews for each of the color list
keywords.
}
\section{Examples}{


Use \code{exibble} to create a gt table with a number of table parts added. Then,
use the \code{opt_stylize()} function to give the table some additional style
(using the \code{"aqua"} color variation and style number \code{6}).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble \%>\%
  gt(rowname_col = "row", groupname_col = "group") \%>\%
  summary_rows(
    groups = "grp_a",
    columns = c(num, currency),
    fns = list(
      min = ~min(., na.rm = TRUE),
      max = ~max(., na.rm = TRUE)
    )) \%>\%
  grand_summary_rows(
    columns = currency,
    fns = list(
      total = ~sum(., na.rm = TRUE)
    )) \%>\%
  tab_source_note(source_note = "This is a source note.") \%>\%
  tab_footnote(
    footnote = "This is a footnote.",
    locations = cells_body(columns = 1, rows = 1)
  ) \%>\%
  tab_header(
    title = "The title of the table",
    subtitle = "The table's subtitle"
  ) \%>\%
  opt_stylize(style = 6, color = "aqua")
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_opt_stylize_1.png" style="width:100\%;">
}}
}

\section{Function ID}{

9-10
}

\seealso{
Other Table Option Functions: 
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_css}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_horizontal_padding}()},
\code{\link{opt_row_striping}()},
\code{\link{opt_table_font}()},
\code{\link{opt_table_lines}()},
\code{\link{opt_table_outline}()},
\code{\link{opt_vertical_padding}()}
}
\concept{Table Option Functions}
