% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_table_font}
\alias{opt_table_font}
\title{Option to define a custom font for the table}
\usage{
opt_table_font(data, font, weight = NULL, style = NULL, add = TRUE)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{font}{Either the name of a font available in the user system or a call
to \code{\link[=google_fonts]{google_fonts()}}, which has a large selection of typefaces.}

\item{weight}{The weight of the font. Can be a text-based keyword such as
\code{"normal"}, \code{"bold"}, \code{"lighter"}, \code{"bolder"}, or, a numeric value between
\code{1} and \code{1000}, inclusive. Note that only variable fonts may support the
numeric mapping of weight.}

\item{style}{The text style. Can be one of either \code{"normal"}, \code{"italic"}, or
\code{"oblique"}.}

\item{add}{Should this font be added to the front of the already-defined
fonts for the table? By default, this is \code{TRUE} and is recommended since
the list serves as fallbacks when certain fonts are not available.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{opt_table_font()} function makes it possible to define a custom font for
the entire \strong{gt} table. The standard fallback fonts are still set by default
but the font defined here will take precedence. You could still have
different fonts in select locations in the table, and for that you would need
to use \code{\link[=tab_style]{tab_style()}} in conjunction with the \code{\link[=cell_text]{cell_text()}} helper function.

We have the option to supply either a system font for the \code{font_name}, or, a
font available at the Google Fonts service by use of the \code{\link[=google_fonts]{google_fonts()}}
helper function.
}
\section{Function ID}{

9-7
}

\seealso{
Other Table Option Functions: 
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_row_striping}()},
\code{\link{opt_table_lines}()},
\code{\link{opt_table_outline}()}
}
\concept{Table Option Functions}
