% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opt_vertical_padding}
\alias{opt_vertical_padding}
\title{Option to expand or contract vertical padding}
\usage{
opt_vertical_padding(data, steps = 0)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{steps}{A positive or negative factor by which the vertical padding will
be adjusted. Must be a number between \code{-1.0} and \code{2.0}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Increase or decrease the vertical padding throughout all locations of a gt
table by an arbitrary number of 'steps', which here are defined by a real
number between -1.0 and 2.0. This function serves as a shortcut for
setting the following eight options in \code{\link[=tab_options]{tab_options()}}:
\itemize{
\item \code{heading.padding}
\item \code{column_labels.padding}
\item \code{data_row.padding}
\item \code{row_group.padding}
\item \code{summary_row.padding}
\item \code{grand_summary_row.padding}
\item \code{footnotes.padding}
\item \code{source_notes.padding}
}
}
\section{Function ID}{

9-4
}

\seealso{
Other Table Option Functions: 
\code{\link{opt_align_table_header}()},
\code{\link{opt_all_caps}()},
\code{\link{opt_css}()},
\code{\link{opt_footnote_marks}()},
\code{\link{opt_horizontal_padding}()},
\code{\link{opt_row_striping}()},
\code{\link{opt_table_font}()},
\code{\link{opt_table_lines}()},
\code{\link{opt_table_outline}()}
}
\concept{Table Option Functions}
