% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_latex.R
\name{output_latex}
\alias{output_latex}
\title{Output a gt table object as LateX}
\usage{
output_latex(data)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}
}
\description{
Take a \code{gt_tbl} table object and emit LateX.
}
\examples{
\dontrun{
# Create a `gt` table using the `sleep`
# dataset and then write the table to
# an RTF file
gt(data = sleep) \%>\%
  output_latex()
}
}
\seealso{
Other table export functions: \code{\link{extract_summary}},
  \code{\link{write_rtf}}
}
\concept{table export functions}
