% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{peeps}
\alias{peeps}
\title{A table of personal information for people all over the world}
\format{
A tibble with 100 rows and 14 variables:
\describe{
\item{name_given, name_family}{The given and family name of individual.}
\item{address}{The street address of the individual.}
\item{city}{The name of the city or locality in which the individual
resides.}
\item{state_prov}{The state or province associated with the \code{city} and
\code{address}. This is \code{NA} where data is not needed when generating an
acceptable mailing address.}
\item{postcode}{The post code for the \code{address}.}
\item{country}{The 3-letter country code representative of the resident's
country.}
\item{email_addr}{The individual's email address.}
\item{phone_number, country_code}{The individual's phone number and the
country code associated with the phone number.}
\item{gender}{The gender of the individual.}
\item{dob}{The individual's date of birth (DOB) in the 'YYYY-MM-DD' format.}
\item(height_cm, weight_kg){The height and weight of the individual in
centimeters (cm) and kilograms (kg), respectively.}
}
}
\usage{
peeps
}
\description{
This dataset containing records for one hundred people residing in ten
different countries.
}
\section{Examples}{


Here is a glimpse at the data available in \code{peeps}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dplyr::glimpse(peeps)
#> Rows: 100
#> Columns: 14
#> $ name_given   <chr> "Ruth", "Peter", "Fanette", "Judyta", "Leonard", "Maymun"~
#> $ name_family  <chr> "Conte", "Möller", "Gadbois", "Borkowska", "Jacobs", "Kho~
#> $ address      <chr> "4299 Bobcat Drive", "3705 Hidden Pond Road", "4200 Swick~
#> $ city         <chr> "Baileys Crossroads", "Red Boiling Springs", "New Orleans~
#> $ state_prov   <chr> "MD", "TN", "LA", "NY", "CA", "OH", "IN", "MA", "CA", "TX~
#> $ postcode     <chr> "22041", "37150", "70112", "14125", "90036", "45013", "46~
#> $ country      <chr> "USA", "USA", "USA", "USA", "USA", "USA", "USA", "USA", "~
#> $ email_addr   <chr> "rcconte@example.com", "pmoeller@example.com", "fan_gadbo~
#> $ phone_number <chr> "240-783-7630", "615-699-3517", "985-205-2970", "585-948-~
#> $ country_code <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 32, 32, 32, 32, 32, 32, 32,~
#> $ gender       <chr> "female", "male", "female", "female", "male", "male", "fe~
#> $ dob          <date> 1949-03-16, 1939-11-22, 1970-12-20, 1965-07-19, 1985-10-~
#> $ height_cm    <int> 153, 175, 167, 156, 177, 172, 168, 165, 181, 187, 164, 15~
#> $ weight_kg    <dbl> 76.4, 74.9, 61.6, 54.5, 113.2, 88.4, 63.5, 61.3, 99.7, 10~
}\if{html}{\out{</div>}}
}

\section{Dataset ID and Badge}{

DATA-8

\if{html}{\out{
<div style="text-align:center;"><img src="https://raw.githubusercontent.com/rstudio/gt/master/images/dataset_peeps.png" alt="This image of that of a dataset badge." style="width:50\%;padding-bottom:20px;"></div>
}}
}

\section{Dataset Introduced}{

\emph{In Development}
}

\seealso{
Other datasets: 
\code{\link{constants}},
\code{\link{countrypops}},
\code{\link{exibble}},
\code{\link{gtcars}},
\code{\link{illness}},
\code{\link{metro}},
\code{\link{photolysis}},
\code{\link{pizzaplace}},
\code{\link{reactions}},
\code{\link{rx_addv}},
\code{\link{rx_adsl}},
\code{\link{sp500}},
\code{\link{sza}},
\code{\link{towny}}
}
\concept{datasets}
\keyword{datasets}
