% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gt_group}
\alias{print.gt_group}
\title{Print a collection of \strong{gt} tables}
\usage{
\method{print}{gt_group}(x, ..., view = interactive())
}
\arguments{
\item{x}{An object of class \code{gt_group}.}

\item{...}{Any additional parameters.}

\item{view}{The value for \code{print()}s \code{browse} argument.}
}
\description{
This facilitates printing of multiple HTML tables (in a \code{gt_group} object) to
the R console.
}
\keyword{internal}
