% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gt_multi}
\alias{print.gt_multi}
\title{Print a collection of tables}
\usage{
\method{print}{gt_multi}(x, ..., view = interactive())
}
\arguments{
\item{x}{An object of class \code{gt_multi}.}

\item{...}{Any additional parameters.}

\item{view}{The value for \code{print()}s \code{browse} argument.}
}
\description{
This facilitates printing of multiple HTML tables (in a \code{gt_multi} object) to
the R console.
}
\keyword{internal}
