% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gt_tbl}
\alias{print.gt_tbl}
\title{Print the table}
\usage{
\method{print}{gt_tbl}(x, ..., view = interactive())
}
\arguments{
\item{x}{an object of class \code{gt_tbl}.}
}
\description{
This facilitates printing of the HTML table to the R console.
}
\keyword{internal}
