% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gti_tbl}
\alias{print.gti_tbl}
\title{Print the table}
\usage{
\method{print}{gti_tbl}(x, ..., view = interactive())
}
\arguments{
\item{x}{An object of class \code{gti_tbl}.}

\item{...}{Any additional parameters.}

\item{view}{The value for \code{print()}s \code{browse} argument.}
}
\description{
This facilitates printing of the HTML table to the R console.
}
\keyword{internal}
