% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.html_table}
\alias{print.html_table}
\title{Print the table}
\usage{
\method{print}{html_table}(x, ...)
}
\arguments{
\item{x}{an object of class
\code{html_table}.}
}
\description{
This function will provide a
print out of the HTML table.
}
\keyword{internal}
