% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rtf_text}
\alias{print.rtf_text}
\title{Print RTF text}
\usage{
\method{print}{rtf_text}(x, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Any additional parameters.}
}
\description{
Print RTF text
}
\keyword{internal}
