% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_html.R
\name{process_html}
\alias{process_html}
\title{Transform a gt table object to an HTML table}
\usage{
process_html(tbl)
}
\arguments{
\item{tbl}{a \code{gt_tbl} object that is
created using the \code{gt()} function.}
}
\value{
a character object with an HTML table
}
\description{
Take a \code{gt_tbl} table object and transform
it to an HTML table.
}
