% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_multi.R
\name{pull_gt_tbls}
\alias{pull_gt_tbls}
\title{Pull out a \strong{gt} table from a \code{gt_multi} container object}
\usage{
pull_gt_tbls(data, which)
}
\arguments{
\item{data}{A \code{gt_multi} container object, typically generated through use of
the \code{\link[=gt_multi]{gt_multi()}} function along with one or more \code{gt_tbl} objects.}

\item{which}{An index value denoting which \code{gt_tbl} table should be obtained
from the \code{gt_multi} object.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Should you have a \code{gt_multi} object, created through use of the \code{\link[=gt_multi]{gt_multi()}}
function, you may have a need to extract a \strong{gt} table from that container.
The \code{pull_gt_tbls()} function makes this possible, returning a \code{gt_tbl}
object. The only thing you need to provide is the index value for the \strong{gt}
table within the \code{gt_multi} object.
}
\section{Function ID}{

14-2
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other multiple-table functions: 
\code{\link{add_gt_tbls}()},
\code{\link{gt_multi}()},
\code{\link{opts_gt_tbls}()},
\code{\link{replace_gt_tbls}()},
\code{\link{rm_gt_tbls}()}
}
\concept{multiple-table functions}
