% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{px}
\alias{px}
\title{Helper for providing a numeric value as pixels value}
\usage{
px(x)
}
\arguments{
\item{x}{the numeric value to format as a string (e.g., \code{"12px"}) for
some \code{\link[=tab_options]{tab_options()}} arguments that can take values as units of
pixels (e.g., \code{table.font.size}).}
}
\description{
Helper for providing a numeric value as pixels value
}
\seealso{
Other Helper Functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{escape_latex}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{location_cells}},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{random_id}()}
}
\concept{Helper Functions}
