% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_table.R
\name{quick_table}
\alias{quick_table}
\title{Quickly create an HTML table from tabular data}
\usage{
quick_table(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
a character object with an HTML
fragment.
}
\description{
Create a simple HTML table by providing a
\code{data.frame} or tibble object. The
function will sensibly style the table and
output an HTML fragment.
}
\details{
If more control over the table styling is
required, then an alternative is to use
\code{gt()} and other API functions for
building a table.
}
\examples{
# Quickly create an html table using the
# `quick_table()` with the iris dataset
html <- quick_table(iris)

# The resulting object is an
# HTML fragment; and it can be viewed
# with `render_table()`
html \%>\% render_table()
}
