% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{remove_columns}
\alias{remove_columns}
\title{Move one or more columns to the end}
\usage{
remove_columns(html_tbl, columns)
}
\arguments{
\item{html_tbl}{an HTML table object that
is created using the \code{gt()} function.}

\item{columns}{the column names to remove
from the table. The order of the remaining
columns will be preserved. Values provided
that do not correspond to column names will
be disregarded.}
}
\value{
an object of class \code{html_table}.
}
\description{
Move one or more columns to the end
}
