% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_table.R
\name{render_table}
\alias{render_table}
\title{Render an HTML table object}
\usage{
render_table(x)
}
\arguments{
\item{x}{a \code{data.frame} object, a tibble,
or an object of class \code{html_table}.}
}
\description{
View an HTML table in the viewer pane.
}
