% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder all columns}
\usage{
reorder_columns(html_tbl, columns)
}
\arguments{
\item{html_tbl}{an HTML table object that
is created using the \code{gt()} function.}

\item{columns}{a vector of column names in
the revised order. All column names must
be provided without any repeats.}
}
\value{
an object of class \code{html_table}.
}
\description{
Reorder all columns
}
