% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_multi.R
\name{replace_gt_tbls}
\alias{replace_gt_tbls}
\title{Replace one or more \strong{gt} tables in a \code{gt_multi} container object}
\usage{
replace_gt_tbls(.data, ..., .list = list2(...), .which)
}
\arguments{
\item{.data}{A \code{gt_multi} container object, typically generated through use
of the \code{\link[=gt_multi]{gt_multi()}} function along with one or more \code{gt_tbl} objects.}

\item{...}{One or more \strong{gt} table (\code{gt_tbl}) objects, typically generated
via the \code{\link[=gt]{gt()}} function.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}

\item{.which}{Index values denoting which \code{gt_tbl} tables should be replaced
in the \code{gt_multi} object.}
}
\value{
An object of class \code{gt_multi}.
}
\description{
The \code{\link[=gt_multi]{gt_multi()}} function can be used to create a container for multiple
\strong{gt} tables. In some circumstances, you might want to replace a specific
\code{gt_tbl} object (or multiple) with a different one. This can be done with the
\code{replace_gt_tbls()} function. The important thing is that the number of
\strong{gt} tables provided must equal the number of indices for tables present
in the \code{gt_multi} object.
}
\section{Function ID}{

14-4
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other multiple-table functions: 
\code{\link{add_gt_tbls}()},
\code{\link{gt_multi}()},
\code{\link{opts_gt_tbls}()},
\code{\link{pull_gt_tbls}()},
\code{\link{rm_gt_tbls}()}
}
\concept{multiple-table functions}
