% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace missing values}
\usage{
replace_missing(html_tbl, columns = NULL, types = NULL, replacement)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{an optional vector of column names
for which the left alignment should be applied.}

\item{types}{an optional vector of column types
for which the left alignment should be applied.}

\item{replacement}{the value or string that
should be used to replace the missing values.}
}
\value{
an object of class \code{html_table}.
}
\description{
Replace any missing values in specified columns
or columns of a given type.
}
