% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{resolving_helpers}
\alias{resolving_helpers}
\alias{reexports}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{one_of}
\alias{everything}
\title{Helpers for resolving column names or row labels}
\arguments{
\item{match}{a string.}

\item{ignore.case}{if \code{TRUE}, the default, case is ignored when
matching names.}
}
\description{
We can resolve one or multiple column names or row labels by using a select
  helper function like \code{starts_with}, \code{contains}, etc., with an
  argument like \code{columns}, \code{rows}, or \code{groups}. This works
  inside the the \code{\link{location_cells}} helper functions, which, in
  turn, are to be used within \code{\link{tab_footnote}()} or
  \code{\link{tab_footnote}()} (both at the \code{locations} argument).
}
\details{
The helper functions are:

\itemize{
  \item \code{starts_with()}: targets the table title or the table headnote
    depending on the value given to the \code{groups} argument
    (\code{"title"} or \code{"headnote"}).
  \item \code{ends_with()}: targets captions for the column headers (the
    \code{columns} argument) or group spanners (the \code{groups} argument)
    in the table boxhead.
  \item \code{contains()}: targets the group headings in the stub blocks
    using the \code{groups} argument.
  \item \code{matches()}: targets row captions in the table stub using the
    \code{rows} argument.
  \item \code{one_of()}: targets data cells in the table field using
    intersections of \code{columns} and \code{rows}.
  \item \code{everything()}: targets data cells in the table field using
    intersections of \code{columns} and \code{rows}.
}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{everything}}}
}}

