% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_remove.R
\name{rm_caption}
\alias{rm_caption}
\title{Remove the stubhead label}
\usage{
rm_caption(data)
}
\arguments{
\item{data}{A table object of class \code{gt_tbl}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
We can easily remove the caption text from a \strong{gt} table with
\code{rm_caption()}. The caption may exist if it were set through the \code{\link[=gt]{gt()}}
\code{caption} argument or via \code{\link[=tab_caption]{tab_caption()}}.

This function for removal is useful if you have received a \strong{gt} table
(perhaps through an API that returns \strong{gt} objects) but would prefer that
the table not have a caption at all. This function is safe to use even if
there is no table caption set in the input \code{gt_tbl} object.
}
\section{Function ID}{

6-6
}

\seealso{
Other part removal functions: 
\code{\link{rm_footnotes}()},
\code{\link{rm_header}()},
\code{\link{rm_source_notes}()},
\code{\link{rm_spanners}()},
\code{\link{rm_stubhead}()}
}
\concept{part removal functions}
