% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_remove.R
\name{rm_footnotes}
\alias{rm_footnotes}
\title{Remove table footnotes}
\usage{
rm_footnotes(data, footnotes = everything())
}
\arguments{
\item{data}{A table object of class \code{gt_tbl}.}

\item{footnotes}{A specification of which footnotes should be removed.
The footnotes to be removed can be given as a vector of integer values
(they are stored as integer positions, in order of creation, starting at
\code{1}). A select helper can also be used and, by default, this is
\code{everything()} (whereby all footnotes will be removed).}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
If you have one or more footnotes that ought to be removed, the
\code{rm_footnotes()} function allows for such a selective removal. The table
footer is an optional table part that is positioned below the table body,
containing areas for both the footnotes and source notes.

This function for removal is useful if you have received a \strong{gt} table
(perhaps through an API that returns \strong{gt} objects) but would prefer that
some or all of the footnotes be removed. This function is safe to use even if
there are no footnotes in the input \code{gt_tbl} object so long as select helpers
(such as the default \code{everything()}) are used instead of explicit integer
values.
}
\section{Function ID}{

6-4
}

\seealso{
Other part removal functions: 
\code{\link{rm_header}()},
\code{\link{rm_source_notes}()},
\code{\link{rm_spanners}()},
\code{\link{rm_stubhead}()}
}
\concept{part removal functions}
