% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_multi.R
\name{rm_gt_tbls}
\alias{rm_gt_tbls}
\title{Remove one or more \strong{gt} tables from a \code{gt_multi} container object}
\usage{
rm_gt_tbls(data, which)
}
\arguments{
\item{data}{A \code{gt_multi} container object, typically generated through use of
the \code{\link[=gt_multi]{gt_multi()}} function along with one or more \code{gt_tbl} objects.}

\item{which}{An index value denoting which \code{gt_tbl} table should be removed
from the \code{gt_multi} object.}
}
\value{
An object of class \code{gt_multi}.
}
\description{
A \code{gt_multi} object, created through use of the \code{\link[=gt_multi]{gt_multi()}} function, can
hold a multiple of \strong{gt} tables. However, you might want to delete one or
more \code{gt_tbl} objects table from that container. With \code{rm_gt_tbls()}, this
possible and safe to perform. What's returned is a \code{gt_multi} object with the
specified \code{gt_tbl} objects gone. The only thing you need to provide is the
index value for the \strong{gt} table within the \code{gt_multi} object.
}
\section{Function ID}{

14-5
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other multiple-table functions: 
\code{\link{add_gt_tbls}()},
\code{\link{gt_multi}()},
\code{\link{opts_gt_tbls}()},
\code{\link{pull_gt_tbls}()},
\code{\link{replace_gt_tbls}()}
}
\concept{multiple-table functions}
