% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_remove.R
\name{rm_source_notes}
\alias{rm_source_notes}
\title{Remove table source notes}
\usage{
rm_source_notes(data, source_notes = everything())
}
\arguments{
\item{data}{A table object of class \code{gt_tbl}.}

\item{source_notes}{A specification of which source notes should be removed.
The source notes to be removed can be given as a vector of integer values
(they are stored as integer positions, in order of creation, starting at
\code{1}). A select helper can also be used and, by default, this is
\code{everything()} (whereby all source notes will be removed).}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
If you have one or more source notes that ought to be removed, the
\code{rm_source_notes()} function allows for such a selective removal. The table
footer is an optional table part that is positioned below the table body,
containing areas for both the footnotes and the source notes.

This function for removal is useful if you have received a \strong{gt} table
(perhaps through an API that returns \strong{gt} objects) but would prefer that
some or all of the source notes be removed. This function is safe to use even
if there are no source notes in the input \code{gt_tbl} object so long as select
helpers (such as the default \code{everything()}) are used instead of explicit
integer values.
}
\section{Function ID}{

10-2
}

\seealso{
Other part removal functions: 
\code{\link{rm_header}()}
}
\concept{part removal functions}
