% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_remove.R
\name{rm_spanners}
\alias{rm_spanners}
\title{Remove column spanner labels}
\usage{
rm_spanners(data, spanners = everything())
}
\arguments{
\item{data}{A table object of class \code{gt_tbl}.}

\item{spanners}{A specification of which spanner column labels should be
removed. Those to be removed can be given as a vector of spanner ID values
(every spanner column label has one, either set by the user or by \strong{gt}
when using \code{\link[=tab_spanner_delim]{tab_spanner_delim()}}). A select helper can also be used and, by
default, this is \code{everything()} (whereby all spanner column labels will be
removed).}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
If you would like to remove column spanner labels then the \code{rm_spanners()}
function can make this possible. Column spanner labels appear above the
column labels and can occupy several levels via stacking either though
\code{\link[=tab_spanner]{tab_spanner()}} or \code{\link[=tab_spanner_delim]{tab_spanner_delim()}}. Spanner column labels are
distinguishable and accessible by their ID values.

This function for removal is useful if you have received a \strong{gt} table
(perhaps through an API that returns \strong{gt} objects) but would prefer that
some or all of the column spanner labels be removed. This function is safe to
use even if there are no column spanner labels in the input \code{gt_tbl} object
so long as select helpers (such as the default \code{everything()}) are used
instead of explicit ID values.
}
\section{Function ID}{

10-3
}

\seealso{
Other part removal functions: 
\code{\link{rm_footnotes}()},
\code{\link{rm_header}()},
\code{\link{rm_source_notes}()},
\code{\link{rm_stubhead}()}
}
\concept{part removal functions}
