% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_remove.R
\name{rm_stubhead}
\alias{rm_stubhead}
\title{Remove the stubhead label}
\usage{
rm_stubhead(data)
}
\arguments{
\item{data}{A table object of class \code{gt_tbl}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
We can easily remove the stubhead label from a \strong{gt} table with
\code{rm_stubhead()}. The stubhead location only exists if there is a table stub
and the text in that cell is added through the \code{\link[=tab_stubhead]{tab_stubhead()}} function.

This function for removal is useful if you have received a \strong{gt} table
(perhaps through an API that returns \strong{gt} objects) but would prefer that
the table not contain any content in the stubhead. This function is safe to
use even if there is no stubhead label in the input \code{gt_tbl} object.
}
\section{Function ID}{

10-2
}

\seealso{
Other part removal functions: 
\code{\link{rm_footnotes}()},
\code{\link{rm_header}()},
\code{\link{rm_source_notes}()}
}
\concept{part removal functions}
