% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_rows.R
\name{row_group_order}
\alias{row_group_order}
\title{Modify the ordering of any row groups}
\usage{
row_group_order(data, groups)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{groups}{a vector of row group names, or, a numeric vector of indices
corresponding to the new ordering. Either vector must correspond to
assigned group names or the index positions. Also, either type of vector is
not required to have all of the row group names or available index
positions within it; any omitted values will be added to the end while
preserving the original ordering.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Modify the ordering of any row groups
}
\concept{row modification functions}
