% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rownames.R
\name{rownames_with}
\alias{rownames_with}
\title{Obtain rownames that fit a naming pattern}
\usage{
rownames_with(pattern)
}
\arguments{
\item{pattern}{a matching pattern for
targeting rownames.}
}
\description{
Obtain rownames that fit a naming pattern
}
