% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_add.R
\name{rows_add}
\alias{rows_add}
\title{Add one or more rows to a \strong{gt} table}
\usage{
rows_add(
  .data,
  ...,
  .list = list2(...),
  .before = NULL,
  .after = NULL,
  .n_empty = NULL
)
}
\arguments{
\item{.data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{...}{Expressions for the assignment of cell values to the new rows
by column name in \code{.data}. Name-value pairs, in the form of
\verb{<column> = <value vector>} will work, so long as the \verb{<column>} value
exists in the table. Two-sided formulas with column-resolving expressions
(e.g, \verb{<expr> ~ <value vector>}) can also be used, where the left-hand side
corresponds to selections of columns. The length of the longest vector in
\verb{<value vector>} determines how many new rows will be added. Single values
in \verb{<value vector>} will be repeated down in cases where there are multiple
rows to be added.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}

\item{.before, .after}{A single row-resolving expression or row index can be
given to either \code{.before} or \code{.after}. This specifies where the new rows
should be positioned among the existing rows in the input data table. If
nothing is provided for either argument the new rows will be placed at the
bottom.}

\item{.n_empty}{An option to add empty rows in lieu of rows containing data
that would otherwise be supplied to \code{...} or \code{.list}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
It's possible to add new rows to your table with \code{rows_add()} by supplying
the new row data through name value pairs.
}
\section{Function ID}{

6-4
}

\section{Function Introduced}{

\verb{0.9.1} (in development)
}

\seealso{
Other row addition/modification functions: 
\code{\link{grand_summary_rows}()},
\code{\link{row_group_order}()},
\code{\link{summary_rows}()}
}
\concept{row addition/modification functions}
