% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_hide.R
\name{rows_hide}
\alias{rows_hide}
\title{Hide one or more rows}
\usage{
rows_hide(data, rows)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}

\item{rows}{\emph{Rows to target}

\verb{<row-targeting expression>} // \strong{required}

The rows to hide in the output display table. Can either be a series of row
IDs provided in \code{c()}, a vector of row indices, or a select helper
function (e.g. \code{\link[=starts_with]{starts_with()}}, \code{\link[=ends_with]{ends_with()}}, \code{\link[=contains]{contains()}}, \code{\link[=matches]{matches()}},
\code{\link[=num_range]{num_range()}}, and \code{\link[=everything]{everything()}}). We can also use expressions to filter
down to the rows we need (e.g., \verb{[column_1] > 100 & [column_2] < 50}).}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
\code{rows_hide()} allows us to hide one or more rows from appearing in the final
output table. While it's possible and often desirable to omit rows from the
input table data before introduction to \code{\link[=gt]{gt()}}, there can be cases where the
data in certain rows is useful (as a row reference during formatting of other
rows or for column-based aggregations) but the final display of those rows is
not necessary.
}
\details{
The hiding of rows is internally a rendering directive, so, all rows that are
'hidden' are still accessible and useful in any expression provided to a
\code{columns} argument (for column-level operations). Furthermore, \code{rows_hide()}
(as with many \strong{gt} functions) can be placed anywhere in a pipeline of
\strong{gt} function calls (acting as a promise to hide rows when the timing is
right). However, there's perhaps greater readability when placing this call
closer to the end of such a pipeline.

\code{rows_hide()} quietly changes the visible state of a row (much like
\code{\link[=rows_unhide]{rows_unhide()}}) and doesn't yield warnings or messages when changing the
state of already-hidden rows.
}
\section{Examples}{


Let's use a small portion of the \code{\link{countrypops}} dataset to create a \strong{gt}
table. We can hide rows where the population is below a certain threshold
using \code{rows_hide()}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{countrypops |>
  dplyr::filter(
    country_name == "Egypt",
    year \%in\% 2017:2021
  ) |>
  gt(rowname_col = "year") |>
  fmt_integer(columns = population) |>
  rows_hide(rows = population < 100E6)
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_rows_hide_1.png" alt="This image of a table was generated from the first code example in the `rows_hide()` help file." style="width:100\%;">
}}

Using another \code{\link{countrypops}}-based \strong{gt} table, we can hide specific rows
by targeting them with their row labels (from the stub). Note that the
footnote placement still works correctly, skipping the hidden rows in its
mark indexing.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{countrypops |>
  dplyr::filter(
    country_name == "Brazil",
    year \%in\% 2017:2021
  ) |>
  gt(rowname_col = "year") |>
  fmt_integer(columns = population) |>
  rows_hide(rows = c("2018", "2020")) |>
  tab_footnote(
    footnote = "Population above 210,000,000.",
    locations = cells_body(
      columns = population,
      rows = population > 210E6
    )
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_rows_hide_2.png" alt="This image of a table was generated from the second code example in the `rows_hide()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

6-7
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other row addition/modification functions: 
\code{\link{grand_summary_rows}()},
\code{\link{row_group_order}()},
\code{\link{row_order}()},
\code{\link{rows_add}()},
\code{\link{rows_unhide}()},
\code{\link{summary_columns}()},
\code{\link{summary_rows}()}
}
\concept{row addition/modification functions}
