% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_hide.R
\name{rows_unhide}
\alias{rows_unhide}
\title{Unhide one or more rows}
\usage{
rows_unhide(data, rows)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}

\item{rows}{\emph{Rows to target}

\verb{<row-targeting expression>} // \emph{default:} \code{everything()}

The rows to unhide in the output display table. Can either be a series of
row IDs provided in \code{c()}, a vector of row indices, or a select helper
function (e.g. \code{\link[=starts_with]{starts_with()}}, \code{\link[=ends_with]{ends_with()}}, \code{\link[=contains]{contains()}}, \code{\link[=matches]{matches()}},
\code{\link[=num_range]{num_range()}}, and \code{\link[=everything]{everything()}}). We can also use expressions to filter
down to the rows we need (e.g., \verb{[column_1] > 100 & [column_2] < 50}).}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
\code{rows_unhide()} allows us to take one or more hidden rows (usually done
via \code{\link[=rows_hide]{rows_hide()}}) and make them visible in the final output table. This may
be important in cases where the user obtains a \code{gt_tbl} object with hidden
rows and there is motivation to reveal one or more of those.
}
\details{
The hiding and unhiding of rows is internally a rendering directive, so, all
rows that are 'hidden' are still accessible and useful in any expression
provided to a \code{columns} argument. The \code{rows_unhide()} function quietly
changes the visible state of a row (much like the \code{\link[=rows_hide]{rows_hide()}} function) and
doesn't yield warnings or messages when changing the state of already-visible
rows.
}
\section{Examples}{


Let's use a small portion of the \code{\link{countrypops}} dataset to create a \strong{gt}
table. We'll hide all rows but then unhide specific ones.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tab_1 <-
  countrypops |>
  dplyr::filter(
    country_name == "Japan",
    year \%in\% 2017:2021
  ) |>
  gt(rowname_col = "year") |>
  fmt_integer(columns = population) |>
  rows_hide(rows = everything())

tab_1
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_rows_unhide_1.png" alt="This image of a table was generated from the first code example in the `rows_unhide()` help file." style="width:100\%;">
}}

If we obtain the \code{tab_1} object without access to the code or source data,
and we want to reveal otherwise hidden rows, then \code{rows_unhide()} becomes
useful.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tab_1 |> rows_unhide(rows = c("2019", "2021"))
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_rows_unhide_2.png" alt="This image of a table was generated from the second code example in the `rows_unhide()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

6-8
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other row addition/modification functions: 
\code{\link{grand_summary_rows}()},
\code{\link{row_group_order}()},
\code{\link{row_order}()},
\code{\link{rows_add}()},
\code{\link{rows_hide}()},
\code{\link{summary_columns}()},
\code{\link{summary_rows}()}
}
\concept{row addition/modification functions}
