% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_cols_align_right}
\alias{set_cols_align_right}
\title{Set columns to be aligned right}
\usage{
set_cols_align_right(data, columns = NULL, types = NULL)
}
\description{
Set columns to be aligned right
}
