% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_cols_move_to_end}
\alias{set_cols_move_to_end}
\title{Move one or more columns to the end}
\usage{
set_cols_move_to_end(data, columns)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to move to
the right-most side of the table. The order in
which columns are provided will be preserved
(as is the case with the remaining columns).
Values provided that do not correspond to
column names will be disregarded.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Move one or more columns to the end
}
