% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_cols_remove}
\alias{set_cols_remove}
\title{Remove one or more columns}
\usage{
set_cols_remove(data, columns)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to remove from the table.
The order of the remaining columns will be preserved. Values
provided that do not correspond to column names will be disregarded.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Remove one or more columns
}
