% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_fmt}
\alias{set_fmt}
\title{Set a column format with a formatter function}
\usage{
set_fmt(data, columns, formatter)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an option specify which columns are to be
formatted.}

\item{formatter}{a formatting function.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a column format with a formatter function
}
