% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_fmt_date}
\alias{set_fmt_date}
\title{Format values as dates}
\usage{
set_fmt_date(data, columns = NULL, date_style = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to format.}

\item{date_style}{the date style to use.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as dates
}
