% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_fmt_datetime}
\alias{set_fmt_datetime}
\title{Format values as date-times}
\usage{
set_fmt_datetime(data, columns = NULL, date_style = NULL,
  time_style = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to format.}

\item{date_style}{the date style to use.}

\item{time_style}{the time style to use.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as date-times
}
