% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_fmt_scientific}
\alias{set_fmt_scientific}
\title{Format values to scientific notation}
\usage{
set_fmt_scientific(data, columns, decimals, sep_mark = "", dec_mark = ".",
  negative_style = "signed", drop0trailing = TRUE)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an option specify which columns are to be
formatted.}

\item{decimals}{an option to specify exactly the number
of decimal places to use.}

\item{sep_mark}{the mark to use as a separator between
groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}

\item{negative_style}{the formatting to use
for negative numbers. With \code{signed} (the default),
negative numbers will be shown with a negative sign.
Using \code{parens} will show the negative value in
parentheses. The \code{red} option will display the
number in red. Finally, \code{parens-red} will display
negative numbers as red and enclosed in parentheses.}

\item{drop0trailing}{a logical value that allows for
removal of trailing zeros (those redundant zeros after
the decimal mark).}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values to scientific notation
}
