% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_fmt_time}
\alias{set_fmt_time}
\title{Format values as times}
\usage{
set_fmt_time(data, columns = NULL, time_style = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an option specify which columns are to be
formatted.}

\item{time_style}{the time style to use.}
}
\description{
Format values as times
}
