% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_footnote}
\alias{set_footnote}
\title{Add a table footnote}
\usage{
set_footnote(data, footnote, location)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{footnote}{text to be used in the footnote.}

\item{location}{the cell or set of cells to be
associated with the footnote.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a footnote with a glyph attached to the
targeted cells, rows, or columns.
}
