% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_group}
\alias{set_group}
\title{Arrange a stub into blocks}
\usage{
set_group(data, group, rows, others = NULL)
}
\arguments{
\item{html_tbl}{a table object that is created using
the \code{gt()} function.}

\item{spanner}{the spanner heading name.}

\item{columns}{the columns to be components of
the spanner heading.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a group with a name and mappings to rows extant
in the table. This creates a stub block with group
headings and row captions.
}
\examples{
# Create a table based on `rock` where
# there are column headings grouped under
# spanner headings
gt(data = mtcars \%>\% tibble::rownames_to_column()) \%>\%
  set_group(
    group = "perimeter",
    rows = c("Mazda RX4", "Mazda RX4 Wag"))
}
