% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_heading}
\alias{set_heading}
\title{Add a table heading}
\usage{
set_heading(data, title, headnote = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{title}{text to be used in the table title.}

\item{headnote}{optional text to be used as the table's
headnote.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a title and optional headnote to the heading
part of the table.
}
