% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_source_note}
\alias{set_source_note}
\title{Add a source note citation}
\usage{
set_source_note(data, source_note)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{source_note}{text to be used in the source note.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a source note citation to the source note
part of the table.
}
