% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_spanner}
\alias{set_spanner}
\title{Arrange a boxhead into panels}
\usage{
set_spanner(data, spanner, columns)
}
\arguments{
\item{spanner}{the spanner heading name.}

\item{columns}{the columns to be components of
the spanner heading.}

\item{html_tbl}{a table object that is
created using the \code{gt()} function.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a spanner with a name and mappings
to columns extant in the table. This
creates a boxhead panel with spanner
headings and column headings.
}
\examples{
# Create a table based on `rock` where
# there are column headings grouped under
# spanner headings
gt(data = rock) \%>\%
  set_spanner(
    spanner = "perimeter",
    columns = c("peri", "shape"))
}
