% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{starts_with}
\alias{starts_with}
\title{Helper for resolving colnames or rownames that start with a prefix}
\usage{
starts_with(match, ignore.case = TRUE)
}
\arguments{
\item{match}{a string.}
}
\description{
See \code{tidyselect::\link[tidyselect]{select_helpers}} for details.
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{columns_with}},
  \code{\link{contains}}, \code{\link{ends_with}},
  \code{\link{everything}}, \code{\link{funs}},
  \code{\link{html}}, \code{\link{matches}},
  \code{\link{md}}, \code{\link{not_in_group}},
  \code{\link{one_of}}, \code{\link{pct}},
  \code{\link{px}}, \code{\link{rownames_with}},
  \code{\link{rows_where}}, \code{\link{vars}}
}
\concept{helper functions}
