% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stub_rows}
\alias{stub_rows}
\title{Helper for targeting stub rows}
\usage{
stub_rows(rows = NULL)
}
\arguments{
\item{rows}{either a vector of row names, a vector of rows indices, row
captions provided in \code{c()}, or a helper function focused on
selections. The select helper functions are: \code{\link{starts_with}()},
\code{\link{ends_with}()}, \code{\link{contains}()},
\code{\link{matches}()}, \code{\link{one_of}()}, and
\code{\link{everything}()}.}
}
\value{
a list object of class \code{stub_rows}.
}
\description{
This function is to be used to target individual stub row captions.
}
