% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{system_fonts}
\alias{system_fonts}
\title{Get a themed font stack that works well across systems}
\usage{
system_fonts(name)
}
\arguments{
\item{name}{A keyword from the set of \code{"system-ui"}, \code{"transitional"},
\code{"old-style"}, \code{"humanist"}, \code{"geometric-humanist"}, \code{"classical-humanist"},
\code{"neo-grotesque"}, \code{"monospace-slab-serif"}, \code{"monospace-code"},
\code{"industrial"}, \code{"rounded-sans"}, \code{"slab-serif"}, \code{"antique"}, \code{"didone"},
and \code{"handwritten"}.}
}
\value{
A character vector of font names.
}
\description{
A font stack can be obtained from \code{system_fonts()} using one or various
keywords such as \code{"system-ui"}, \code{"old-style"}, and \code{"humanist"} (there are 15
in total). These sets comprise a themed font family that has been tested to
work across a wide range of computer systems. This is useful when specifying
\code{font} values in the \code{\link[=cell_text]{cell_text()}} function (itself used in the \code{\link[=tab_style]{tab_style()}}
function). If using \code{\link[=opt_table_font]{opt_table_font()}} we can invoke this function in its
\code{stack} argument.
}
\section{Examples}{


Use \code{\link{sp500}} to create a \strong{gt} table with 10 rows. For the \code{date} column
and the column labels, let's use a different font stack (the \code{"industrial"}
one). The system fonts used in this particular stack are \code{"Bahnschrift"},
\code{"DIN Alternate"}, \code{"Franklin Gothic Medium"}, and \code{"Nimbus Sans Narrow"}
(the generic \code{"sans-serif-condensed"} and \code{"sans-serif"} are used if the
aforementioned fonts aren't available).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sp500 |>
  dplyr::slice(1:10) |>
  dplyr::select(-volume, -adj_close) |>
  gt() |>
  fmt_currency() |>
  tab_style(
    style = cell_text(
      font = system_fonts(name = "industrial"),
      size = px(18)
    ),
    locations = list(
      cells_body(columns = date),
      cells_column_labels()
    )
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_system_fonts_1.png" alt="This image of a table was generated from the first code example in the `system_fonts()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

8-29
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_footnotes}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_source_notes}()},
\code{\link{cells_stub_grand_summary}()},
\code{\link{cells_stub_summary}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()}
}
\concept{helper functions}
