% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_boxhead_panel}
\alias{tab_boxhead_panel}
\title{Arrange a boxhead into panels}
\usage{
tab_boxhead_panel(data, spanner, columns)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{spanner}{the spanner heading name.}

\item{columns}{the columns to be components of the
spanner heading.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a spanner with a name and mappings to columns extant
in the table. This creates a boxhead panel with spanner
headings and column headings.
}
\examples{
# Create a table based on `rock` where
# there are column headings grouped under
# spanner headings
gt(data = rock) \%>\%
  tab_boxhead_panel(
    spanner = "perimeter",
    columns = c("peri", "shape"))
}
