% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_create_modify.R
\name{tab_caption}
\alias{tab_caption}
\title{Add a table caption}
\usage{
tab_caption(data, caption)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{caption}{The table caption to use for cross-referencing in R Markdown,
Quarto, or \strong{bookdown}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Add a caption to a \strong{gt} table, which is handled specially for a table
within an R Markdown, Quarto, or \strong{bookdown} context. The addition of
captions makes tables cross-referencing across the containing document. The
caption location (i.e., top, bottom, margin) is handled at the document level
in each of these system.
}
\section{Function ID}{

2-9
}

\seealso{
Other part creation/modification functions: 
\code{\link{tab_footnote}()},
\code{\link{tab_header}()},
\code{\link{tab_options}()},
\code{\link{tab_row_group}()},
\code{\link{tab_source_note}()},
\code{\link{tab_spanner_delim}()},
\code{\link{tab_spanner}()},
\code{\link{tab_stub_indent}()},
\code{\link{tab_stubhead}()},
\code{\link{tab_style}()}
}
\concept{part creation/modification functions}
