% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_create.R
\name{tab_create}
\alias{tab_create}
\title{Create gt table object}
\usage{
tab_create(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
an object of class \code{html_table}.
}
\description{
Create a gt table object so that we can perform
styling transformations before transforming
the entire object to a display table. Using this
function is the first step in that process.
}
\examples{
# Create a table object using the
# `mtcars` dataset
tab <-
  tab_create(tbl = mtcars)

# The resulting object can be used
# in transformations
tab_2 <-
  tab \%>\%
  theme_striped() \%>\%
  cols_align_right() \%>\%
  fmt_number(
    columns = c("drat", "wt", "qsec"),
    decimals = 1
    )

# The object of this type can be
# displayed in the Viewer
tab_2
}
