% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_options.R
\name{tab_footnote_marks}
\alias{tab_footnote_marks}
\title{Modify the set of footnote marks}
\usage{
tab_footnote_marks(data, set = NULL)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{set}{The set of sequential figures or characters used to identify the
footnotes. We can either supply the keyword \code{"numbers"} (the default,
indicating that we want numeric glyphs), the keywords \code{"letters"} or
\code{"LETTERS"} (indicating that we want letters as glyphs, either lowercase or
uppercase), or, a vector of character values representing the series of
glyphs. A series of glyphs is recycled when its usage goes beyond the
length of the set. At each cycle, the glyphs are simply combined (e.g., \code{*}
-> \code{**} -> \code{***}).}
}
\description{
Alter the footnote marks for any footnotes that may be present in the table.
}
