% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_header}
\alias{tab_header}
\title{Add a table header}
\usage{
tab_header(data, title, subtitle = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{title, subtitle}{text to be used in the table title and, optionally, for
the table subtitle. We can optionally use the \code{\link{md}()} and
\code{\link{html}()} functions to style the text as Markdown or to retain
HTML elements in the text.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a table header part with a table title and optional subtitle.
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a header
# part to contain a title and subtitle
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_header(
      title = md("Data listing from **mtcars**"),
      subtitle = md("`mtcars` is an R dataset"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_footnote}},
  \code{\link{tab_options}}, \code{\link{tab_row_group}},
  \code{\link{tab_source_note}}, \code{\link{tab_spanner}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
