% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{tab_heading}
\alias{tab_heading}
\title{Add a table heading}
\usage{
tab_heading(html_tbl, title, headnote = NULL, table_number = NULL,
  font = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{title}{text to be used in the table title.}

\item{headnote}{text to be used as the table's
headnote.}

\item{table_number}{a number to use as the
table number.}

\item{font}{the name of the font to use for
the specified columns. This could be provided
as a vector of fonts where subsequent font names
provide fallbacks in the case that fonts are
not available.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a title, headnote, and table number
information to the heading part of the table.
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a heading
# to describe the table
tab_create(tbl = mtcars) \%>\%
  theme_striped() \%>\%
  tab_heading(
    title = md("Data listing from **mtcars**"),
    headnote = md("`mtcars` is an R dataset")
    )
}
