% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_info.R
\name{tab_info}
\alias{tab_info}
\title{Understand what's been set inside of a \strong{gt} table object}
\usage{
tab_info(data)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
It can become increasingly difficult to recall the ID values associated with
different labels in a \strong{gt} table. Further to this, there are also
situations where \strong{gt} will generate ID values on your behalf (e.g., with
\code{\link[=tab_spanner_delim]{tab_spanner_delim()}}, etc.) while ensuring that duplicate ID values aren't
produced. For the latter case, it is impossible to know what those ID values
are unless one were to carefully examine to correct component of the \code{gt_tbl}
object.

Because it's so essential to know these ID values for targeting purposes
(when styling with \code{\link[=tab_style]{tab_style()}}, adding footnote marks with
\code{\link[=tab_footnote]{tab_footnote()}}, etc.), the \code{tab_info()} function can help with all of this.
It summarizes (by location) all of the table's ID values and their associated
labels. The product is an informational \strong{gt} table, designed for easy
retrieval of the necessary values.
}
\section{Function ID}{

2-12
}

\seealso{
Other part creation/modification functions: 
\code{\link{tab_caption}()},
\code{\link{tab_footnote}()},
\code{\link{tab_header}()},
\code{\link{tab_options}()},
\code{\link{tab_row_group}()},
\code{\link{tab_source_note}()},
\code{\link{tab_spanner_delim}()},
\code{\link{tab_spanner}()},
\code{\link{tab_stub_indent}()},
\code{\link{tab_stubhead}()},
\code{\link{tab_style}()}
}
\concept{part creation/modification functions}
