% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_spanner}
\alias{tab_spanner}
\title{Add a spanner column label}
\usage{
tab_spanner(data, label, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns to be components of the spanner heading.}

\item{group}{the name to assign to the spanner heading.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a spanner column label by mapping it to columns already in the table.
This label that is placed above one or more column labels, spanning the width
of those columns.
}
\examples{
# Create a table based on `rock` where
# there are column headings grouped
# under a spanner column label
gt_tbl <-
  gt(data = rock) \%>\%
    tab_spanner(
      group = "perimeter",
      columns = c("peri", "shape"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_footnote}},
  \code{\link{tab_heading}}, \code{\link{tab_options}},
  \code{\link{tab_row_group}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
