% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{tab_spanner_headings}
\alias{tab_spanner_headings}
\title{Apply spanner headings}
\usage{
tab_spanner_headings(html_tbl, use_names = FALSE)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{use_names}{an options to use the the
naming convention of \code{[spanner_heading].[column_heading]}
to generate spanner headings that correctly
enclose column headings.}
}
\value{
an object of class \code{html_table}.
}
\description{
If column names in the \code{tbl} supplied
use the naming convention of
\code{[spanner_heading].[column_heading]},
we can invoke \code{tab_spanner_headings()}
with the \code{use_names} option set to
\code{TRUE}. This allows for parsing and
expansion of column headings into one or
more boxhead panels with spanner headings
and column headings.
}
