% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_stub_block}
\alias{tab_stub_block}
\title{Arrange a stub into blocks}
\usage{
tab_stub_block(data, group, rows, others = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{group}{the stub block group heading name.}

\item{rows}{the rows to be made components of the stub block.}

\item{others}{an optional group heading to use for any rows not part of a
stub block.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a group with a name and mappings to rows extant in the table. This
creates a stub block with group headings and row captions.
}
\examples{
# Create a table based on `mtcars` where
# there are group headings grouped inside
# stub blocks
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  tab_stub_block(
    group = "perimeter",
    rows = c("Mazda RX4", "Mazda RX4 Wag"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stubhead_caption}}
}
