% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{tab_stubhead_caption}
\alias{tab_stubhead_caption}
\title{Add caption text to the stubhead}
\usage{
tab_stubhead_caption(html_tbl, caption, alignment = "center")
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{caption}{the text to be used as the
stubhead caption.}

\item{alignment}{an option to modify the
alignment of the stubhead caption. By
default this is \code{center}. Other options
are \code{left} and \code{right}.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a caption to the stubhead of a table. If
a stub does not exist, no change will be made
but a message to this effect will be emitted.
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a caption
# to describe what is in the stub
tab_create(tbl = mtcars) \%>\%
  theme_striped() \%>\%
  tab_stubhead_caption(
    caption = md("car *make* and *model*")
    ) \%>\%
  cols_align_left(types = "character")
}
