% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style.R
\name{tab_style}
\alias{tab_style}
\title{Add custom styles to one or more cells}
\usage{
tab_style(data, style, locations)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{style}{a vector of styles to use. The \code{\link{cells_styles}()}
helper function can be used here to more easily generate valid styles.}

\item{locations}{the cell or set of cells to be associated with the style
Supplying an object with the \code{\link{cells_data}()} helper function is
a useful way to specify the cell that is associated with the style.
Additionally, we can supply a list with multiple calls to
\code{\link{cells_data}()} if we wish to combine multiple cell selections.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add custom styles to one or more cells
}
\examples{
# Add a style that is to be applied
# to a single table cell
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_style(
      style = cells_styles(bkgd_color = "steelblue"),
      locations = cells_data(
        columns = vars(hp),
        rows = c("Datsun 710", "Valiant")))
}
\seealso{
\code{\link{cells_styles}()} as a helper for defining custom styles
  and \code{\link{cells_data}()} as a useful helper function for targeting
  one or more data cells to be styled.

Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_footnote}},
  \code{\link{tab_header}}, \code{\link{tab_options}},
  \code{\link{tab_row_group}},
  \code{\link{tab_source_note}}, \code{\link{tab_spanner}},
  \code{\link{tab_stubhead_label}}
}
\concept{table-part creation/modification functions}
