% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{test_image}
\alias{test_image}
\title{Generate a path to a test image}
\usage{
test_image(type = c("png", "svg"))
}
\arguments{
\item{type}{the type of the image. Can either be \code{png} (the default) or
\code{svg}.}
}
\description{
Two test images are available within the \code{gt} package. Both contain the
same imagery (sized at 200px by 200px) but one is a PNG file while the other
is an SVG file.
}
