% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{text_case_match}
\alias{text_case_match}
\title{Perform whole text replacements which complete matches}
\usage{
text_case_match(.data, ..., .default = NULL, .locations = cells_body())
}
\arguments{
\item{.data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{...}{A sequence of two-sided formulas matching this general
construction: \code{old_text} ~ \code{new_text}. The left hand side (LHS) determines
which values to match on and it can be any length (allowing for \code{new_text}
to replace different values of \code{old_text}). The right hand side (RHS)
provides the replacement text (it must resolve to a single value of the
\code{character} class).}

\item{.default}{The replacement text to use when cell values aren't matched
by any of the LHS inputs. If \code{NULL}, the default, no replacement text will
be used.}

\item{.locations}{The cell or set of cells to be associated with the text
transformation. Only the \code{\link[=cells_body]{cells_body()}}, \code{\link[=cells_stub]{cells_stub()}},
\code{\link[=cells_column_labels]{cells_column_labels()}}, and \code{\link[=cells_row_groups]{cells_row_groups()}} helper functions can be
used here. We can enclose several of these calls within a \code{list()} if we
wish to make the transformation happen at different locations.}
}
\description{
Perform whole text replacements which complete matches
}
\section{Function ID}{

4-4
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other text transforming functions: 
\code{\link{text_case_when}()},
\code{\link{text_replace}()},
\code{\link{text_transform}()}
}
\concept{text transforming functions}
