% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{text_transform}
\alias{text_transform}
\title{Perform targeted text transformation with a function}
\usage{
text_transform(data, locations, fn)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{locations}{the cell or set of cells to be associated with the text
transformation. Supplying any of the \code{cells_*()} helper functions is a
useful way to target the location cells that are associated with the
footnote text. These helper functions are: \code{\link{cells_title}()},
\code{\link{cells_boxhead}()}, \code{\link{cells_group}()},
\code{\link{cells_stub}()}, \code{\link{cells_data}()}, and
\code{\link{cells_summary}()}. Please see the help article
\link{location_cells} for more information on how these helper functions
can be used. Additionally, we can enclose several \code{cells_*()} calls
within a \code{list()} if we wish to link the footnote text to different
types of locations (e.g., cell data values, stub group headings, the table
title, etc.).}

\item{fn}{the function to use for text transformation.}
}
\description{
Perform targeted text transformation with a function
}
\examples{
# Add a footnote that is in
# reference to a single table cell
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    text_transform(
      locations = cells_data(columns = vars(mpg)),
      fn = function(x) {
        ifelse(x > 20,
               paste(x, "(good)"),
               paste(x, "(worse)"))
      })
}
