% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tgt}
\alias{tgt}
\title{Helper for targeting a series of row labels or columns}
\usage{
tgt(...)
}
\arguments{
\item{...}{one or more column names, given as bare
column names.}
}
\description{
Helper for targeting a series of row labels or columns
}
