% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{theme_horizontal}
\alias{theme_horizontal}
\title{Use the 'horizontal' table theme}
\usage{
theme_horizontal(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply the 'horizontal' theme to the table. This
employs line styles with a minimal set of
horizontal rules for demarcation of the
boxhead and the field. This adds reasonably
thick horizontal rules enclosing the boxhead
and stubhead and a similar horizontal rule
below the field and stub.
}
\examples{
# Create a table object using the
# `mtcars` dataset and use the
# 'horizontal' theme
tab_create(tbl = mtcars) \%>\%
  theme_horizontal()
}
