% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{theme_schematic}
\alias{theme_schematic}
\title{Use the 'schematic' table theme}
\usage{
theme_schematic(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply the 'schematic' theme to the table.
This allows you to inspect the table's parts
since parts and subparts will be decorated with
with background fills and border effects.
}
\examples{
# Create a table object using the
# `mtcars` dataset and use the
# 'schematic' theme
tab_create(tbl = mtcars) \%>\%
  theme_schematic()
}
