% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{theme_striped}
\alias{theme_striped}
\title{Use the striped table theme}
\usage{
theme_striped(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply the striped theme to the table. This
adds striped rows, reasonably thick horizontal
rules enclosing the boxhead and stubhead and
a similar horizontal rule below the field and
stub.
}
\examples{
# Create a table object using the
# `mtcars` dataset and use the
# 'striped' theme
tab_create(tbl = mtcars) \%>\%
  theme_striped()
}
