% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{typst}
\alias{typst}
\title{Interpret input text as Typst-formatted text}
\usage{
typst(text)
}
\arguments{
\item{text}{\emph{Typst text}

\verb{scalar<character>} // \strong{required}

The text that is understood to be Typst markup, which is to be preserved in
the Typst output context.}
}
\value{
A character object of class \code{from_typst}. It's tagged as a Typst
fragment that is not to be sanitized.
}
\description{
For certain pieces of text (like in column labels or table headings) we may
want to express them as raw Typst markup. The \code{typst()} function will guard
the input from being escaped in the Typst output context.
}
\section{Function ID}{

8-32
}

\section{Function Introduced}{

\code{v0.13.0} (2026-03-26)
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{latex}()},
\code{\link{md}()},
\code{\link{nanoplot_options}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{row_group}()},
\code{\link{stub}()},
\code{\link{system_fonts}()},
\code{\link{unit_conversion}()}
}
\concept{helper functions}
