% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Get a conversion factor across two measurement units of a given class}
\usage{
unit_conversion(from, to)
}
\arguments{
\item{from}{\emph{Units for the input value}

\verb{scalar<character>} // \strong{required}

The keyword representing the units for the value that requires unit
conversion. In the case where the value has units of miles, the necessary
input is \code{"length.mile"}.}

\item{to}{\emph{Desired units for the value}

\verb{scalar<character>} // \strong{required}

The keyword representing the target units for the value with units defined
in \code{from}. In the case where input value has units of miles and we would
rather want the value to be expressed as kilometers, the \code{to} value should
be \code{"length.kilometer"}.}
}
\value{
A single numerical value.
}
\description{
The \code{unit_conversion()} helper function gives us a conversion factor for
transforming a value from one form of measurement units to a target form.
For example if you have a length value that is expressed in miles you could
transform that value to one in kilometers through multiplication of the value
by the conversion factor (in this case \code{1.60934}).
}
\section{Function ID}{

8-7
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{nanoplot_options}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()},
\code{\link{system_fonts}()}
}
\concept{helper functions}
