% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{web_image}
\alias{web_image}
\title{Helper function for adding an image from the web}
\usage{
web_image(url, height = 30)
}
\arguments{
\item{url}{a url that resolves to an image file.}

\item{height}{the absolute height (px) of the image in the table cell.}
}
\value{
a character object with an HTML fragment that can be placed inside of
a cell.
}
\description{
Add a web image inside a table with this helper function.
}
\examples{
\dontrun{
# Create an HTML fragment that
# contains an image
img_file_html <-
  web_image(
    file = "https://.../web_image.png")
}
}
