% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rtf.R
\name{write_rtf}
\alias{write_rtf}
\title{Save a gt table object as an RTF file}
\usage{
write_rtf(data, file)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{file}{a filename to use for writing out the RTF file.}
}
\description{
Take a \code{gt_tbl} table object and render it to an RTF file with the
extension \code{.rtf}.
}
\examples{
\dontrun{
# Create a `gt` table using the `sleep`
# dataset and then write the table to
# an RTF file
gt(data = sleep) \%>\%
  write_rtf(file = "sleep.rtf")
}
}
\concept{table export functions}
