% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vec.R
\name{x_date}
\alias{x_date}
\title{Format a vector as date values}
\usage{
x_date(x, date_style = 2, pattern = "{x}", output = c("html", "latex", "rtf"))
}
\arguments{
\item{x}{A numeric vector.}

\item{date_style}{The date style to use. Supply a number (from \code{1} to \code{14})
that corresponds to the preferred date style, or, provide a named date
style (\code{"wday_month_day_year"}, \code{"m_day_year"}, \code{"year.mn.day"}, etc.). Use
\code{\link[=info_date_style]{info_date_style()}} to see the different numbered and named date presets.}

\item{pattern}{A formatting pattern that allows for decoration of the
formatted value. The value itself is represented by \code{{x}} and all other
characters are taken to be string literals.}

\item{output}{The output style of the resulting character vector. This can
either be \code{"html"} (the default), \code{"latex"}, or \code{"rtf"}.}
}
\value{
A character vector.
}
\description{
Format vector values to date values using one of fourteen presets. Input can
be in the form of \code{POSIXt} (i.e., date-times), the \code{Date} type, or
\code{character} (must be in the ISO 8601 form of \verb{YYYY-MM-DD HH:MM:SS} or
\code{YYYY-MM-DD}).

We can simply apply a preset date style to format the dates. The following
date styles are available for use (all using the input date of \code{2000-02-29}
in the example output dates):
\enumerate{
\item \code{"iso"}: \code{2000-02-29}
\item \code{"wday_month_day_year"}: \verb{Tuesday, February 29, 2000}
\item \code{"wd_m_day_year"}: \verb{Tue, Feb 29, 2000}
\item \code{"wday_day_month_year"}: \verb{Tuesday 29 February 2000}
\item \code{"month_day_year"}: \verb{February 29, 2000}
\item \code{"m_day_year"}: \verb{Feb 29, 2000}
\item \code{"day_m_year"}: \verb{29 Feb 2000}
\item \code{"day_month_year"}: \verb{29 February 2000}
\item \code{"day_month"}: \verb{29 February}
\item \code{"year"}: \code{2000}
\item \code{"month"}: \code{February}
\item \code{"day"}: \code{29}
\item \code{"year.mn.day"}: \code{2000/02/29}
\item \code{"y.mn.day"}: \code{00/02/29}
}

We can use the \code{\link[=info_date_style]{info_date_style()}} function for a useful reference on all of
the possible inputs to \code{date_style}.
}
\section{Function ID}{

14-9
}

\seealso{
Other Vector Formatters: 
\code{\link{x_bytes}()},
\code{\link{x_currency}()},
\code{\link{x_datetime}()},
\code{\link{x_engineering}()},
\code{\link{x_fraction}()},
\code{\link{x_integer}()},
\code{\link{x_markdown}()},
\code{\link{x_number}()},
\code{\link{x_percent}()},
\code{\link{x_scientific}()},
\code{\link{x_time}()}
}
\concept{Vector Formatters}
