% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vec.R
\name{x_engineering}
\alias{x_engineering}
\title{Format a vector as values in engineering notation}
\usage{
x_engineering(
  x,
  decimals = 2,
  drop_trailing_zeros = FALSE,
  scale_by = 1,
  pattern = "{x}",
  sep_mark = ",",
  dec_mark = ".",
  force_sign = FALSE,
  locale = NULL,
  output = c("html", "latex", "rtf")
)
}
\arguments{
\item{x}{A numeric vector.}

\item{decimals}{An option to specify the exact number of decimal places to
use. The default number of decimal places is \code{2}.}

\item{drop_trailing_zeros}{A logical value that allows for removal of
trailing zeros (those redundant zeros after the decimal mark).}

\item{scale_by}{A value to scale the input. The default is \code{1.0}. All numeric
values will be multiplied by this value first before undergoing formatting.}

\item{pattern}{A formatting pattern that allows for decoration of the
formatted value. The value itself is represented by \code{{x}} and all other
characters are taken to be string literals.}

\item{sep_mark}{The mark to use as a separator between groups of digits
(e.g., using \code{sep_mark = ","} with \code{1000} would result in a formatted value
of \verb{1,000}).}

\item{dec_mark}{The character to use as a decimal mark (e.g., using
\code{dec_mark = ","} with \code{0.152} would result in a formatted value of
\verb{0,152}).}

\item{force_sign}{Should the positive sign be shown for positive values
(effectively showing a sign for all values except zero)? If so, use \code{TRUE}
for this option. The default is \code{FALSE}, where only negative numbers will
display a minus sign.}

\item{locale}{An optional locale ID that can be used for formatting the value
according the locale's rules. Examples include \code{"en_US"} for English
(United States) and \code{"fr_FR"} for French (France). The use of a valid
locale ID will override any values provided in \code{sep_mark} and \code{dec_mark}.
We can use the \code{\link[=info_locales]{info_locales()}} function as a useful reference for all of
the locales that are supported.}

\item{output}{The output style of the resulting character vector. This can
either be \code{"html"} (the default), \code{"latex"}, or \code{"rtf"}.}
}
\value{
A character vector.
}
\description{
With numeric values in a vector, we can perform formatting so that the input
values are rendered into engineering notation within the output character
vector. The following major options are available:
\itemize{
\item decimals: choice of the number of decimal places, option to drop
trailing zeros, and a choice of the decimal symbol
\item digit grouping separators: choice of separator symbol
\item scaling: we can choose to scale targeted values by a multiplier value
\item pattern: option to use a text pattern for decoration of the formatted
values
\item locale-based formatting: providing a locale ID will result in
formatting specific to the chosen locale
}
}
\section{Function ID}{

14-4
}

\seealso{
Other Vector Formatters: 
\code{\link{x_bytes}()},
\code{\link{x_currency}()},
\code{\link{x_datetime}()},
\code{\link{x_date}()},
\code{\link{x_fraction}()},
\code{\link{x_integer}()},
\code{\link{x_markdown}()},
\code{\link{x_number}()},
\code{\link{x_percent}()},
\code{\link{x_scientific}()},
\code{\link{x_time}()}
}
\concept{Vector Formatters}
