% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vec.R
\name{x_markdown}
\alias{x_markdown}
\title{Format a vector containing Markdown text}
\usage{
x_markdown(x, output = c("html", "latex", "rtf"))
}
\arguments{
\item{x}{A numeric vector.}

\item{output}{The output style of the resulting character vector. This can
either be \code{"html"} (the default), \code{"latex"}, or \code{"rtf"}.}
}
\value{
A character vector.
}
\description{
Any Markdown-formatted text in the input vector will be transformed to the
appropriate output type.
}
\section{Function ID}{

14-12
}

\seealso{
Other vector formatting functions: 
\code{\link{x_bytes}()},
\code{\link{x_currency}()},
\code{\link{x_datetime}()},
\code{\link{x_date}()},
\code{\link{x_engineering}()},
\code{\link{x_fraction}()},
\code{\link{x_integer}()},
\code{\link{x_number}()},
\code{\link{x_percent}()},
\code{\link{x_scientific}()},
\code{\link{x_time}()}
}
\concept{vector formatting functions}
