% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{add_body_style}
\alias{add_body_style}
\title{Add inline CSS styles to the body of an HTML table object}
\usage{
add_body_style(html_tbl, property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{build_html_table()}
function.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This could be provided as a single-length
character string representing to exact string
to be inserted, or, as a vector of values where
the individual values for the property will
be transformed to a space-separated string.}
}
\value{
an HTML table object.
}
\description{
Modify the style of the body for a table
within an HTML table object. This is part of
an intermediate set of step in the
\code{build_html_table()} ->
\code{add_..._style()} -> \code{emit_html()}
pattern.
}
