% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{add_stubhead_caption}
\alias{add_stubhead_caption}
\title{Add caption text to the stubhead}
\usage{
add_stubhead_caption(html_tbl, text)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}

\item{text}{the text to be used as the stubhead.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a caption to the stubhead of a table. If
a stub does not exist, no change will be made
but a message to this effect will be emitted.
}
