% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.tags.gt_tbl}
\alias{as.tags.gt_tbl}
\title{Knit print the table as RTF}
\usage{
as.tags.gt_tbl(x, ...)
}
\arguments{
\item{x}{an object of class \code{gt_rtf}.}
}
\description{
This facilitates printing of the RTF table within a knitr code chunk.
}
\keyword{internal}
