% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{as_latex}
\alias{as_latex}
\title{Output a gt table object as LaTeX}
\usage{
as_latex(data)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}
}
\description{
Take a \code{gt_tbl} table object and emit LaTeX.
}
\examples{
# Use `gtcars` to create a gt table;
# add a header and then export as
# LaTeX code (use `as.character()`
# to get just the code and no LaTeX
# dependencies for R Markdown)
tab_1 <-
  gtcars \%>\%
  dplyr::select(mfr, model, msrp) \%>\%
  dplyr::slice(1:5) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_latex()
}
\seealso{
Other table export functions: \code{\link{as_raw_html}},
  \code{\link{as_rtf}}, \code{\link{extract_summary}}
}
\concept{table export functions}
