% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_latex}
\alias{as_latex}
\title{Output a \strong{gt} object as LaTeX}
\usage{
as_latex(data)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}
}
\description{
Get the LaTeX content from a \code{gt_tbl} object as a \code{knit_asis} object. This
object contains the LaTeX code and attributes that serve as LaTeX
dependencies (i.e., the LaTeX packages required for the table). Using
\code{as.character()} on the created object will result in a single-element vector
containing the LaTeX code.
}
\details{
LaTeX packages required to generate tables are:
booktabs, caption, longtable, colortbl, array.

In the event packages are not automatically added during the render phase
of the document, please create and include a style file to load them.

Inside the document's YAML metadata, please include:

\preformatted{
output:
  pdf_document: # Change to appropriate LaTeX template
    includes:
      in_header: 'gt_packages.sty'
}

The \code{gt_packages.sty} file would then contain the listed dependencies above:

\preformatted{
  \usepackage{booktabs, caption, longtable, colortbl, array}
}
}
\section{Examples}{


Use a subset of the \code{\link{gtcars}} dataset to create a \strong{gt} table. Add a header
with \code{\link[=tab_header]{tab_header()}} and then export the table as LaTeX code using the
\code{as_latex()} function.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tab_latex <-
  gtcars |>
  dplyr::select(mfr, model, msrp) |>
  dplyr::slice(1:5) |>
  gt() |>
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) |>
  as_latex()
}\if{html}{\out{</div>}}

What's returned is a \code{knit_asis} object, which makes it easy to include in R
Markdown documents that are knit to PDF. We can use \code{as.character()} to get
just the LaTeX code as a single-element vector.
}

\section{Function ID}{

13-3
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other table export functions: 
\code{\link{as_raw_html}()},
\code{\link{as_rtf}()},
\code{\link{as_word}()},
\code{\link{extract_cells}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{table export functions}
