% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raw_html.R
\name{as_raw_html}
\alias{as_raw_html}
\title{Get the HTML content of a \pkg{gt} table}
\usage{
as_raw_html(data, inline_css = TRUE)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{inline_css}{an option to supply styles to table elements as inlined CSS
styles.}
}
\description{
Get the HTML content from a \code{gt_tbl} object as a single-element
character vector. By default, the generated HTML will have inlined styles,
where CSS styles (that were previously contained in CSS rule sets external to
the \code{<table> element}) are included as \code{style} attributes in the
HTML table's tags. This option is preferable when using the output HTML table
in an emailing context.
}
\examples{
# Use `gtcars` to create a gt table;
# add a header and then export as
# HTML code with CSS inlined
tab_html <-
  gtcars \%>\%
  dplyr::select(mfr, model, msrp) \%>\%
  dplyr::slice(1:5) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_raw_html()

# `tab_html` is a single-element vector
# containing inlined HTML for the table;
# it has only the `<table>...</table>` part
# so it's not a complete HTML document but
# rather an HTML fragment
tab_html \%>\%
  substr(1, 700) \%>\%
  cat()

}
\seealso{
Other table export functions: \code{\link{as_latex}},
  \code{\link{as_rtf}}, \code{\link{extract_summary}}
}
\concept{table export functions}
