% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rtf.R
\name{as_rtf}
\alias{as_rtf}
\title{Save a gt table as an RTF file}
\usage{
as_rtf(data)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}
}
\description{
Take a \code{gt_tbl} table object and render it to an RTF file with the
extension \code{.rtf}.
}
\examples{
\dontrun{
# Create a `gt` table using the `sleep`
# dataset and then output the table to
# a raw RTF object
gt(data = sleep) \%>\%
  as_rtf(file = "sleep.rtf")
}
}
\seealso{
Other table export functions: \code{\link{as_latex}},
  \code{\link{as_raw_html}}, \code{\link{extract_summary}},
  \code{\link{gt_pdf}}, \code{\link{gt_png}}
}
\concept{table export functions}
