% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rtf.R
\name{as_rtf}
\alias{as_rtf}
\title{Output a \pkg{gt} object as RTF}
\usage{
as_rtf(data)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}
}
\description{
Get the RTF content from a \code{gt_tbl} object as as a single-element
character vector. This object can be used with \code{writeLines()} to
generate a valid .rtf file that can be opened by RTF readers.
}
\examples{
# Use `gtcars` to create a gt table;
# add a header and then export as
# RTF code
tab_rtf <-
  gtcars \%>\%
  dplyr::select(mfr, model) \%>\%
  dplyr::slice(1:2) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_rtf()

# `tab_rtf` is a single element character
# vector
tab_rtf \%>\% cat()

}
\seealso{
Other table export functions: \code{\link{as_latex}},
  \code{\link{as_raw_html}}, \code{\link{extract_summary}}
}
\concept{table export functions}
