% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as_word}
\alias{as_word}
\title{Output a \strong{gt} object as Word}
\usage{
as_word(
  data,
  align = "center",
  caption_location = c("top", "bottom", "embed"),
  caption_align = "left",
  split = FALSE,
  keep_with_next = TRUE,
  autonum = TRUE
)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}

\item{align}{\emph{Table alignment}

\verb{scalar<character>} // \emph{default:} \code{"center"}

An option for table alignment. Can either be \code{"center"}, \code{"left"}, or
\code{"right"}.}

\item{caption_location}{\emph{Caption location}

\verb{singl-kw:[top|bottom|embed]} // \emph{default:} \code{"top"}

Determines where the caption should be positioned. This can either be
\code{"top"}, \code{"bottom"}, or \code{"embed"}.}

\item{caption_align}{\emph{Caption alignment}

Determines the alignment of the caption. This is
either \code{"left"} (the default), \code{"center"}, or \code{"right"}. This option is
only used when \code{caption_location} is not set as \code{"embed"}.}

\item{split}{\emph{Allow splitting of a table row across pages}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

A logical value that indicates whether to activate the Word option
\verb{Allow row to break across pages}.}

\item{keep_with_next}{\emph{Keeping rows together}

\verb{scalar<logical>} // \emph{default:} \code{TRUE}

A logical value that indicates whether a table should use Word option
\verb{Keep rows together}.}

\item{autonum}{Automatic Table Numbering

\verb{scalar<logical>} // \emph{default:} \code{TRUE}

A logical value that indicates whether a table should use Words built-in
auto table numbering option in the caption.
\verb{Automatic Table Numbering}.}
}
\description{
Get the Open Office XML table tag content from a \code{gt_tbl} object as a
single-element character vector.
}
\section{Examples}{


Use a subset of the \code{\link{gtcars}} dataset to create a \strong{gt} table. Add a header
with \code{\link[=tab_header]{tab_header()}} and then export the table as OOXML code for Word using
\code{as_word()}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tab_rtf <-
  gtcars |>
  dplyr::select(mfr, model) |>
  dplyr::slice(1:2) |>
  gt() |>
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) |>
  as_word()
}\if{html}{\out{</div>}}
}

\section{Function ID}{

13-5
}

\section{Function Introduced}{

\code{v0.7.0} (August 25, 2022)
}

\seealso{
Other table export functions: 
\code{\link{as_gtable}()},
\code{\link{as_latex}()},
\code{\link{as_raw_html}()},
\code{\link{as_rtf}()},
\code{\link{extract_body}()},
\code{\link{extract_cells}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{table export functions}
