% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorizing.R
\name{bg_color_fill_manual}
\alias{bg_color_fill_manual}
\title{Set background data cell colors using provided colors}
\usage{
bg_color_fill_manual(data, column, values, colors, alpha = 1,
  na_color = "darkgrey")
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{column}{the column wherein changes to cell data background colors
should occur.}

\item{values}{the cell values to which the background \code{colors} should be
applied. The length of \code{values} must match that of \code{colors} since
they are considered to be one-to-one mappings of value to color for the
target \code{column}.}

\item{colors}{a vector of colors to use for each of the provided
\code{values}. Each color value provided must either be a color name (in
the set of colors provided by \code{grDevices::colors()}) or hexadecimal
strings in the form of "#RRGGBB" or "#RRGGBBAA".}

\item{alpha}{a fixed alpha transparency value that will be applied to all of
the \code{colors} provided.}

\item{na_color}{the default color for any unmapped data cells in the target
\code{column}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set background data cell colors using provided colors
}
