% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_rows.R
\name{blocks_arrange}
\alias{blocks_arrange}
\title{Modify the ordering of the stub block groups}
\usage{
blocks_arrange(data, groups)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{groups}{a vector of stub block group heading names.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Modify the ordering of the stub block groups
}
