% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{build_html_table}
\alias{build_html_table}
\title{Create an HTML table object that can be styled}
\usage{
build_html_table(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
an HTML table object.
}
\description{
Create an HTML table object such that we can
perform styling transformations before transforming
the entire object to an HTML fragment (for
inclusion into the email message body). This is
the first step in the \code{build_html_table()} ->
\code{add_column_style()} -> \code{emit_html()}
pattern.
}
