% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_selection_helpers.R
\name{cell}
\alias{cell}
\title{Define a table cell}
\usage{
cell(row, column)
}
\arguments{
\item{row}{an integer representing the table
row to target (as part of the \code{row}-\code{column}
dyad).}
}
\description{
Define a table cell
}
