% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cell_borders}
\alias{cell_borders}
\title{Helper for defining custom borders for table cells}
\usage{
cell_borders(sides = "all", color = "#000000", style = "solid",
  weight = px(1))
}
\arguments{
\item{sides}{The border sides to be modified. Options include \code{"left"},
\code{"right"}, \code{"top"}, and \code{"bottom"}. For all borders surrounding the
selected cells, we can use the \code{"all"} option.}

\item{color}{The border color. The default \code{"#000000"} (black) will be used
as a default.}

\item{style}{The border style. Can be one of either \code{"solid"} (the default),
\code{"dashed"}, or \code{"dotted"}.}

\item{weight}{The weight of the border lines. The default value is \code{"1px"}.}
}
\description{
The \code{cell_borders()} helper function is to be used with the \code{\link[=tab_style]{tab_style()}}
function, which itself allows for the setting of custom styles to one or more
cells. Specifically, the call to \code{cell_borders()} should be bound to the
\code{styles} argument of \code{\link[=tab_style]{tab_style()}}. The \code{selection} argument is where we
define which borders should be modified (e.g., \code{"left"}, \code{"right"}, etc.).
With that selection, the \code{color}, \code{style}, and \code{weight} of the selected
borders can then be modified.
}
\seealso{
Other helper functions: \code{\link{cell_fill}},
  \code{\link{cell_text}}, \code{\link{cells_styles}},
  \code{\link{escape_latex}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}},
  \code{\link{random_id}}
}
\concept{helper functions}
