% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cell_borders}
\alias{cell_borders}
\title{Helper for defining custom borders for table cells}
\usage{
cell_borders(sides = "all", color = "#000000", style = "solid", weight = px(1))
}
\arguments{
\item{sides}{\emph{Border sides}

\verb{vector<character>} // \emph{default:} \code{"all"}

The border sides to be modified. Options include \code{"left"}, \code{"right"},
\code{"top"}, and \code{"bottom"}. For all borders surrounding the selected cells, we
can use the \code{"all"} option.}

\item{color}{\emph{Border color}

\verb{scalar<character>|NULL} // \emph{default:} \code{"#000000"}

The border \code{color} can be defined with a color name or with a hexadecimal
color code. The default \code{color} value is \code{"#000000"} (black). Borders for
any defined \code{sides} can be removed by supplying \code{NULL} here.}

\item{style}{\emph{Border line style}

\verb{scalar<character>|NULL} // \emph{default:} \code{"solid"}

The border \code{style} can be one of either \code{"solid"} (the default),
\code{"dashed"}, \code{"dotted"}, \code{"hidden"}, or \code{"double"}. Borders for any defined
\code{sides} can be removed by supplying \code{NULL} here.}

\item{weight}{\emph{Border weight}

\verb{scalar<character>|NULL} // \emph{default:} \code{px(1)}

The default value for \code{weight} is \code{"1px"} and higher values will become
more visually prominent. Borders for any defined \code{sides} can be removed by
supplying \code{NULL} to any of \code{color}, \code{style}, or \code{weight}.}
}
\value{
A list object of class \code{cell_styles}.
}
\description{
\code{cell_borders()} is to be used with \code{\link[=tab_style]{tab_style()}}, which itself allows for
the setting of custom styles to one or more cells. Specifically, the call to
\code{cell_borders()} should be bound to the \code{styles} argument of \code{\link[=tab_style]{tab_style()}}.
The \code{sides} argument is where we define which borders should be modified
(e.g., \code{"left"}, \code{"right"}, etc.). With that selection, the \code{color}, \code{style},
and \code{weight} of the selected borders can then be modified.
}
\section{Examples}{


We can add horizontal border lines for all table body rows in a \strong{gt} table
based on the \code{\link{exibble}} dataset. For this, we need to use \code{\link[=tab_style]{tab_style()}}
(targeting all cells in the table body with \code{\link[=cells_body]{cells_body()}}) in conjunction
with \code{cell_borders()} in the \code{style} argument. Both top and bottom borders
will be added as \code{"solid"} and \code{"red"} lines with a line width of 1.5 px.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  gt() |>
  tab_style(
    style = cell_borders(
      sides = c("top", "bottom"),
      color = "red",
      weight = px(1.5),
      style = "solid"
    ),
    locations = cells_body()
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_cell_borders_1.png" alt="This image of a table was generated from the first code example in the `cell_borders()` help file." style="width:100\%;">
}}

It's possible to incorporate different horizontal and vertical (\code{"left"} and
\code{"right"}) borders at several different locations. This uses multiple
\code{cell_borders()} and \code{\link[=cells_body]{cells_body()}} calls within their own respective lists.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  gt() |>
  tab_style(
    style = list(
      cell_borders(
        sides = c("top", "bottom"),
        color = "#FF0000",
        weight = px(2)
      ),
      cell_borders(
        sides = c("left", "right"),
        color = "#0000FF",
        weight = px(2)
      )
    ),
    locations = list(
      cells_body(
        columns = num,
        rows = is.na(num)
      ),
      cells_body(
        columns = currency,
        rows = is.na(currency)
      )
    )
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_cell_borders_2.png" alt="This image of a table was generated from the second code example in the `cell_borders()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

8-27
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{latex}()},
\code{\link{md}()},
\code{\link{nanoplot_options}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{row_group}()},
\code{\link{stub}()},
\code{\link{system_fonts}()},
\code{\link{unit_conversion}()}
}
\concept{helper functions}
