% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cell_fill}
\alias{cell_fill}
\title{Helper for defining custom fills for table cells}
\usage{
cell_fill(color = "#D3D3D3", alpha = NULL)
}
\arguments{
\item{color}{The fill color. If nothing is provided, then \code{"#D3D3D3"} (light
gray) will be used as a default.}

\item{alpha}{An optional alpha transparency value for the \code{color} as single
value in the range of \code{0} (fully transparent) to \code{1} (fully opaque). If not
provided the fill color will either be fully opaque or use alpha
information from the color value if it is supplied in the #RRGGBBAA format.}
}
\description{
The \code{cell_fill()} helper function is to be used with the \code{\link[=tab_style]{tab_style()}}
function, which itself allows for the setting of custom styles to one or more
cells. Specifically, the call to \code{cell_fill()} should be bound to the
\code{styles} argument of \code{\link[=tab_style]{tab_style()}}.
}
\seealso{
Other helper functions: 
\code{\link{cell_borders}()},
\code{\link{cell_text}()},
\code{\link{cells_styles}()},
\code{\link{currency}()},
\code{\link{escape_latex}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{location_cells}},
\code{\link{location_colors}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()}
}
\concept{helper functions}
