% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cells_group}
\alias{cells_group}
\title{Location helper for targeting row groups (deprecated)}
\usage{
cells_group(groups = everything())
}
\arguments{
\item{groups}{\emph{Specification of row group IDs}

\verb{<row-group-targeting expression>} // \emph{default:} \code{everything()}

The row groups to which targeting operations are constrained. Can either be
a series of row group ID values provided in \code{\link[=c]{c()}} or a select helper
function. Examples of select helper functions include \code{\link[=starts_with]{starts_with()}},
\code{\link[=ends_with]{ends_with()}}, \code{\link[=contains]{contains()}}, \code{\link[=matches]{matches()}}, \code{\link[=one_of]{one_of()}}, \code{\link[=num_range]{num_range()}}, and
\code{\link[=everything]{everything()}}.}
}
\description{
Location helper for targeting row groups (deprecated)
}
\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\keyword{internal}
