% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cells_styles}
\alias{cells_styles}
\title{Helper for defining custom styles for table cells}
\usage{
cells_styles(bkgd_color = NULL, text_color = NULL, text_font = NULL,
  text_size = NULL, text_align = NULL, text_style = NULL,
  text_weight = NULL, text_stretch = NULL, text_indent = NULL,
  text_decorate = NULL, text_transform = NULL)
}
\arguments{
\item{bkgd_color}{the background color of the cell.}

\item{text_color}{the text color.}

\item{text_font}{the font or collection of fonts (subsequent font names are)
used as fallbacks.}

\item{text_size}{the size of the font. Can be provided as a number that is
assumed to represent \code{px} values (or could be wrapped in the
\code{\link{px}()}) helper function. We can also use one of the following
absolute size keywords: \code{xx-small}, \code{x-small}, \code{small},
\code{medium}, \code{large}, \code{x-large}, or \code{xx-large}.}

\item{text_align}{the text alignment. Can be one of either \code{"center"},
\code{"left"}, \code{"right"}, or \code{"justify"}.}

\item{text_style}{the text style. Can be one of either \code{"center"},
\code{"normal"}, \code{"italic"}, or \code{"oblique"}.}

\item{text_weight}{the weight of the font. Can be a text-based keyword such
as \code{"normal"}, \code{"bold"}, \code{"lighter"}, \code{"bolder"}, or, a
numeric value between \code{1} and \code{1000}, inclusive. Note that only
variable fonts may support the numeric mapping of weight.}

\item{text_stretch}{allows for text to either be condensed or expanded. We
can use the following text-based keywords to describe the degree of
condensation/expansion: \code{ultra-condensed}, \code{extra-condensed},
\code{condensed}, \code{semi-condensed}, \code{normal},
\code{semi-expanded}, \code{expanded}, \code{extra-expanded}, and
\code{ultra-expanded}. Alternatively, we can supply percentage values from
\code{0\%} to \code{200\%}, inclusive. Negative percentage values are not
allowed.}

\item{text_indent}{the indentation of the text.}

\item{text_decorate}{allows for text decoration effect to be applied. Here,
we can use \code{"overline"}, \code{"line-through"}, or \code{"underline"}.}

\item{text_transform}{allows for the transformation of text. Options are
\code{"uppercase"}, \code{"lowercase"}, or \code{"capitalize"}.}
}
\value{
a character vector containing formatted styles.
}
\description{
This helper function is to be used with the \code{\link{tab_style}()}
function, which itself allows for the setting of custom styles to one or more
cells. We can also define several styles within a single call of
\code{cells_styles} and \code{\link{tab_style}()} will reliably apply those
styles to the targeted element.
}
\seealso{
Other helper functions: \code{\link{escape_latex}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}}
}
\concept{helper functions}
