% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{color_bin}
\alias{color_bin}
\title{Colorizing helper function that colors by bin membership}
\usage{
color_bin(palette, domain = NULL, bins = 7, pretty = TRUE,
  na_color = "#808080")
}
\arguments{
\item{palette}{the colors or color function that values will be mapped to.}

\item{domain}{the possible values that should be mapped.}

\item{bins}{either a numeric vector of two or more unique cut points, or, a
single number (>= 2) giving the number of intervals into which the
\code{domain} values are to be cut.}

\item{pretty}{an option for whether to create pretty breaks based on the
specified number of \code{bins}. By default, this is set to \code{TRUE}.}

\item{na_color}{the default color for cells where data isn't mapped.}
}
\description{
Colorizing helper function that colors by bin membership
}
