% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{color_factor}
\alias{color_factor}
\title{Colorizing helper function that colors according to factor levels}
\usage{
color_factor(palette, domain = NULL, levels = NULL, ordered = FALSE,
  na_color = "#808080")
}
\arguments{
\item{palette}{the colors or color function that values will be mapped to.}

\item{domain}{the possible values that should be mapped.}

\item{levels}{levels}

\item{ordered}{ordered}

\item{na_color}{the default color for cells where data isn't mapped.}
}
\description{
Colorizing helper function that colors according to factor levels
}
