% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align}
\alias{cols_align}
\title{Set the alignment of columns}
\usage{
cols_align(data, align = "center", columns = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{align}{the alignment direction. This can either
be \code{"center"}, \code{"left"}, or \code{"right"}.}

\item{columns}{an optional vector of column names
for which the alignment should be applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set the alignment of columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and align all
# columns to the left
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  cols_align("left")
}
