% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align}
\alias{cols_align}
\title{Set the alignment of columns}
\usage{
cols_align(data, align = "center", columns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{align}{the alignment direction. This can either be \code{"center"},
\code{"left"}, or \code{"right"}.}

\item{columns}{a vector of column names for which the alignment should be
applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set the alignment of columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and align the `mpg`
# column to the left
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    cols_align(
      align = "left",
      columns = vars(mpg))
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_hide}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
