% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align_center}
\alias{cols_align_center}
\title{Set columns to be aligned to the center}
\usage{
cols_align_center(data, columns)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an vector of column names for which the
center alignment should be applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set columns to be aligned to the center
}
\examples{
# Create a table object using the
# `mtcars` dataset and center align
# all columns
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  cols_align_center()
}
