% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{cols_align_left}
\alias{cols_align_left}
\title{Use left alignment}
\usage{
cols_align_left(html_tbl, columns = NULL, types = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{an optional vector of column names
for which the left alignment should be applied.}

\item{types}{an optional vector of column types
for which the left alignment should be applied.}
}
\value{
an object of class \code{html_table}.
}
\description{
Use left alignment on one or more table
columns. Options exist to selectively apply the
alignment to specified column names or to
columns of certain types.
}
\examples{
# Create a table object using the
# `mtcars` dataset and align all
# columns to the left
tab_create(tbl = mtcars) \%>\%
  theme_striped() \%>\%
  cols_align_left()
}
