% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align_left}
\alias{cols_align_left}
\title{Set columns to be aligned left}
\usage{
cols_align_left(data, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{columns}{an vector of column names for which the left alignment should
be applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set columns to be aligned left
}
\examples{
# Create a table object using the
# `mtcars` dataset and align all
# columns to the left
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  cols_align_left(mpg)
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_remove}},
  \code{\link{cols_split_delim}}
}
