% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align_right}
\alias{cols_align_right}
\title{Set columns to be aligned right}
\usage{
cols_align_right(data, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{a vector of column names for which the alignment should be
applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set columns to be aligned right
}
\examples{
# Create a table object using the
# `mtcars` dataset and align the
# `qsec` column to the right
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    cols_align_right(columns = vars(qsec))
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}}, \code{\link{cols_align}},
  \code{\link{cols_hide}}, \code{\link{cols_label}},
  \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
