% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_align_right}
\alias{cols_align_right}
\title{Set columns to be aligned right}
\usage{
cols_align_right(data, columns = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an optional vector of column names
for which the right alignment should be applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set columns to be aligned right
}
\examples{
# Create a table object using the
# `mtcars` dataset and align all
# columns to the right
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  cols_align_right()
}
