% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cells.R
\name{cols_color_scale}
\alias{cols_color_scale}
\title{Set data cell colors using a helper function}
\usage{
cols_color_scale(data, columns, colors)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns wherein changes to cell data colors
should occur.}

\item{colors}{a vector of colors to use for each of the provided
\code{values}. Each color value provided must either be a color name (in
the set of colors provided by \code{grDevices::colors()}) or hexadecimal
strings in the form of "#RRGGBB" or "#RRGGBBAA".}
}
\description{
Set data cell colors using a helper function
}
