% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_hide}
\alias{cols_hide}
\title{Hide one or more columns}
\usage{
cols_hide(data, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the column names to hide from the output display table. The
order of the remaining columns will be preserved. Values provided that do
not correspond to column names will be disregarded.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Hide one or more columns
}
\examples{
# Use `countrypops` to create a gt table;
# Hide the columns `country_code_2` and
# `country_code_3`
tab_1 <-
  countrypops \%>\%
  dplyr::filter(country_name == "Mongolia") \%>\%
  tail(5) \%>\%
  gt() \%>\%
  cols_hide(
    columns = vars(
      country_code_2, country_code_3)
  )

# Use `countrypops` to create a gt table;
# Use the `population` column to provide
# the conditional placement of footnotes,
# then hide that column and one other
tab_2 <-
  countrypops \%>\%
  dplyr::filter(country_name == "Mongolia") \%>\%
  tail(5) \%>\%
  gt() \%>\%
  tab_footnote(
    footnote = "Population above 3,000,000.",
    locations = cells_data(
      columns = vars(year),
      rows = population > 3000000)
  ) \%>\%
  cols_hide(
    columns = vars(country_code_3, population)
  )
}
\seealso{
Other column modification functions: \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
