% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_label}
\alias{cols_label}
\title{Re-label one or more columns}
\usage{
cols_label(data, labels)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{labels}{a named vector of column names and their
labels for display of the column headers.}
}
\description{
Re-label one or more columns
}
