% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_label}
\alias{cols_label}
\title{Relabel one or more columns}
\usage{
cols_label(data, labels)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{labels}{a named vector of column names and their labels for display of
the column headers. We can use the \code{col_labels()} function to more
easily specify column names and column labels, since we can also wrap the
column labels with \code{md()} (to interpret text as Markdown) and
\code{HTML()} (to interpret text as HTML).}
}
\description{
Relabel one or more columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and apply different
# column labels to the `mpg` and `qsec`
# columns (the column labels will be
# inherited from the column names)
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    cols_label(
      labels = col_labels(
        mpg = md("*MPG*"),
        qsec = "QMT, seconds"))
}
\seealso{
\code{\link{col_labels}()} as a useful helper function for
  processing column labels.

Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_remove}},
  \code{\link{cols_split_delim}}
}
\concept{column modification functions}
