% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_label}
\alias{cols_label}
\title{Relabel one or more columns}
\usage{
cols_label(data, ..., .list = list2(...))
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{...}{one or more named arguments of column names from the input `data`
table along with their labels for display as the column labels. We can
optionally wrap the column labels with \code{\link{md}()} (to interpret
text as Markdown) or \code{\link{html}()} (to interpret text as HTML).}
}
\description{
Relabel one or more columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and apply different
# column labels to the `mpg` and `qsec`
# columns (the column labels will be
# inherited from the column names)
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    cols_label(
      mpg = md("*MPG*"),
      qsec = "QMT, seconds")
}
\seealso{
Other column modification functions: \code{\link{cols_align}},
  \code{\link{cols_hide}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
