% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_merge}
\alias{cols_merge}
\title{Merge two columns to a single column with a formatter}
\usage{
cols_merge(data, col_1, col_2, pattern = "{1} {2}")
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{col_1}{a column that contains values for the start of the range.}

\item{col_2}{a column that contains values for the end of the range.}

\item{pattern}{a formatting pattern that specifies the arrangement of the
\code{col_1} and \code{col_1} values and any string literals. The
\code{col_1} column is represented as \code{{1}} whereas \code{col_2} is
\code{{2}}. All other characters are taken to be string literals.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Merge two columns to a single column with a formatter
}
\examples{
# Create a table object using the
# `sleep` dataset and merge the `group`
# and `ID` columns together (where the
# `ID` number is in parentheses after
# the `group` number)
gt_tbl <-
  gt(sleep) \%>\%
    cols_merge(
      col_1 = vars(group),
      col_2 = vars(ID),
      pattern = "{1} ({2})")
}
\seealso{
Other column modification functions: \code{\link{cols_align}},
  \code{\link{cols_hide}}, \code{\link{cols_label}},
  \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}},
  \code{\link{data_color}}
}
\concept{column modification functions}
