% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_merge}
\alias{cols_merge}
\title{Merge two columns to a single column with a formatter}
\usage{
cols_merge(data, col_1, col_2, format = "{1} {2}")
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{col_1}{a column that contains values for the start of the range.}

\item{col_2}{a column that contains values for the end of the range.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Merge two columns to a single column with a formatter
}
\examples{
# Create a table object using the
# `sleep` dataset and merge the `group`
# and `ID` columns together (where the
# `ID` number is in parentheses after
# the `group` number
gt(sleep) \%>\%
  cols_merge(
    col_1 = vars(group),
    col_2 = vars(ID),
    format = "{1} ({2})")
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_remove}},
  \code{\link{cols_split_delim}}
}
