% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_merge_n_pct}
\alias{cols_merge_n_pct}
\title{Merge two columns to combine counts and percentages}
\usage{
cols_merge_n_pct(data, col_n, col_pct, autohide = TRUE)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{col_n}{A column that contains values for the count component.}

\item{col_pct}{A column that contains fractional values for the percentage
component.}

\item{autohide}{An option to automatically hide the column specified as
\code{col_pct}. Any columns with their state changed to hidden will behave
the same as before, they just won't be displayed in the finalized table.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{cols_merge_n_pct()} function is a specialized variant of the
\code{\link[=cols_merge]{cols_merge()}} function. It operates by taking two columns that constitute
both a count (\code{col_n}) and a fraction of the total population (\code{col_pct}) and
merges them into a single column. What results is a column containing both
counts and their associated percentages (e.g., \verb{12 (23.2\%)}). The column
specified in \code{col_pct} is dropped from the output table.
}
\section{Function ID}{

4-10
}

\seealso{
Other Modify Columns: 
\code{\link{cols_align}()},
\code{\link{cols_hide}()},
\code{\link{cols_label}()},
\code{\link{cols_merge_range}()},
\code{\link{cols_merge_uncert}()},
\code{\link{cols_merge}()},
\code{\link{cols_move_to_end}()},
\code{\link{cols_move_to_start}()},
\code{\link{cols_move}()},
\code{\link{cols_width}()}
}
\concept{Modify Columns}
