% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{cols_merge_range}
\alias{cols_merge_range}
\title{Merge two columns to a value range column}
\usage{
cols_merge_range(data, col_begin, col_end)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{col_begin}{a column that contains values for the start of the range.}

\item{col_end}{a column that contains values for the end of the range.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Merge two columns to a value range column
}
