% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_merge_range}
\alias{cols_merge_range}
\title{Merge two columns to a value range column}
\usage{
cols_merge_range(data, col_begin, col_end)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{col_begin}{a column that contains values for the start of the range.}

\item{col_end}{a column that contains values for the end of the range.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Merge two columns to a value range column
}
\examples{
library(tidyverse)

# Create a tibble with two columns
# that are both numeric
data_tbl <-
  dplyr::tribble(
    ~val_1, ~val_2,
    1.6,    2.6,
    1.9,    3.3,
    2.6,    3.8,
    2.1,    4.2)

# Create a table object using this
# dataset where the `val_1` and
# `val_2` columns are merged together
# with an em-dash between them (the
# `val_1` column is retained)
gt(data_tbl) \%>\%
  cols_merge_range(
    col_begin = vars(val_1),
    col_end = vars(val_2))
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_label}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_remove}},
  \code{\link{cols_split_delim}}
}
