% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_move}
\alias{cols_move}
\title{Move one or more columns}
\usage{
cols_move(data, columns, after)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{columns}{the column names to move to as a group to a different
position. The order of the remaining columns will be preserved. Values
provided that do not correspond to column names will be disregarded.}

\item{after}{a column name used to anchor the insertion of the moved columns.
All of the moved columns will be placed to the right of this column.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Move one or more columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and move the `mpg`,
# `cyl`, and `disp` columns to the right
# of `qsec`
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  cols_move(
    columns = vars(mpg, cyl, disp),
    after = vars(qsec))
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_remove}}, \code{\link{cols_split_delim}}
}
