% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_remove}
\alias{cols_remove}
\title{Remove one or more columns}
\usage{
cols_remove(data, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{columns}{the column names to remove from the table. The order of the
remaining columns will be preserved. Values provided that do not correspond
to column names will be disregarded.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Remove one or more columns
}
\examples{
# Create a table object using the
# `mtcars` dataset and remove the `mpg`,
# `cyl`, and `disp` columns
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  cols_remove(
    columns = vars(mpg, cyl, disp))
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
