% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_split_delim}
\alias{cols_split_delim}
\title{Create group names and column labels via delimited column names}
\usage{
cols_split_delim(data, delim, columns = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{delim}{the delimiter to use to split an input column name. The
delimiter supplied will be autoescaped for the internal splitting
procedure. The first component of the split will become the group name and
the second component will be the column label.}

\item{columns}{an optional vector of column names that this operation should
be limited to. The default is to consider all columns in the table.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
This option will split delimited column names such that the first component
  is promoted to the group name (and hence will span over column labels) and
  subsequent components will represent the column label. Please note that
  reference to individual columns must continue to be through the original
  column names.
}
\examples{
# Create a table object using the
# `iris` dataset and split any columns
# that are dot separated between column
# groups and column labels.
gt_tbl <-
 gt(iris) \%>\%
    cols_split_delim(delim = ".")
}
\seealso{
Other column modification functions: \code{\link{cols_align_center}},
  \code{\link{cols_align_left}},
  \code{\link{cols_align_right}}, \code{\link{cols_align}},
  \code{\link{cols_label}}, \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_remove}}
}
\concept{column modification functions}
