% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns.R
\name{cols_widths}
\alias{cols_widths}
\title{Set the widths of columns}
\usage{
cols_widths(data, ..., .list = list2(...), .others = px(100))
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{...}{One or more named arguments of column names from the input \code{data}
table along with their width values.}

\item{.list}{Allows for the use of a list as an input alternative to \code{...}.}

\item{.others}{The width to set for all other columns not specified in \code{...}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Manual adjustments to column widths can be performed using the
\code{cols_widths()} function. Normally, column widths are automatically set to
span across the width of the container (both table and container widths can
be individually modified with the \code{table.width} and \code{container.width} options
within \code{\link[=tab_options]{tab_options()}}). When using \code{cols_widths()} though, the \code{table.width}
is disregarded in favor of the pixel values set for each column. We choose
which columns get specific widths (in pixels, usually by use of the \code{\link[=px]{px()}}
helper function) and all other columns are assigned a default width value
though the \code{.others} argument.
}
\seealso{
Other column modification functions: \code{\link{cols_align}},
  \code{\link{cols_hide}}, \code{\link{cols_label}},
  \code{\link{cols_merge_range}},
  \code{\link{cols_merge_uncert}},
  \code{\link{cols_merge}}, \code{\link{cols_move_to_end}},
  \code{\link{cols_move_to_start}},
  \code{\link{cols_move}}, \code{\link{cols_split_delim}}
}
\concept{column modification functions}
