% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{columns_with}
\alias{columns_with}
\title{Helper for targeting multiple columns with a pattern}
\usage{
columns_with(pattern)
}
\arguments{
\item{pattern}{a pattern that can help capture one or more column names.}
}
\value{
a list object of class \code{columns_with}.
}
\description{
Helper for targeting multiple columns with a pattern
}
\examples{
# With the `mtcars` dataset we can target
# number formatting to those column labels
# that fit a specified pattern
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  fmt_number(
    columns = columns_with("^c"),
    decimals = 3)
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{not_in_group}},
  \code{\link{rownames_with}}, \code{\link{rows_where}},
  \code{\link{target_cells}}, \code{\link{target_cell}},
  \code{\link{vars}}
}
