% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{columns_with}
\alias{columns_with}
\title{Helper for targeting multiple columns with a pattern}
\usage{
columns_with(pattern)
}
\arguments{
\item{pattern}{a pattern that can help capture one or more column names.}
}
\value{
a list object of class \code{columns_with}.
}
\description{
Helper for targeting multiple columns with a pattern
}
