% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{constants}
\alias{constants}
\title{The fundamental physical constants}
\format{
A tibble with 354 rows and 4 variables:
\describe{
\item{name}{The name of the constant.}
\item{value}{The value of the constant.}
\item{uncert}{The uncertainty associated with the value. If \code{NA} then the
value is seen as an 'exact' value (e.g., an electron volt has the exact value
of 1.602 176 634 e-19 J)}
\item{units}{The units associated with the constant.}
}
}
\usage{
constants
}
\description{
This dataset contains values for over 300 basic fundamental constants in
nature. The values originate from the 2018 adjustment which is based on the
latest relevant precision measurements and improvements of theoretical
calculations. Such work has been carried out under the authority of the
\emph{Task Group on Fundamental Constants} (TGFC) of the
\emph{Committee on Data of the International Science Council} (CODATA). These
updated values became available on May 20, 2019. They are published at
\url{http://physics.nist.gov/constants}, a website of the
\emph{Fundamental Constants Data Center} of the
\emph{National Institute of Standards and Technology} (NIST), Gaithersburg,
Maryland, USA.
}
\section{Examples}{


Here is a glimpse at the data available in \code{constants}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dplyr::glimpse(constants)
#> Rows: 354
#> Columns: 4
#> $ name   <chr> "alpha particle-electron mass ratio", "alpha particle mass", "a~
#> $ value  <dbl> 7.294300e+03, 6.644657e-27, 5.971920e-10, 3.727379e+03, 4.00150~
#> $ uncert <dbl> 2.4e-07, 2.0e-36, 1.8e-19, 1.1e-06, 6.3e-11, 1.2e-12, 2.2e-10, ~
#> $ units  <chr> NA, "kg", "J", "MeV", "u", "kg mol^-1", NA, NA, "m", "kg", "J",~
}\if{html}{\out{</div>}}
}

\section{Dataset ID and Badge}{

DATA-9
}

\section{Dataset Introduced}{

\emph{In Development}
}

\seealso{
Other datasets: 
\code{\link{countrypops}},
\code{\link{exibble}},
\code{\link{gtcars}},
\code{\link{illness}},
\code{\link{metro}},
\code{\link{pizzaplace}},
\code{\link{rx_addv}},
\code{\link{rx_adsl}},
\code{\link{sp500}},
\code{\link{sza}},
\code{\link{towny}}
}
\concept{datasets}
\keyword{datasets}
