% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_lines.R
\name{create_summary_lines}
\alias{create_summary_lines}
\title{Create lines with summarized data}
\usage{
create_summary_lines(html_tbl, groups = NULL, columns, fcn, summary_labels)
}
\arguments{
\item{html_tbl}{an HTML table object that
is created using the \code{gt()} function.}

\item{groups}{the groups to use when summarizing.}

\item{columns}{the columns to summarize.}

\item{fcn}{the name of a function to apply per
group per column.}

\item{summary_labels}{the labels to use in the stub
for the created summary rows.}
}
\description{
Create lines with summarized data
}
