% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_cells}
\alias{data_cells}
\title{Helper for targeting multiple table cells}
\usage{
data_cells(rows = NULL, columns = NULL)
}
\arguments{
\item{row}{a single row or a range of rows to target.}

\item{column}{a single column or a range of columns to target.}
}
\value{
a list object of class \code{multi_cell_target}.
}
\description{
Helper for targeting multiple table cells
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{columns_with}},
  \code{\link{html}}, \code{\link{md}},
  \code{\link{not_in_group}}, \code{\link{rownames_with}},
  \code{\link{rows_where}}, \code{\link{vars}}
}
\concept{helper functions}
