% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_color.R
\name{data_color}
\alias{data_color}
\title{Set data cell colors using a palette or a colorizing helper function}
\usage{
data_color(data, columns, colors, alpha = NULL, apply_to = "bkgd",
  autocolor_text = TRUE)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns wherein changes to cell data colors should occur.}

\item{colors}{either a color mapping function from the \code{scales} package
or a vector of colors to use for each distinct value or level in each of
the provided \code{columns}. The color mapping functions are:
\code{scales::col_quantile()}, \code{scales::col_bin()},
\code{scales::col_numeric()}, and \code{scales::col_factor()}. If providing
a vector of colors as a palette, each color value provided must either be a
color name (in the set of colors provided by \code{grDevices::colors()}) or
hexadecimal strings in the form of "#RRGGBB" or "#RRGGBBAA".}

\item{alpha}{an optional, fixed alpha transparency value that will be applied
to all of the \code{colors} provided if they are provided as a vector of
colors. If using a colorizing helper function for \code{colors} then this
option is ignored (each of the colorizing helper functions has its own
\code{alpha} argument).}

\item{apply_to}{which style element should the colors be applied to? Options
include the cell background (the default, given as \code{bkgd}) or the cell
text (\code{text}).}

\item{autocolor_text}{an option to let \code{gt} modify the coloring of text
within cells undergoing background coloring. This will in some cases yield
more optimal text-to-background color contrast. By default, this is set to
\code{TRUE}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set data cell colors using a palette or a colorizing helper function
}
\concept{data cell modification}
