% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emit_html.R
\name{emit_html}
\alias{emit_html}
\title{Transform an HTML table object to an HTML fragment}
\usage{
emit_html(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{gt()} function.}
}
\value{
a character object with an HTML fragment.
}
\description{
Take a suitably styled HTML table object and
transform it to an HTML fragment. This is the
final step in the \code{gt()} -> ... ->
\code{emit_html()} pattern.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- gt(tbl = iris)

# The resulting object can be used
# in transformations
html_table_transformed <-
  html_table \%>\%
  add_column_style(
    columns = 1,
    property = "color",
    values = "blue")

# We can extract the HTML from the
# HTML table object using `emit_html()`
html_table_transformed \%>\%
  emit_html()
}
