% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{escape_latex}
\alias{escape_latex}
\title{Perform Latex escaping}
\usage{
escape_latex(text)
}
\arguments{
\item{text}{a character vector containing the text that is to be
Latex-escaped.}
}
\description{
Text may contain several characters with special meanings in Latex. This
function will transform a character vector so that it is safe to use within
Latex tables.
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{html}}, \code{\link{md}}, \code{\link{pct}},
  \code{\link{px}}
}
\concept{helper functions}
