% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{escape_latex}
\alias{escape_latex}
\title{Perform LaTeX escaping}
\usage{
escape_latex(text)
}
\arguments{
\item{text}{a character vector containing the text that is to be
LaTeX-escaped.}
}
\description{
Text may contain several characters with special meanings in LaTeX. This
function will transform a character vector so that it is safe to use within
LaTeX tables.
}
\seealso{
Other helper functions: \code{\link{cells_styles}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}, \code{\link{px}}
}
\concept{helper functions}
