% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{escape_latex}
\alias{escape_latex}
\title{Perform LaTeX escaping}
\usage{
escape_latex(text)
}
\arguments{
\item{text}{\emph{LaTeX text}

\verb{vector<character>} // \strong{required}

A character vector containing the text that is to be LaTeX-escaped.}
}
\value{
A character vector.
}
\description{
Text may contain several characters with special meanings in LaTeX. The
\code{escape_latex()} function will transform a character vector so that it is
safe to use within LaTeX tables.
}
\section{Function ID}{

8-27
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_footnotes}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_source_notes}()},
\code{\link{cells_stub_grand_summary}()},
\code{\link{cells_stub_summary}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{define_units}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()},
\code{\link{system_fonts}()}
}
\concept{helper functions}
