% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{escape_latex}
\alias{escape_latex}
\title{Perform LaTeX escaping}
\usage{
escape_latex(
  text,
  unicode_conversion = getOption("gt.latex.unicode_convert", default = FALSE)
)
}
\arguments{
\item{text}{\emph{LaTeX text}

\verb{vector<character>} // \strong{required}

A character vector containing the text that is to be LaTeX-escaped.}

\item{unicode_conversion}{\emph{Boolean}

\verb{scalar<boolean>} // \emph{default:} \code{TRUE}

A boolean value indicating whether unicode in text should be turned into
LaTeX.}
}
\value{
A character vector.
}
\description{
Text may contain several characters with special meanings in LaTeX.
\code{escape_latex()} will transform a character vector so that it is safe to use
within LaTeX tables.
}
\section{Function ID}{

8-29
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{latex}()},
\code{\link{md}()},
\code{\link{nanoplot_options}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{row_group}()},
\code{\link{stub}()},
\code{\link{system_fonts}()},
\code{\link{unit_conversion}()}
}
\concept{helper functions}
