% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{extract_body}
\alias{extract_body}
\title{Extract the table body from a \strong{gt} object}
\usage{
extract_body(data, stage = NULL, output = c("html", "latex", "rtf", "word"))
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[=gt]{gt()}} function.}
}
\value{
A data frame or tibble object containing the table body.
}
\description{
We can extract the body of a \strong{gt} table, even at various stages of its
rendering, from a \code{gt_tbl} object using the \code{extract_body()} function.
}
\section{Function ID}{

13-6
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other table export functions: 
\code{\link{as_latex}()},
\code{\link{as_raw_html}()},
\code{\link{as_rtf}()},
\code{\link{as_word}()},
\code{\link{extract_cells}()},
\code{\link{extract_summary}()},
\code{\link{gtsave}()}
}
\concept{table export functions}
