% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract a summary table from a gt table object}
\usage{
extract_summary(data)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}
}
\value{
a list of data frames containing summary data.
}
\description{
Get a list of summary row data frames from a \code{gt_tbl} object where
summary rows were added via the \code{\link{summary_rows}()} function. The
output data frames contain the \code{groupname} and \code{rowname} columns,
whereby \code{rowname} contains descriptive stub labels for the summary rows.
}
\seealso{
Other table export functions: \code{\link{as_latex}},
  \code{\link{as_raw_html}}, \code{\link{as_rtf}}
}
\concept{table export functions}
