% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract a summary table from a gt table object}
\usage{
extract_summary(data)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}
}
\value{
a data frame containing summary data.
}
\description{
Take a \code{gt_tbl} table object and transform it to an HTML table.
}
\seealso{
Other table export functions: \code{\link{write_rtf}}
}
\concept{table export functions}
