% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract a summary table from a gt table object}
\usage{
extract_summary(data)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}
}
\value{
a data frame containing summary data.
}
\description{
Get a summary row data frame from a \code{gt_tbl} object where summary rows
were added via the \code{summary_rows()} function. The output data frame
contains the \code{groupname} and \code{rowname} columns, where
\code{rowname} contains descriptive stub labels for the summary rows.
The remaining columns are a subset of the original dataset, where only those
columns that had aggregation functions applied are retained. The output data
frame format is suitable for generating a summary table using the
\code{\link{gt}()} function.
}
\seealso{
Other table export functions: \code{\link{write_rtf}}
}
\concept{table export functions}
