% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt}
\alias{fmt}
\title{Set a column format with a formatter function}
\usage{
fmt(data, columns = NULL, rows = NULL, fcns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns to format. Can either be a series of column names
provided in \code{vars()}, a vector of column indices, or a helper function
focused on selections. The select helper functions are:
\code{\link{starts_with}()}, \code{\link{ends_with}()},
\code{\link{contains}()}, \code{\link{matches}()}, \code{\link{one_of}()},
and \code{\link{everything}()}.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted. Can either be a vector of row
captions provided \code{c()}, a vector of row indices, or a helper function
focused on selections. The select helper functions are:
\code{\link{starts_with}()}, \code{\link{ends_with}()},
\code{\link{contains}()}, \code{\link{matches}()}, \code{\link{one_of}()},
and \code{\link{everything}()}. We can also use expressions to filter down
to the rows we need (e.g., \code{[colname_1] > 100 & [colname_2] < 50}).}

\item{fcns}{a single formatting function or a named list of functions.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a column format with a formatter function
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_missing}}, \code{\link{fmt_number}},
  \code{\link{fmt_passthrough}}, \code{\link{fmt_percent}},
  \code{\link{fmt_scientific}}, \code{\link{fmt_time}}
}
\concept{data formatting functions}
