% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_currency}
\alias{fmt_currency}
\title{Format values as currencies}
\usage{
fmt_currency(data, columns, rows = NULL, currency, use_subunits = TRUE,
  negative_val = "signed", decimals = NULL, sep_mark = ",",
  dec_mark = ".", placement = "left", locale = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{currency}{the currency to use for the numeric value.}

\item{use_subunits}{an option for whether the subunits portion of a currency
value should be displayed.}

\item{negative_val}{the formatting to use for negative numbers. With
\code{signed} (the default), negative numbers will be shown with a negative
sign. Using \code{parens} will show the negative value in parentheses. The
\code{red} option will display the number in red. Finally, \code{parens-red}
will display negative numbers as red and enclosed in parentheses.}

\item{decimals}{an option to specify the exact number of decimal places to
use.}

\item{sep_mark}{the mark to use as a separator between groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}

\item{placement}{the placement of the currency symbol. This can be either be
\code{left} (the default) or \code{right}.}

\item{locale}{an optional locale ID that can be used for formatting the value
according the locale's rules. Examples include \code{"en_US"} for English
(United States) and \code{"fr_FR"} for French (France).}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as currencies
}
\examples{
library(tidyverse)

# Create a tibble with two columns
# that are both numeric
data_tbl <-
  dplyr::tribble(
    ~val_usd, ~val_jpy,
    23.346,   2606.95,
    832.830,  93002.54,
    72.013,   8041.73,
    83.1,     9279.82)

# Create a table object using this
# dataset and where the `val_usd`
# and `val_jpy` are formatted as
# the `USD` and `JPY` currencies
gt(data_tbl) \%>\%
  fmt_currency(
    columns = vars(val_usd),
    currency = "USD") \%>\%
  fmt_currency(
    columns = vars(val_jpy),
    currency = "JPY")
}
\seealso{
Other data formatting functions: \code{\link{fmt_datetime}},
  \code{\link{fmt_date}}, \code{\link{fmt_missing}},
  \code{\link{fmt_number}}, \code{\link{fmt_percent}},
  \code{\link{fmt_scientific}}, \code{\link{fmt_time}},
  \code{\link{set_fmt}}
}
