% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_currency}
\alias{fmt_currency}
\title{Format values as currencies}
\usage{
fmt_currency(data, columns, currency, use_subunits = TRUE, decimals = NULL,
  sep_mark = "", dec_mark = ".", placement = "left",
  negative_style = "signed")
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to format.}

\item{currency}{the currency to use for the numeric value.}

\item{use_subunits}{an option for whether the subunits
portion of a currency value should be displayed.}

\item{decimals}{an option to specify the exact number
of decimal places to use.}

\item{sep_mark}{the mark to use as a separator between
groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}

\item{placement}{the placement of the currency symbol.
This can be either be \code{left} (the default) or
\code{right}.}

\item{negative_style}{the formatting to use
for negative numbers. With \code{signed} (the default),
negative numbers will be shown with a negative sign.
Using \code{parens} will show the negative value in
parentheses. The \code{red} option will display the
number in red. Finally, \code{parens-red} will display
negative numbers as red and enclosed in parentheses.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as currencies
}
