% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_date}
\alias{fmt_date}
\title{Format values as dates}
\usage{
fmt_date(data, columns, rows = NULL, date_style = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{date_style}{the date style to use. Supply a number (from \code{1} to
\code{14}) that corresponds to the preferred date style. Use
\code{date_style_info()} to see the different numbered and named date
presets.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as dates
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_missing}},
  \code{\link{fmt_number}}, \code{\link{fmt_percent}},
  \code{\link{fmt_scientific}}, \code{\link{fmt_time}},
  \code{\link{set_fmt}}
}
