% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_datetime}
\alias{fmt_datetime}
\title{Format values as date-times}
\usage{
fmt_datetime(data, columns, rows = NULL, where = NULL, date_style,
  time_style)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns to format. Can either be a series of column names
provided in \code{vars()}, a vector of column indices, or a helper function
focused on selections. The select helper functions are:
\code{\link{starts_with}()}, \code{\link{ends_with}()},
\code{\link{contains}()}, \code{\link{matches}()}, \code{\link{one_of}()},
and \code{\link{everything}()}.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted. Can either be a vector of row
captions provided \code{c()}, a vector of row indices, or a helper function
focused on selections. The select helper functions are:
\code{\link{starts_with}()}, \code{\link{ends_with}()},
\code{\link{contains}()}, \code{\link{matches}()}, \code{\link{one_of}()},
and \code{\link{everything}()}.}

\item{where}{a conditional expression that operates across all of the rows
captured by \code{rows} to further constrain the row selection.}

\item{date_style}{the date style to use. Supply a number (from \code{1} to
\code{14}) that corresponds to the preferred date style. Use
\code{\link{info_date_style}()} to see the different numbered and named date
presets.}

\item{time_style}{the time style to use. Supply a number (from \code{1} to
\code{5}) that corresponds to the preferred time style. Use
\code{\link{info_time_style}()} to see the different numbered and named time
presets.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as date-times
}
\examples{
library(tidyverse)

# Create a tibble with a column
# that contains 24-hour time strings
data_tbl <-
  dplyr::tribble(
    ~time,
    "2017-06-10 12:35:23",
    "2017-07-12 15:01:34",
    "2017-08-05 09:45:23",
    "2017-10-23 01:32:00")

# Create a table object using this
# dataset and format the `time`
# column with `time_style` '3'
gt_tbl <-
  gt(data_tbl) \%>\%
    fmt_datetime(
      columns = vars(time),
      date_style = 2,
      time_style = 3)
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_date}}, \code{\link{fmt_missing}},
  \code{\link{fmt_number}}, \code{\link{fmt_passthrough}},
  \code{\link{fmt_percent}}, \code{\link{fmt_scientific}},
  \code{\link{fmt_time}}, \code{\link{fmt}}
}
\concept{data formatting functions}
