% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_duration}
\alias{fmt_duration}
\title{Format values as time durations}
\usage{
fmt_duration(
  data,
  columns,
  rows = everything(),
  input_units = NULL,
  output_units = NULL,
  duration_style = c("narrow", "wide", "colon-sep", "iso"),
  trim_zero_units = TRUE,
  max_output_units = NULL,
  pattern = "{x}"
)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{columns}{The columns to format. Can either be a series of column names
provided in \code{\link[=c]{c()}}, a vector of column indices, or a helper function
focused on selections. The select helper functions are: \code{\link[=starts_with]{starts_with()}},
\code{\link[=ends_with]{ends_with()}}, \code{\link[=contains]{contains()}}, \code{\link[=matches]{matches()}}, \code{\link[=one_of]{one_of()}}, \code{\link[=num_range]{num_range()}}, and
\code{\link[=everything]{everything()}}.}

\item{rows}{Optional rows to format. Providing \code{\link[=everything]{everything()}} (the
default) results in all rows in \code{columns} being formatted. Alternatively,
we can supply a vector of row captions within \code{\link[=c]{c()}}, a vector of row
indices, or a helper function focused on selections. The select helper
functions are: \code{\link[=starts_with]{starts_with()}}, \code{\link[=ends_with]{ends_with()}}, \code{\link[=contains]{contains()}}, \code{\link[=matches]{matches()}},
\code{\link[=one_of]{one_of()}}, \code{\link[=num_range]{num_range()}}, and \code{\link[=everything]{everything()}}. We can also use expressions
to filter down to the rows we need (e.g.,
\verb{[colname_1] > 100 & [colname_2] < 50}).}

\item{input_units}{If one or more selected columns contains numeric values,
the keyword provided for \code{input_units} will determine how those values are
interpreted in terms of a single duration type. If numeric columns are
included in \code{columns} then one of the following keywords, representative of
time units, must be provided: seconds (\code{"secs"}), \code{"mins"} (minutes),
\code{"hours"}, \code{"days"}, or \code{"weeks"}.}

\item{output_units}{Controls the output time units. The default, \code{NULL},
means that \strong{gt} will automatically choose time units based on the input
duration value. To control which time units are to be considered for output
(before trimming with \code{trim_zero_units}) we can specify a range of time
units with a two-element vector consisting of these keywords: \code{"weeks"},
\code{"days"}, \code{"hours"}, \code{"mins"}, or \code{"secs"}. If only a single time unit is
needed for output, a one-element vector can be used with any of the
aforementioned keywords.}

\item{duration_style}{A choice of three formatting styles for the output
duration values. With \code{"narrow"} (the default style), duration values will
be formatted with single letter time-part units (e.g., 1.35 days will be
styled as \verb{"1d 8h 24m}). With \code{"wide"}, this example value will be expanded
to \code{"1 day 8 hours 24 minutes"} after formatting. The \code{"colon-sep"} style
will put days, hours, minutes, and seconds in the \code{"[D]/[HH]:[MM]:[SS]"}
format. The \code{"iso"} style will produce a value that conforms to the ISO
8601 rules for duration values (e.g., 1.35 days will become \code{"P1DT8H24M"}).}

\item{trim_zero_units}{Provides methods to remove output time units that have
zero values. By default this is \code{TRUE} and duration values that might
otherwise be formatted as \code{"0w 1d 0h 4m 19s"} with
\code{trim_zero_units = FALSE} are instead displayed as \code{"1d 4m 19s"}. Aside
from using \code{TRUE}/\code{FALSE} we could provide a vector of keywords for more
precise control. These keywords are: (1) \code{"leading"}, to omit all leading
zero-value time units (e.g., \code{"0w 1d"} -> \code{"1d"}), (2) \code{"trailing"}, to
omit all trailing zero-value time units (e.g., \code{"3d 5h 0s"} -> \code{"3d 5h"}),
and \code{"internal"}, which removes all internal zero-value time units (e.g.,
\code{"5d 0h 33m"} -> \code{"5d 33m"}).}

\item{max_output_units}{If \code{output_units} is \code{NULL}, where the output time
units are unspecified and left to \strong{gt} to handle, a numeric value
provided for \code{max_output_units} will be taken as the maximum number of time
units to display in all output time duration values. By default, this is
\code{NULL} and all possible time units will be displayed.}

\item{pattern}{A formatting pattern that allows for decoration of the
formatted value. The value itself is represented by \code{{x}} and all other
characters are taken to be string literals.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Format input values to time durations whether those input values are numbers
or of the \code{difftime} class. We can specify which time units any numeric input
values have (as days, hours, minutes, or seconds) and the output can be
customized with a duration style (corresponding to short, long, and
colon-separated forms) and a choice of output units from weeks to seconds.
}
\details{
Targeting of values is done through \code{columns} and additionally by \code{rows} (if
nothing is provided for \code{rows} then entire columns are selected). Conditional
formatting is possible by providing a conditional expression to the \code{rows}
argument. See the \emph{Arguments} section for more information on this.
}
\section{Function ID}{

3-11
}

\seealso{
Other Format Data: 
\code{\link{data_color}()},
\code{\link{fmt_bytes}()},
\code{\link{fmt_currency}()},
\code{\link{fmt_datetime}()},
\code{\link{fmt_date}()},
\code{\link{fmt_engineering}()},
\code{\link{fmt_fraction}()},
\code{\link{fmt_integer}()},
\code{\link{fmt_markdown}()},
\code{\link{fmt_missing}()},
\code{\link{fmt_number}()},
\code{\link{fmt_passthrough}()},
\code{\link{fmt_percent}()},
\code{\link{fmt_scientific}()},
\code{\link{fmt_time}()},
\code{\link{fmt}()},
\code{\link{text_transform}()}
}
\concept{Format Data}
