% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formats.R
\name{fmt_fraction}
\alias{fmt_fraction}
\title{Format values to fractions}
\usage{
fmt_fraction(html_tbl, columns = NULL, accuracy = 3)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{an option specify which columns
are to be formatted.}

\item{accuracy}{an option to specify the
accuracy of the transformation from decimal
values to fractional values. Options are
\code{1}, \code{2}, or \code{3} for
accuracies with up to 1, 2, or 3 digits in
the denominator. Other options are \code{halves},
\code{quarters}, \code{eigths}, \code{sixteenths},
\code{tenths}, and \code{hundreths}, where all
of these round fractions to the nearest namesake
value.}
}
\value{
an object of class \code{html_table}.
}
\description{
Format values to fractions
}
