% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_missing}
\alias{fmt_missing}
\title{Format missing values}
\usage{
fmt_missing(data, columns, rows = NULL, missing_text = "---")
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{missing_text}{the text to be used in place of \code{NA} values in the
rendered table.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Wherever there is missing data (i.e., \code{NA} values) a customizable mark
may present better than the standard `NA` text that would otherwise appear.
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_number}}, \code{\link{fmt_passthrough}},
  \code{\link{fmt_percent}}, \code{\link{fmt_scientific}},
  \code{\link{fmt_time}}, \code{\link{fmt}}
}
\concept{data formatting functions}
