% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formats.R
\name{fmt_number}
\alias{fmt_number}
\title{Format values as a number}
\usage{
fmt_number(html_tbl, columns = NULL, decimals = NULL,
  use_big_mark = FALSE, drop_trailing_zeros = FALSE,
  negative_style = "signed")
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{columns}{an option specify which columns
are to be formatted.}

\item{decimals}{an option to specify exactly
the number of decimal places to use.}

\item{use_big_mark}{a logical value that
allows for inclusion of a number grouping
separator for larger numbers.}

\item{drop_trailing_zeros}{a logical value that
allows for removal of trailing zeros (those
redundant zeros after the decimal mark).}

\item{negative_style}{the formatting to use
for negative numbers. With \code{signed},
negative numbers will be shown with a negative
sign. Using \code{parens} will show the
negative value in parentheses. The \code{red}
option will display the number in red.
Finally, \code{parens-red} will display
negative numbers as red and enclosed in
parentheses.}
}
\value{
an object of class \code{html_table}.
}
\description{
Format values as a number
}
