% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_percent}
\alias{fmt_percent}
\title{Format values as a percentage}
\usage{
fmt_percent(data, columns, rows = NULL, decimals = 2,
  drop0trailing = FALSE, negative_val = "signed", sep_mark = ",",
  dec_mark = ".", incl_space = FALSE, placement = "right",
  locale = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{decimals}{an option to specify the exact number of decimal places to
use. The default number of decimal places is \code{2}.}

\item{drop0trailing}{a logical value that allows for removal of trailing
zeros (those redundant zeros after the decimal mark).}

\item{negative_val}{the formatting to use for negative numbers. With
\code{signed} (the default), negative numbers will be shown with a negative
sign. Using \code{parens} will show the negative value in parentheses. The
\code{red} option will display the number in red. Finally, \code{parens-red}
will display negative numbers as red and enclosed in parentheses.}

\item{sep_mark}{the mark to use as a separator between groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}

\item{incl_space}{an option on whether to include a space between the value
and the percent sign. The default is to not introduce a space character.}

\item{placement}{the placement of the percent sign. This can be either be
\code{right} (the default) or \code{left}.}

\item{locale}{an optional locale ID that can be used for formatting the value
according the locale's rules. Examples include \code{"en_US"} for English
(United States) and \code{"fr_FR"} for French (France).}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as a percentage
}
\examples{
library(tidyverse)

# Create a tibble with two columns
# that are both numeric
data_tbl <-
  dplyr::tribble(
    ~val_1, ~val_2,
    0.345,  2.6,
    0.234,  3.3,
    0.53,   3.8,
    0.391,  4.2)

# Create a table object using this
# dataset and where the `val_1` columns
# are formatted as percentage values
gt(data_tbl) \%>\%
  fmt_percent(columns = vars(val_1))
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_missing}}, \code{\link{fmt_number}},
  \code{\link{fmt_scientific}}, \code{\link{fmt_time}},
  \code{\link{set_fmt}}
}
