% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_percent}
\alias{fmt_percent}
\title{Format values as a percentage}
\usage{
fmt_percent(data, columns, rows = NULL, decimals = 2,
  drop_trailing_zeros = FALSE, negative_val = "signed",
  use_seps = TRUE, pattern = "{x}", sep_mark = ",", dec_mark = ".",
  incl_space = FALSE, placement = "right", locale = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()} function.}

\item{columns}{the column names to format.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted.}

\item{decimals}{an option to specify the exact number of decimal places to
use. The default number of decimal places is \code{2}.}

\item{drop_trailing_zeros}{a logical value that allows for removal of
trailing zeros (those redundant zeros after the decimal mark).}

\item{negative_val}{the formatting to use for negative values. With
\code{signed} (the default), negative values will be shown with a negative
sign. Using \code{parens} will show the negative value in parentheses.}

\item{use_seps}{an option to use digit group separators. The type of digit
group separator is set by \code{sep_mark} and overridden if a locale ID is
provided to \code{locale}. This setting is \code{TRUE} by default.}

\item{pattern}{a formatting pattern that allows for decoration of the
formatted value. The value itself is represented by \code{{x}} and all
other characters are taken to be string literals.}

\item{sep_mark}{the mark to use as a separator between groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}

\item{incl_space}{an option on whether to include a space between the value
and the percent sign. The default is to not introduce a space character.}

\item{placement}{the placement of the percent sign. This can be either be
\code{right} (the default) or \code{left}.}

\item{locale}{an optional locale ID that can be used for formatting the value
according the locale's rules. Examples include \code{"en_US"} for English
(United States) and \code{"fr_FR"} for French (France). The use of a valid
locale ID will override any values provided in \code{sep_mark} and
\code{dec_mark}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values as a percentage
}
\examples{
library(tidyverse)

# Create a tibble with two columns
# that are both numeric
data_tbl <-
  dplyr::tribble(
    ~val_1, ~val_2,
    0.345,  2.6,
    0.234,  3.3,
    0.53,   3.8,
    0.391,  4.2)

# Create a table object using this
# dataset and foramt the `val_1` column
# as percentage values and `val_2`
# as numeric values
gt_tbl <-
  gt(data_tbl) \%>\%
    fmt_percent(
      columns = vars(val_1),
      decimals = 1) \%>\%
    fmt_number(
      columns = vars(val_2),
      decimals = 2)
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_missing}}, \code{\link{fmt_number}},
  \code{\link{fmt_passthrough}},
  \code{\link{fmt_scientific}}, \code{\link{fmt_time}},
  \code{\link{fmt}}
}
\concept{data formatting functions}
