% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{fmt_scientific}
\alias{fmt_scientific}
\title{Format values to scientific notation}
\usage{
fmt_scientific(data, columns, rows = NULL, where = NULL,
  decimals = 2, drop_trailing_zeros = FALSE, scale_by = 1,
  pattern = "{x}", sep_mark = ",", dec_mark = ".", locale = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{columns}{the columns to format. Can either be a series of column names
provided in \code{vars()}, a vector of column indices, or a helper function
focused on selections. The select helper functions are:
\code{\link{starts_with}()}, \code{\link{ends_with}()},
\code{\link{contains}()}, \code{\link{matches}()}, \code{\link{one_of}()},
and \code{\link{everything}()}.}

\item{rows}{optional rows to format. Not providing any value results in all
rows in \code{columns} being formatted. Can either be a vector of row
captions provided \code{c()}, a vector of row indices, or a helper function
focused on selections. The select helper functions are:
\code{\link{starts_with}()}, \code{\link{ends_with}()},
\code{\link{contains}()}, \code{\link{matches}()}, \code{\link{one_of}()},
and \code{\link{everything}()}.}

\item{where}{a conditional expression that operates across all of the rows
captured by \code{rows} to further constrain the row selection.}

\item{decimals}{an option to specify the exact number of decimal places to
use. The default number of decimal places is \code{2}.}

\item{drop_trailing_zeros}{a logical value that allows for removal of
trailing zeros (those redundant zeros after the decimal mark).}

\item{scale_by}{a value to scale the input. The default is \code{1.0}.}

\item{pattern}{a formatting pattern that allows for decoration of the
formatted value. The value itself is represented by \code{{x}} and all
other characters are taken to be string literals.}

\item{sep_mark}{the mark to use as a separator between groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}

\item{locale}{an optional locale ID that can be used for formatting the value
according the locale's rules. Examples include \code{"en_US"} for English
(United States) and \code{"fr_FR"} for French (France). The use of a valid
locale ID will override any values provided in \code{sep_mark} and
\code{dec_mark}.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Format values to scientific notation
}
\examples{
# Create a table object using the
# `mtcars` dataset and format specified
# numeric columns to display values in
# scientific notation
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    fmt_scientific(
      columns = vars(disp),
      decimals = 2)
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_missing}}, \code{\link{fmt_number}},
  \code{\link{fmt_passthrough}}, \code{\link{fmt_percent}},
  \code{\link{fmt_time}}, \code{\link{fmt}}
}
\concept{data formatting functions}
