% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{fmt_uncertainty}
\alias{fmt_uncertainty}
\title{Apply uncertainty values to a column of data}
\usage{
fmt_uncertainty(html_tbl, value_column, uncert_column)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{value_column}{the column containing the
values where uncertainties should be applied.}

\item{uncert_column}{the column containing
uncertainties related to quantities in
\code{value column}.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply uncertainty values to a column of data
}
