% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{ggplot_image}
\alias{ggplot_image}
\title{Helper function for adding a ggplot}
\usage{
ggplot_image(plot_object, height = 100, aspect_ratio = 1)
}
\arguments{
\item{plot_object}{the ggplot plot object.}

\item{height}{the absolute height (px) of the image in the table cell.}

\item{aspect_ratio}{the plot's final aspect ratio.}
}
\value{
a character object with an HTML fragment that can be placed inside of
  a cell.
}
\description{
Add a ggplot plot inside a table with this helper function.
}
\examples{
\dontrun{
library(ggplot2)

# Create a ggplot plot
plot_object <-
  ggplot(
    data = mtcars,
    aes(x = disp, y = hp,
        color = wt, size = mpg)) +
  geom_point()

# Create an HTML fragment that
# contains an the ggplot as an
# embedded plot that's 100px high
plot_html <-
  plot_object \%>\%
  ggplot_image(
    height = 200)

# Create a table that contains the
# the html fragment with the plot
# then pass it to `gt()`
tribble(
  ~row, ~plot,
  1,  plot_html) \%>\%
  gt()
}
}
