% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gt-options}
\alias{gt-options}
\title{\strong{gt} package options}
\description{
\strong{gt} package options
}
\section{Package options}{


\strong{gt} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{gt.html_tag_check}: A logical scalar indicating whether or not to print a
warning when HTML tags are found in a table that is being rendered to LaTeX.
\item \code{gt.row_group.sep}: A separator between groups for the row group
label.
}
}

\keyword{internal}
