% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gt-options}
\alias{gt-options}
\title{\strong{gt} package options}
\description{
\strong{gt} package options
}
\section{Package options}{


\strong{gt} uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{gt.stub_group.sep}: a separator between groups for the
stub group label.

}
}

