% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gt-options}
\alias{gt-options}
\title{gt package options}
\description{
gt package options
}
\section{Package options}{


gt uses the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{gt.stub_group.sep}: a separator

}
}

