% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gt-options}
\alias{gt-options}
\title{\pkg{gt} package options}
\description{
\pkg{gt} package options
}
\section{Package options}{


\pkg{gt} uses the following \code{\link{options}} to configure behavior:

\itemize{
  \item \code{gt.stub_group.sep}: a separator between groups for the
  stub group label.

}
}

