% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-package.R
\docType{package}
\name{gt-package}
\alias{gt-package}
\alias{_PACKAGE}
\title{gt: Easily Create Presentation-Ready Display Tables}
\description{
Build display tables from tabular data using with an easy-to-use
    API. With its progressive approach, we can construct display tables with a
    clear separation of concerns: you don't have to decide how the tabular data
    gets transformed and structured whilst also worrying about aesthetics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/gt}
  \item Report bugs at \url{https://github.com/rstudio/gt/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{rich@rstudio.com} (0000-0003-3925-190X)

Authors:
\itemize{
  \item Joe Cheng \email{joe@rstudio.com}
  \item Barret Schloerke \email{barret@rstudio.com} (0000-0001-9986-114X)
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
