% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-package.R
\docType{package}
\name{gt-package}
\alias{gt-package}
\alias{_PACKAGE}
\title{gt: Easily Create Presentation-Ready Display Tables}
\description{
Build display tables from tabular data with an easy-to-use set of
    functions. With its progressive approach, we can construct display tables
    with a cohesive set of table parts. Table values can be formatted using any
    of the included formatting functions. Footnotes and cell styles can be 
    precisely added through a location targeting system. The way in which 'gt'
    handles things for you means that you don't often have to worry about the
    fine details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/gt}
  \item Report bugs at \url{https://github.com/rstudio/gt/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{rich@rstudio.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})

Authors:
\itemize{
  \item Joe Cheng \email{joe@rstudio.com}
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
