% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-package.R
\docType{package}
\name{gt-package}
\alias{gt-package}
\title{gt: Easily Create Presentation-Ready Display Tables}
\description{
Build display tables from tabular data with an easy-to-use set of functions. With its progressive approach, we can construct display tables with a cohesive set of table parts. Table values can be formatted using any of the included formatting functions. Footnotes and cell styles can be precisely added through a location targeting system. The way in which 'gt' handles things for you means that you don't often have to worry about the fine details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gt.rstudio.com}
  \item \url{https://github.com/rstudio/gt}
  \item Report bugs at \url{https://github.com/rstudio/gt/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{rich@posit.co} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})

Authors:
\itemize{
  \item Joe Cheng \email{joe@posit.co}
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Shannon Haughton \email{shannon.l.haughton@gsk.com}
  \item Ellis Hughes \email{ellis.h.hughes@gsk.com} (\href{https://orcid.org/0000-0003-0637-4436}{ORCID})
  \item Alexandra Lauer \email{alexandralauer1@gmail.com} (\href{https://orcid.org/0000-0002-4191-6301}{ORCID})
  \item JooYoung Seo \email{jseo1005@illinois.edu} (\href{https://orcid.org/0000-0002-4064-6012}{ORCID})
  \item Ken Brevoort \email{ken@brevoort.com} (\href{https://orcid.org/0000-0002-4001-8358}{ORCID})
  \item Olivier Roy
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
