% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt}
\alias{gt}
\title{Create gt table object}
\usage{
gt(data, rownames_to_stub = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} object or a
tibble.}

\item{rownames_to_stub}{an option to take
rownames from the input \code{data} table
as row captions in the stub.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Create a gt table object so that we can perform
styling transformations before transforming
the entire object to a display table. Using this
function is the first step in that process.
}
\examples{
# Create a table object using the
# `mtcars` dataset
tab <-
  gt(data = mtcars)

# The resulting object can be used
# in transformations
tab_2 <-
  tab \%>\%
  cols_align_right() \%>\%
  fmt_numeric(
    columns = c("drat", "wt", "qsec"),
    decimals = 1)

# The object of this type can be
# displayed in the Viewer
tab_2
}
