% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt}
\alias{gt}
\title{Create gt table object}
\usage{
gt(data, rownames_to_stub = FALSE, split_cols_dots = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} object or a
tibble.}

\item{rownames_to_stub}{an option to take
rownames from the input \code{data} table
as row captions in the stub.}

\item{split_cols_dots}{an option to split
dot-separated column names such that the
first component is promoted to a spanner and
subsequent components represent the column label.
Please note that reference to individual columns
must still be via the original column names.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Create a gt table object so that we can perform
styling transformations before transforming
the entire object to a display table. Using this
function is the first step in that process.
}
\examples{
# Create a table object using the
# `mtcars` dataset
tab <-
  gt(mtcars, rownames_to_stub = TRUE)

# The resulting object can be used
# in transformations
tab_2 <-
  tab \%>\%
  cols_align_right() \%>\%
  fmt_number(
    columns = c("drat", "wt", "qsec"),
    decimals = 1)

# The object of this type can be
# displayed in the Viewer
tab_2
}
