% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_html_table.R
\name{gt}
\alias{gt}
\title{Create an HTML table object that can be styled}
\usage{
gt(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
an object of class \code{html_table}.
}
\description{
Create an HTML table object such that we can
perform styling transformations before transforming
the entire object to an HTML fragment. This is
the first step in the \code{gt()} -> ... ->
\code{emit_html()} pattern.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- gt(tbl = iris)

# The resulting object can be used
# in transformations
html_table_transformed <-
  html_table \%>\%
  add_column_style(
    columns = 1,
    property = "color",
    values = "blue")

# The object of this type can also be
# rendered in the Viewer
html_table_transformed \%>\%
  render_table()

# Once the table has been transformed
# suitably, we can extract it as HTML
# using `emit_html()`
html_table_transformed \%>\%
  emit_html()
}
