% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_function.R
\name{gt}
\alias{gt}
\title{Create an HTML table object that can be styled}
\usage{
gt(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
an object of class \code{html_table}.
}
\description{
Create an HTML table object such that we can
perform styling transformations before transforming
the entire object to an HTML fragment. This is
the first step in creating a display table.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- gt(tbl = airquality)

# The resulting object can be used
# in transformations
html_table_transformed <-
  html_table \%>\%
  apply_theme_minimal() \%>\%
  apply_alignment_right() \%>\%
  format_as_number(
    columns = "Wind",
    decimals = 1)

# The object of this type can be
# displayed in the Viewer
html_table_transformed
}
