% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt}
\alias{gt}
\title{Create a gt table object}
\usage{
gt(data, rownames_to_stub = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} object or a tibble.}

\item{rownames_to_stub}{an option to take rownames from the input \code{data}
table as row captions in the stub.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Create a gt table object so that we can perform styling transformations
before transforming the entire object to a display table. Using this function
is the first step in that process, and, there are a few data ingest options
we can consider at this stage.
}
\examples{
# Create a table object using the
# `mtcars` dataset
tab <-
  gt(mtcars, rownames_to_stub = TRUE)

# The resulting object can be used
# in transformations
tab_2 <-
  tab \%>\%
  fmt_number(
    columns = vars(drat, wt, qsec),
    decimals = 1)

# The object of this type can be
# displayed in the Viewer
tab_2
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_options}}, \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}},
  \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
