% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt}
\alias{gt}
\title{Create a gt table object}
\usage{
gt(data, rowname_col = "rowname", groupname_col = "groupname",
  rownames_to_stub = FALSE,
  stub_group.sep = getOption("gt.stub_group.sep", " - "))
}
\arguments{
\item{data}{a \code{data.frame} object or a tibble.}

\item{rowname_col}{the column name in the input \code{data} table to use as
row captions to be placed in the display table stub.}

\item{groupname_col}{the column name in the input \code{data} table to use as
group labels for generation of stub row groups.}

\item{rownames_to_stub}{an option to take rownames from the input \code{data}
table as row captions in the display table stub.}

\item{stub_group.sep}{the separator to use between consecutive group names
(a possibility when providing \code{data} as a \code{grouped_df} with
multiple groups) in the displayed stub row group label.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Create a gt table object so that we can perform styling transformations
before transforming the entire object to a display table. Using this function
is the first step in that process, and, there are a few data ingest options
we can consider at this stage.
}
\examples{
# Create a table object using the
# `mtcars` dataset; take the rownames
# available in this table and use them
# as row captions in the table stub
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE)

# The resulting object can be used
# in transformations (with `tab_*()`,
# `fmt_*()`, `cols_*()` functions)
gt_tbl_2 <-
  gt_tbl \%>\%
  tab_heading(title = "mtcars") \%>\%
  fmt_number(
    columns = vars(drat, wt, qsec),
    decimals = 1) \%>\%
  cols_label(hp = "HP")
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_options}}, \code{\link{tab_row_group}},
  \code{\link{tab_source_note}}, \code{\link{tab_spanner}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
