% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gt_latex_dependencies}
\alias{gt_latex_dependencies}
\title{Get the LaTeX dependencies required for a \strong{gt} table}
\usage{
gt_latex_dependencies()
}
\value{
An object of class \code{knit_asis}.
}
\description{
When working with Rnw (Sweave) files or otherwise writing LaTeX code,
including a \strong{gt} table can be problematic if we don't have knowledge
of the LaTeX dependencies. For the most part, these dependencies are the
LaTeX packages that are required for rendering a \strong{gt} table. The
\code{gt_latex_dependencies()} function provides an object that can be
used to provide the LaTeX in an Rnw file, allowing \strong{gt} tables to work
and not yield errors due to missing packages.
}
\details{
Here is an example Rnw document that shows how the
\code{gt_latex_dependencies()} can be used in conjunction with a \strong{gt}
table:

\preformatted{
\%!sweave=knitr

\documentclass{article}

<<echo=FALSE>>=
library(gt)
 @

<<results='asis', echo=FALSE>>=
gt_latex_dependencies()
 @

\begin{document}

<<results='asis', echo=FALSE>>=
exibble %>% gt()
 @

\end{document}
}
}
\section{Function ID}{

7-23
}

\seealso{
Other Helper Functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{google_font}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()}
}
\concept{Helper Functions}
