% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gt_output}
\alias{gt_output}
\title{Create a \pkg{gt} display table output element for Shiny}
\usage{
gt_output(outputId)
}
\arguments{
\item{outputId}{an output variable from which to read the table.}
}
\value{
A \pkg{gt} table output element that can be included in a panel.
}
\description{
Create a \pkg{gt} display table output element for Shiny
}
\seealso{
\link{render_gt}()

Other Shiny functions: \code{\link{render_gt}}
}
\concept{Shiny functions}
