% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pdf.R
\name{gt_png}
\alias{gt_png}
\title{Export the gt table as a PNG file}
\usage{
gt_png(data, file_name = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{file_name}{the output file name of the PNG file.}
}
\description{
Export the gt table as a PNG file
}
\examples{
\dontrun{
# Use `countrypops` to create a gt table;
# export the table as a PNG file
countrypops \%>\%
  dplyr::select(-contains("code")) \%>\%
  dplyr::filter(country_name == "Mongolia") \%>\%
  tail(5) \%>\%
  gt() \%>\%
  cols_align(
    align = "left",
    columns = vars(population)
  ) \%>\%
  gt_png(file_name = "countrypops.png")
}
}
\seealso{
Other table export functions: \code{\link{as_latex}},
  \code{\link{as_raw_html}}, \code{\link{as_rtf}},
  \code{\link{extract_summary}}, \code{\link{gt_pdf}}
}
\concept{table export functions}
