% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_preview.R
\name{gt_preview}
\alias{gt_preview}
\title{Preview a gt table object}
\usage{
gt_preview(data, top_n = 5, bottom_n = 1, incl_rownums = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} object or a tibble.}

\item{top_n}{this value will be used as the number of rows from the top of
the table to display. The default, \code{5}, will show the first five rows
of the table.}

\item{bottom_n}{the value will be used as the number of rows from the bottom
of the table to display. The default, \code{1}, will show the final row of
the table.}

\item{incl_rownums}{an option to include the row numbers for \code{data} in
the table stub.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Create a preview version of table. Here we can specify the first and last
\code{n} rows for the preview table.
}
\examples{
# Create a preview of the `mtcars`
# dataset; invoking the `gt_tbl`
# object will display it
gt_tbl <- gt_preview(mtcars)
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt}},
  \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_options}}, \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
