% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_preview.R
\name{gt_preview}
\alias{gt_preview}
\title{Generate a special \strong{gt} table for previewing a dataset}
\usage{
gt_preview(data, top_n = 5, bottom_n = 1, incl_rownums = TRUE)
}
\arguments{
\item{data}{\emph{Input data table}

\verb{obj:<data.frame>|obj:<tbl_df>} // \strong{required}

A \code{data.frame} object or a tibble (\code{tbl_df}).}

\item{top_n}{\emph{Top n rows to display}

\verb{scalar<numeric|integer>} // \emph{default:} \code{5}

The \code{top_n} value will be used as the number of rows from the top of the
table to display. The default, \code{5}, will show the first five rows of the
table.}

\item{bottom_n}{\emph{Bottom n rows to display}

\verb{scalar<numeric|integer>} // \emph{default:} \code{1}

The \code{bottom_n} value will be used as the number of rows from the bottom
of the table to display. The default, \code{1}, will show the final row of the
table.}

\item{incl_rownums}{\emph{Display row numbers}

\verb{scalar<logical>} // \emph{default:} \code{TRUE}

An option to include the row numbers for \code{data} in the table stub.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Sometimes you may want to see just a small portion of your input data. We can
use \code{gt_preview()} in place of \code{\link[=gt]{gt()}} to get the first x rows of data and the
last y rows of data (which can be set by the \code{top_n} and \code{bottom_n}
arguments). It's not advised to use additional \strong{gt} functions to further
modify the output of \code{gt_preview()}. Furthermore, you cannot pass a \strong{gt}
object to \code{gt_preview()}.
}
\details{
By default, the output table will include row numbers in a stub (including a
range of row numbers for the omitted rows). This row numbering option can be
deactivated by setting \code{incl_rownums} to \code{FALSE}.
}
\section{Examples}{


With three columns from the \code{\link{gtcars}} dataset, let's create a \strong{gt} table
preview with the \code{gt_preview()} function. You'll get only the first five rows
and the last row.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{gtcars |>
  dplyr::select(mfr, model, year) |>
  gt_preview()
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_gt_preview_1.png" alt="This image of a table was generated from the first code example in the `gt_preview()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

1-2
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other table creation functions: 
\code{\link{gt}()}
}
\concept{table creation functions}
