% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsave.R
\name{gtsave}
\alias{gtsave}
\title{Save a \pkg{gt} table as a file}
\usage{
gtsave(data, filename, ...)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{filename}{the file name to create on disk. Ensure that an extension
compatible with the output types is provided (\code{.html}, \code{.tex},
\code{.ltx}, \code{.rtf}). If a custom save function is provided then the
file extension is disregarded.}

\item{...}{all other options passed to the appropriate internal saving
function.}
}
\description{
The \code{gtsave()} function makes it easy to save a \pkg{gt} table to a
file. The function guesses the file type by the extension provided in the
output filename, producing either an HTML, LaTeX, or RTF file.
}
\details{
Output filenames with either the \code{.html} or \code{.htm} extensions will
produce an HTML document. In this case, we can pass a \code{TRUE} or
\code{FALSE} value to the \code{inline_css} option to obtain an HTML document
with inlined CSS styles (the default is \code{FALSE}). More details on CSS
inlining are available at \code{\link{as_raw_html}()}.

If the output filename is either of \code{.tex}, \code{.ltx}, or \code{.rnw},
a LaTeX document is produced. An output filename of \code{.rtf} will generate
an RTF document. The LaTeX and RTF saving function don't have any options to
pass to \code{...}.
}
