% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{html}
\alias{html}
\title{Interpret input text as HTML-formatted text}
\usage{
html(text)
}
\arguments{
\item{text}{the text that is understood to be HTML text, which is to be
preserved.}
}
\value{
a character object that is tagged as an HTML fragment that is not to
be sanitized.

a character object of class \code{preserve_html}.
}
\description{
Interpret input text as HTML-formatted text
}
