% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_tables.R
\name{info_date_style}
\alias{info_date_style}
\title{View a table with info on date styles}
\usage{
info_date_style()
}
\description{
View a table with info on date styles
}
\seealso{
Other information functions: \code{\link{info_paletteer}},
  \code{\link{info_time_style}}
}
\concept{information functions}
