% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_tables.R
\name{info_date_style}
\alias{info_date_style}
\title{View a table with info on date styles}
\usage{
info_date_style(locale = NULL)
}
\arguments{
\item{locale}{\emph{Locale identifier}

\verb{scalar<character>} // \emph{default:} \code{NULL} (\code{optional})

An optional locale identifier that can be used for displaying formatted
date values according to the locale's rules. Examples include \code{"en"} for
English (United States) and \code{"fr"} for French (France). We can call
\code{\link[=info_locales]{info_locales()}} for a useful reference for all of the locales that are
supported.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
\code{\link[=fmt_date]{fmt_date()}} lets us format date-based values in a convenient manner using
preset styles. The table generated by \code{info_date_style()} provides a quick
reference to all styles, with associated format names and example outputs
using a fixed date (\code{2000-02-29}).
}
\section{Examples}{


Get a table of info on the different date-formatting styles (which are used
by supplying a number code to \code{\link[=fmt_date]{fmt_date()}}).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{info_date_style()
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_info_date_style_1.png" alt="This image of a table was generated from the first code example in the `info_date_style()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

11-1
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other information functions: 
\code{\link{info_currencies}()},
\code{\link{info_flags}()},
\code{\link{info_google_fonts}()},
\code{\link{info_icons}()},
\code{\link{info_locales}()},
\code{\link{info_paletteer}()},
\code{\link{info_tf_style}()},
\code{\link{info_time_style}()},
\code{\link{info_unit_conversions}()}
}
\concept{information functions}
