% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_tables.R
\name{info_time_style}
\alias{info_time_style}
\title{View a table with info on time styles}
\usage{
info_time_style()
}
\description{
View a table with info on time styles
}
\seealso{
Other information functions: \code{\link{info_date_style}},
  \code{\link{info_paletteer}}
}
\concept{information functions}
