% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{location_cells}
\alias{location_cells}
\alias{cells_title}
\alias{cells_column_labels}
\alias{cells_group}
\alias{cells_stub}
\alias{cells_data}
\alias{cells_summary}
\title{Helpers for targeting multiple cells in different locations}
\usage{
cells_title(groups = c("title", "subtitle"))

cells_column_labels(columns, groups)

cells_group(groups)

cells_stub(rows = NULL)

cells_data(columns = NULL, rows = NULL)

cells_summary(groups = NULL, columns = NULL, rows = NULL)
}
\arguments{
\item{columns, rows, groups}{either a vector of names, a vector of
indices, values provided by \code{\link{vars}()}, values provided by
\code{c()}, or a select helper function (see Details for information on
these functions).}
}
\value{
a list object of class \code{location_cells}.
}
\description{
These helper functions are used to target individual cells in different
locations.
}
\details{
The following helper functions can be used to target cells (roughly in order
from the top to the bottom of a table):

\itemize{

\item \code{cells_title()}: targets the table title or the table subtitle
depending on the value given to the \code{groups} argument (\code{"title"} or
\code{"subtitle"}).

\item \code{cells_column_labels()}: targets labels in the column labels (the
\code{columns} argument) or the spanner column labels (the \code{groups}
argument) in the table's column labels part.

\item \code{cells_group()}: targets the row group labels in any available
row groups using the \code{groups} argument.

\item \code{cells_stub()}: targets row labels in the table stub using the
\code{rows} argument.

\item \code{cells_data()}: targets data cells in the table field using
intersections of \code{columns} and \code{rows}.

\item \code{cells_summary()}: targets summary cells in the table field using
the \code{groups} argument and intersections of \code{columns} and
\code{rows}.

}

The select helper functions are: \code{\link{starts_with}()},
  \code{\link{ends_with}()}, \code{\link{contains}()},
  \code{\link{matches}()}, \code{\link{one_of}()}, and
  \code{\link{everything}()}.
}
