% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{md}
\alias{md}
\title{Interpret input text as Markdown-formatted text}
\usage{
md(text)
}
\arguments{
\item{text}{the text that is understood to contain Markdown formatting.}
}
\value{
a character object that is tagged for a Markdown-to-HTML
  transformation.

a character object of class \code{from_markdown}.
}
\description{
Interpret input text as Markdown-formatted text
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a caption
# that is to interpreted as Markdown
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_stubhead_caption(
      caption = md("car *make* and *model*"))
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{html}}, \code{\link{pct}}, \code{\link{px}}
}
\concept{helper functions}
