% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nanoplot_options}
\alias{nanoplot_options}
\title{Supply nanoplot options to \code{cols_nanoplot()}}
\usage{
nanoplot_options(
  line_stroke = NULL,
  line_stroke_width = NULL,
  line_fill = NULL,
  data_point_radius = NULL,
  data_point_stroke = NULL,
  data_point_stroke_width = NULL,
  data_point_fill = NULL,
  vertical_guideline_stroke = NULL,
  vertical_guideline_stroke_width = NULL,
  show_data_points = NULL,
  show_curved_data_line = NULL,
  show_lower_area = NULL,
  show_vertical_guidelines = NULL,
  svg_margin_left = NULL,
  svg_margin_right = NULL,
  svg_font_size = NULL,
  svg_overflow = NULL,
  svg_position = NULL
)
}
\arguments{
\item{line_stroke}{The line stroke.}

\item{line_stroke_width}{The width of the line stroke.}

\item{line_fill}{The line fill.}

\item{data_point_radius}{The data point radius.}

\item{data_point_stroke}{The data point stroke color.}

\item{data_point_stroke_width}{The data point stroke width.}

\item{data_point_fill}{The data point fill width.}

\item{vertical_guideline_stroke}{The stroke color for the vertical guidelines.}

\item{vertical_guideline_stroke_width}{The stroke width for the vertical guidelines.}

\item{show_data_points}{Should the data points be shown?}

\item{show_curved_data_line}{Should a curved data line be shown?}

\item{show_lower_area}{Should the area below the data points filled in?}

\item{show_vertical_guidelines}{Should vertical guidelines be shown?}

\item{svg_margin_left}{The SVG left margin width.}

\item{svg_margin_right}{The SVG right margin width.}

\item{svg_font_size}{The SVG font size.}

\item{svg_overflow}{The SVG font overflow setting.}

\item{svg_position}{The SVG font position setting.}
}
\value{
A list object of class \code{nanoplot_options}.
}
\description{
When using \code{\link[=cols_nanoplot]{cols_nanoplot()}}, the defaults for the generated nanoplots
can be modified with \code{nanoplot_options()} within the \code{options} argument.
}
\section{Function ID}{

8-7
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_footnotes}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_source_notes}()},
\code{\link{cells_stub_grand_summary}()},
\code{\link{cells_stub_summary}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{define_units}()},
\code{\link{escape_latex}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()},
\code{\link{system_fonts}()}
}
\concept{helper functions}
