% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pct}
\alias{pct}
\title{Helper for providing a numeric value as percentage}
\usage{
pct(x)
}
\arguments{
\item{x}{the numeric value to format as a string percentage for some
\code{\link{tab_options}()} arguments that can take percentage values (e.g.,
\code{table.width}).}
}
\description{
Helper for providing a numeric value as percentage
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{columns_with}},
  \code{\link{data_cells}}, \code{\link{funs}},
  \code{\link{html}}, \code{\link{md}},
  \code{\link{not_in_group}}, \code{\link{px}},
  \code{\link{rownames_with}}, \code{\link{rows_where}},
  \code{\link{vars}}
}
\concept{helper functions}
