% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pct}
\alias{pct}
\title{Helper for providing a numeric value as percentage}
\usage{
pct(x)
}
\arguments{
\item{x}{the numeric value to format as a string percentage for some
\code{\link{tab_options}()} arguments that can take percentage values
(e.g., \code{table.width}).}
}
\description{
Helper for providing a numeric value as percentage
}
\seealso{
Other helper functions: \code{\link{cells_styles}},
  \code{\link{escape_latex}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{px}}
}
\concept{helper functions}
