% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gt_tbl}
\alias{print.gt_tbl}
\alias{print.gt_group}
\alias{print.rtf_text}
\title{Print a \strong{gt} table}
\usage{
\method{print}{gt_tbl}(x, ..., view = interactive())

\method{print}{gt_group}(x, ..., view = interactive())

\method{print}{rtf_text}(x, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Any additional parameters.}

\item{view}{The value for \code{print()}s \code{browse} argument.}
}
\description{
This facilitates printing of the HTML table to the R console.

This facilitates printing of multiple HTML tables (in a \code{gt_group} object) to
the R console.
}
\keyword{internal}
