% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.html_table}
\alias{print.html_table}
\title{Print the table}
\usage{
\method{print}{html_table}(x, ..., view = interactive())
}
\arguments{
\item{x}{an object of class \code{html_table}.}
}
\description{
This facilitates printing of the
HTML table to the R console.
}
\keyword{internal}
