% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{px}
\alias{px}
\title{Helper for providing a numeric value as pixels value}
\usage{
px(x)
}
\arguments{
\item{x}{the numeric value to format as a string (e.g., \code{"12px"}) for
some \code{\link{tab_options}()} arguments that can take values as units of
pixels (e.g., \code{table.font.size}).}
}
\description{
Helper for providing a numeric value as pixels value
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{columns_with}},
  \code{\link{contains}}, \code{\link{ends_with}},
  \code{\link{everything}}, \code{\link{funs}},
  \code{\link{html}}, \code{\link{matches}},
  \code{\link{md}}, \code{\link{not_in_group}},
  \code{\link{one_of}}, \code{\link{pct}},
  \code{\link{rownames_with}}, \code{\link{rows_where}},
  \code{\link{starts_with}}, \code{\link{vars}}
}
\concept{helper functions}
