% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{px}
\alias{px}
\title{Helper for providing a numeric value as pixels value}
\usage{
px(x)
}
\arguments{
\item{x}{the numeric value to format as a string (e.g., \code{"12px"}) for
some \code{\link{tab_options}()} arguments that can take values as units of
pixels (e.g., \code{table.font.size}).}
}
\description{
Helper for providing a numeric value as pixels value
}
\seealso{
Other helper functions: \code{\link{cells_styles}},
  \code{\link{escape_latex}},
  \code{\link{gt_latex_dependencies}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{pct}}
}
\concept{helper functions}
