% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_id}
\alias{random_id}
\title{Helper for creating a random \code{id} for a \strong{gt} table}
\usage{
random_id(n = 10)
}
\arguments{
\item{n}{\emph{Number of letters}

\verb{scalar<numeric|integer>} // \emph{default:} \code{10}

The \code{n} argument defines the number of lowercase letters to use for the
random ID.}
}
\value{
A character vector containing a single, random ID.
}
\description{
\code{random_id()} can be used to create a random, character-based ID value
argument of variable length (the default is 10 letters).
}
\section{Function ID}{

8-28
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{latex}()},
\code{\link{md}()},
\code{\link{nanoplot_options}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{row_group}()},
\code{\link{stub}()},
\code{\link{system_fonts}()},
\code{\link{unit_conversion}()}
}
\concept{helper functions}
