% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{render_gt}
\alias{render_gt}
\title{A \pkg{gt} display table render function for use in Shiny}
\usage{
render_gt(expr, env = parent.frame(), quoted = FALSE,
  outputArgs = list())
}
\arguments{
\item{expr}{an expression that creates a \code{gt} table object.}

\item{env}{the environment in which to evaluate the \code{expr}.}

\item{quoted}{is expr a quoted expression (with \code{quote()})? This is
useful if you want to save an expression in a variable.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link{gt_output}()} when \code{render_gt} is used in an
interactive R Markdown document.}
}
\description{
A \pkg{gt} display table render function for use in Shiny
}
\seealso{
\link{gt_output}()

Other Shiny functions: \code{\link{gt_output}}
}
\concept{Shiny functions}
