% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rownames_with}
\alias{rownames_with}
\title{Helper for targeting multiple rows with a pattern}
\usage{
rownames_with(pattern)
}
\arguments{
\item{pattern}{a pattern that can help capture one or
more row names and hence the rows themselves.}
}
\value{
a list object of class \code{rownames_with}.
}
\description{
Helper for targeting multiple rows with a pattern
}
