% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rownames_with}
\alias{rownames_with}
\title{Helper for targeting multiple rows with a pattern}
\usage{
rownames_with(pattern)
}
\arguments{
\item{pattern}{a pattern that can help capture one or more row names and
hence the rows themselves.}
}
\value{
a list object of class \code{rownames_with}.
}
\description{
Helper for targeting multiple rows with a pattern
}
\examples{
# With the `mtcars` dataset (preserving the
# rownames in the stub), we can target number
# formatting to those rows that fit a
# specified pattern
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  fmt_number(
    columns = vars(qsec, wt),
    rows = rownames_with("Merc"),
    decimals = 3)
}
\seealso{
Other helper functions: \code{\link{col_labels}},
  \code{\link{columns_with}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{not_in_group}},
  \code{\link{target_cell}}, \code{\link{vars}}
}
