% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rows_where}
\alias{rows_where}
\title{Helper for targeting rows based on conditional data}
\usage{
rows_where(conditions)
}
\arguments{
\item{conditions}{conditional expressions involving column names that can
help capture one or more rows.}
}
\value{
a quosure.
# With the `mtcars` dataset, apply
# number formatting to to `qsec` (3
# decimal places) to those rows where
# `cyl >= 6` and `vs == 1`
gt(mtcars, rownames_to_stub = TRUE) %>%
  fmt_number(
    columns = vars(qsec),
    rows = rows_where(cyl >= 6 & vs == 1),
    decimals = 3)
}
\description{
Helper for targeting rows based on conditional data
}
\seealso{
Other helper functions: \code{\link{apply_styles}},
  \code{\link{col_labels}}, \code{\link{columns_with}},
  \code{\link{contains}}, \code{\link{ends_with}},
  \code{\link{everything}}, \code{\link{funs}},
  \code{\link{html}}, \code{\link{matches}},
  \code{\link{md}}, \code{\link{one_of}},
  \code{\link{pct}}, \code{\link{px}},
  \code{\link{rownames_with}}, \code{\link{starts_with}},
  \code{\link{vars}}
}
\concept{helper functions}
