% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rx_addv}
\alias{rx_addv}
\title{An ADDV-flavored clinical trial toy dataset}
\format{
A tibble with 182 rows and 20 variables:
\describe{
\item{STUDYID, STUDYIDN}{The unique study identifier and its numeric
version.}
\item{USUBJID}{The unique subject identifier.}
\item{TRTA, TRTAN}{The study intervention and its numeric version, which
is either \code{"Placebo"} (\code{1}), \code{"Drug 1"} (\code{2}) or \code{NA} (\code{3}), missing for
screen failures).}
\item{ITTFL}{Intent-to-Treat (ITT) population flag, where \code{"Y"} indicates
a subject belongs to the ITT population and \code{"N"} indicates a subject
is not in the ITT population.}
\item{AGE}{The age of a subject at baseline in years.}
\item{AAGEGR1}{The analysis age group, indicating if a subject was strictly
younger than 40 years at baseline or older.}
\item{SEX}{Sex of a subject. Can be either \code{"Male"}, \code{"Female"} or
\code{"Undifferentiated"}.}
\item{ETHNIC}{Ethnicity of a subject. Can be either \code{"Hispanic or Latino"},
\code{"Not Hispanic or Latino"} or missing (\code{""}).}
\item{BLBMI}{Body Mass Index (BMI) of a subject at baseline in kg/m2.}
\item{DVTERM}{The Protocol Deviation Term.}
\item{PARAMCD, PARAM}{The Parameter Code and decoded parameter description
for the protocol deviation.}
\item{PARCAT1}{Parameter category. Can be \code{OVERALL} for derived PD
summaries or \verb{PROTOCOL DEVIATION} for individual PDs.}
\item{DVCAT}{Category for PD, indicating whether the PD is a major one or
not.}
\item{ACAT1}{Analysis category 1. Only populated for individual PDs, not
for summary scores. High level category for PDs.}
\item{AVAL}{Analysis Value. Either 0 or 1.}
\item{CRIT1, CRIT1FL}{Analysis Criterion 1 and analysis criterion 1 flag,
indicating whether PD is related to COVID-19 or not.}
}
}
\usage{
rx_addv
}
\description{
This tibble contains artificial protocol deviation data for 180 subjects
in the Intent-to-Treat (ITT) population of the \code{GT01} study. The dataset
contains the usual parameters (\code{PARAM}, \code{PARAMCD}) for an addv. There is
summary parameter (\code{PARCAT1 == "OVERALL"}) for each subject of the \code{GT01}
ITT-population, indicating whether or not at least one major protocol
deviation (PD) occurred throughout the course of the study for the respective
subject. Individual records for protocol deviations per subject exist,
indicating which specific type of PD occurred. The additional flag \code{CRIT1FL},
shows a PD was related to COVID-19 or not.

Although the data was intentionally created to mimic a typical clinical
trial dataset following the CDISC format, it might not strictly comply
with CDISC ADaM rules. The intent is to showcase the workflow for clinical
table creation rather than creating a fully CDISC-compliant ADaM dataset.
}
\section{Examples}{


Here is the \code{rx_addv}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{rx_addv
#> # A tibble: 291 x 20
#>    STUDYID STUDYIDN USUBJID TRTA    TRTAN ITTFL   AGE AAGEGR1 SEX   ETHNIC BLBMI
#>    <chr>   <chr>    <chr>   <fct>   <dbl> <chr> <int> <fct>   <fct> <fct>  <dbl>
#>  1 GT01    4001     GT1001  Placebo     1 Y        41 >=40    Male  Not H~  33.4
#>  2 GT01    4001     GT1002  Placebo     1 Y        39 <40     Fema~ Not H~  30.5
#>  3 GT01    4001     GT1002  Placebo     1 Y        39 <40     Fema~ Not H~  30.5
#>  4 GT01    4001     GT1003  Placebo     1 Y        38 <40     Male  Not H~  22.9
#>  5 GT01    4001     GT1003  Placebo     1 Y        38 <40     Male  Not H~  22.9
#>  6 GT01    4001     GT1003  Placebo     1 Y        38 <40     Male  Not H~  22.9
#>  7 GT01    4001     GT1004  Placebo     1 Y        45 >=40    Male  Not H~  23.9
#>  8 GT01    4001     GT1004  Placebo     1 Y        45 >=40    Male  Not H~  23.9
#>  9 GT01    4001     GT1005  Placebo     1 Y        35 <40     Fema~ Hispa~  29.1
#> 10 GT01    4001     GT1005  Placebo     1 Y        35 <40     Fema~ Hispa~  29.1
#> # ... with 281 more rows, and 9 more variables: DVTERM <chr>, PARAMCD <fct>,
#> #   PARAM <fct>, PARCAT1 <chr>, DVCAT <chr>, ACAT1 <chr>, AVAL <dbl>,
#> #   CRIT1 <chr>, CRIT1FL <chr>
}\if{html}{\out{</div>}}
}

\section{Function ID}{

11-7
}

\seealso{
Other datasets: 
\code{\link{countrypops}},
\code{\link{exibble}},
\code{\link{gtcars}},
\code{\link{pizzaplace}},
\code{\link{rx_adsl}},
\code{\link{sp500}},
\code{\link{sza}},
\code{\link{towny}}
}
\concept{datasets}
\keyword{datasets}
