% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rx_adsl}
\alias{rx_adsl}
\title{An ADSL-flavored clinical trial toy dataset}
\format{
A tibble with 182 rows and 14 variables:
\describe{
\item{STUDYID, STUDYIDN}{The unique study identifier and its numeric
version.}
\item{USUBJID}{The unique subject identifier.}
\item{TRTA, TRTAN}{The study intervention and its numeric version, which
is either \code{"Placebo"} (\code{1}), \code{"Drug 1"} (\code{2}) or \code{NA} (\code{3}), missing for
screen failures).}
\item{ITTFL}{Intent-to-Treat (ITT) population flag, where \code{"Y"} indicates
a subject belongs to the ITT population and \code{"N"} indicates a subject
is not in the ITT population.}
\item{RANDFL}{Randomization flag, where \code{"Y"} indicates a subject was
randomized to receive either \code{"Placebo"} or \code{"Drug 1"} and \code{"N"} indicates
a subject was not randomized at all.}
\item{SCRFREAS}{The reason for screen failure. This is either missing
(\code{""}) for non-screen failure subjects or indicates the reason for screen
failure}
\item{AGE}{The age of a subject at baseline in years.}
\item{AAGEGR1}{The analysis age group, indicating if a subject was strictly
younger than 40 years at baseline or older.}
\item{SEX}{Sex of a subject. Can be either \code{"Male"}, \code{"Female"} or
\code{"Undifferentiated"}.}
\item{ETHNIC}{Ethnicity of a subject. Can be either \code{"Hispanic or Latino"},
\code{"Not Hispanic or Latino"} or missing (\code{""}).}
\item{BLBMI}{Body Mass Index (BMI) of a subject at baseline in kg/m2.}
\item{EVNTFL}{Event Flag. Indicates whether the subject experienced a
specific event during the course of the study or not. Can be either \code{"Y"},
if the subject had the event, or \code{"N"}}
}
}
\usage{
rx_adsl
}
\description{
This tibble contains artificial data for 182 subjects of the \code{GT01} study.
Each row corresponds to demographic characteristics of a single trial
participant. Two out of 182 study participants were screen failures and
thus not treated, the rest of the study population was randomized with a
1:1 ratio to receive either \code{Placebo} (as comparator) or \verb{Drug 1}. The
dataset entails subject level demographics such as age, age group, sex,
ethnicity and body mass index (BMI) at baseline, as well as an event
flag, indicating whether the subject experienced a specific event
throughout the course of the study or not.

Although the data was intentionally created to mimic a typical clinical
trial dataset following the CDISC format, it might not strictly comply
with CDISC ADaM rules. The intent is to showcase the workflow for clinical
table creation rather than creating a fully CDISC-compliant ADaM dataset.
}
\section{Examples}{


Here is the \code{rx_adsl}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{rx_adsl
#> # A tibble: 182 x 14
#>    STUDYID STUDYIDN USUBJID TRTA  TRTAN ITTFL RANDFL SCRFR~1   AGE AAGEGR1 SEX  
#>    <chr>   <chr>    <chr>   <fct> <dbl> <chr> <chr>  <chr>   <int> <fct>   <fct>
#>  1 GT01    4001     GT1000  <NA>      3 N     N      "WITHD~    37 <40     Male 
#>  2 GT01    4001     GT1001  Plac~     1 Y     Y      ""         41 >=40    Male 
#>  3 GT01    4001     GT1002  Plac~     1 Y     Y      ""         39 <40     Fema~
#>  4 GT01    4001     GT1003  Plac~     1 Y     Y      ""         38 <40     Male 
#>  5 GT01    4001     GT1004  Plac~     1 Y     Y      ""         45 >=40    Male 
#>  6 GT01    4001     GT1005  Plac~     1 Y     Y      ""         35 <40     Fema~
#>  7 GT01    4001     GT1006  Plac~     1 Y     Y      ""         42 >=40    Fema~
#>  8 GT01    4001     GT1007  Plac~     1 Y     Y      ""         35 <40     Male 
#>  9 GT01    4001     GT1008  Plac~     1 Y     Y      ""         42 >=40    Male 
#> 10 GT01    4001     GT1009  Plac~     1 Y     Y      ""         38 <40     Fema~
#> # ... with 172 more rows, 3 more variables: ETHNIC <fct>, BLBMI <dbl>,
#> #   EVNTFL <chr>, and abbreviated variable name 1: SCRFREAS
}\if{html}{\out{</div>}}
}

\section{Function ID}{

11-7
}

\seealso{
Other datasets: 
\code{\link{countrypops}},
\code{\link{exibble}},
\code{\link{gtcars}},
\code{\link{pizzaplace}},
\code{\link{rx_addv}},
\code{\link{sp500}},
\code{\link{sza}},
\code{\link{towny}}
}
\concept{datasets}
\keyword{datasets}
