% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_cols_align_center}
\alias{set_cols_align_center}
\title{Set columns to be aligned to the center}
\usage{
set_cols_align_center(data, columns = NULL, types = NULL)
}
\description{
Set columns to be aligned to the center
}
