% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{set_cols_align_right}
\alias{set_cols_align_right}
\title{Set columns to be aligned right}
\usage{
set_cols_align_right(data, columns = NULL, types = NULL)
}
\arguments{
\item{data}{a table object that is created using the
\code{gt()} function.}

\item{columns}{an optional vector of column names
for which the right alignment should be applied.}

\item{types}{an optional vector of column types
for which the right alignment should be applied.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set columns to be aligned right
}
