% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{set_fmt}
\alias{set_fmt}
\title{Set a column format with a formatter function}
\usage{
set_fmt(data, columns, rows = TRUE, formatter)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{columns}{the specification for which columns are to be formatted.}

\item{rows}{an optional specification for which rows are to be formatted.}

\item{formatter}{a formatting function or a named list of functions.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a column format with a formatter function
}
\seealso{
Other data formatting functions: \code{\link{fmt_currency}},
  \code{\link{fmt_datetime}}, \code{\link{fmt_date}},
  \code{\link{fmt_missing}}, \code{\link{fmt_number}},
  \code{\link{fmt_percent}}, \code{\link{fmt_scientific}},
  \code{\link{fmt_time}}
}
