% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stub}
\alias{stub}
\title{Select helper for targeting the stub column}
\usage{
stub(column = NULL)
}
\arguments{
\item{column}{\emph{Stub column selection}

\verb{scalar<integer>} // \emph{default:} \code{NULL} (optional)

For tables with multi-column stubs, optionally specify which stub column to
target. Use \code{1} for the rightmost stub column, \code{2} for the second from
right, etc. If \code{NULL} (the default), all stub columns are selected.}
}
\value{
A character vector of class \code{"stub_column"}.
}
\description{
Should you need to target only the stub column for formatting or other
operations, the \code{stub()} select helper can be used. This obviates the need
to use the name of the column that was selected as the stub column.
}
\section{Examples}{


Create a tibble that has a \code{row} column (values from \code{1} to \code{6}), a \code{group}
column, and a \code{vals} column (containing the same values as in \code{row}).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl <-
  dplyr::tibble(
    row = 1:6,
    group = c(rep("Group A", 3), rep("Group B", 3)),
    vals = 1:6
  )
}\if{html}{\out{</div>}}

Create a \strong{gt} table with a two-column stub (incorporating the \code{row} and
\code{group} columns in that). Format the row labels of the stub with
\code{\link[=fmt_roman]{fmt_roman()}} to obtain Roman numerals in the stub; we're selecting the stub
column here with the \code{stub()} select helper.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl |>
  gt(rowname_col = "row", groupname_col = "group") |>
  fmt_roman(columns = stub()) |>
  tab_options(row_group.as_column = TRUE)
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_stub_1.png" alt="This image of a table was generated from the first code example in the `stub()` help file." style="width:100\%;">
}}

For multi-column stubs, you can set widths for all stub columns together or
target specific columns individually:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  dplyr::select(num, char, row, group) |>
  gt(rowname_col = c("group", "row")) |>
  cols_width(
    stub() ~ px(200),        # All stub columns get 200px
    everything() ~ px(100)
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_stub_2.png" alt="This image of a table was generated from the second code example in the `stub()` help file." style="width:100\%;">
}}

Or target specific stub columns (1 = rightmost, 2 = second from right):

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  dplyr::select(num, char, row, group) |>
  gt(rowname_col = c("group", "row")) |>
  cols_width(
    stub(1) ~ px(70),        # Rightmost stub column (row)
    stub(2) ~ px(200),       # Second stub column (group)
    everything() ~ px(100)
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_stub_3.png" alt="This image of a table was generated from the third code example in the `stub()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

8-10
}

\section{Function Introduced}{

\code{v0.8.0} (November 16, 2022)
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{from_column}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{latex}()},
\code{\link{md}()},
\code{\link{nanoplot_options}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{row_group}()},
\code{\link{system_fonts}()},
\code{\link{unit_conversion}()}
}
\concept{helper functions}
