% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rows.R
\name{summary_rows}
\alias{summary_rows}
\title{Add summary rows based on simple aggregations}
\usage{
summary_rows(data, groups = NULL, columns = NULL, funs,
  labels = NULL, decimals = 2, sep_mark = ",", dec_mark = ".")
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{groups}{the stub block groups heading names for which summary rows
will be added.}

\item{columns}{the columns for which the summaries should be calculated.}

\item{funs}{functions used for aggregations. This can include any of
these: \code{mean}, \code{min}, \code{max}, \code{median}, \code{sd},
\code{sum} or \code{n}.}

\item{labels}{a vector of stub labels for the summary rows. The vector length
should match the number of \code{funs} provided. If this is not provided then
the labels will be generated from the function names.}

\item{decimals}{an option to specify the exact number of decimal places to
use. The default number of decimal places is \code{2}.}

\item{sep_mark}{the mark to use as a separator between groups of digits.}

\item{dec_mark}{the character to use as a decimal mark.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a summary rows to one or more stub blocks by using the input data
already provided in the \code{gt()} function.
}
\examples{
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  tab_stub_block(
    group = "Mercs",
    rows = rownames_with("Merc")) \%>\%
  summary_rows(
    funs = dplyr::funs(mean))
}
\concept{row addition functions}
