% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rows.R
\name{summary_rows}
\alias{summary_rows}
\title{Add group-wise summary rows using aggregation functions}
\usage{
summary_rows(
  data,
  groups = everything(),
  columns = everything(),
  fns = NULL,
  fmt = NULL,
  missing_text = "---",
  formatter = NULL,
  ...
)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{groups}{The groups to consider for generation of group-wise summary
rows. By default this is set to \code{everything()}, which means that all
available groups will obtain summary rows. Providing the names of row
groups in \code{c()} will create a group-wise summary and generate summary rows
for the specified groups. Setting this to \code{TRUE} indicates that all
available groups will receive group-wise summary rows.}

\item{columns}{The columns for which the summaries should be calculated. By
default, this is every column that has data cells (given by
\code{everything()}).}

\item{fns}{Functions used for aggregations. This can include base functions
like \code{mean}, \code{min}, \code{max}, \code{median}, \code{sd}, or \code{sum} or any other
user-defined aggregation function. The function(s) should be supplied
within a \code{list()}. Within that list, we can specify the functions by use of
function names in quotes (e.g., \code{"sum"}), as bare functions (e.g., \code{sum}),
or as one-sided R formulas using a leading \code{~}. In the formula
representation, a \code{.} serves as the data to be summarized (e.g., \code{sum(., na.rm = TRUE)}). The use of named arguments is recommended as the names
will serve as summary row labels for the corresponding summary rows data
(the labels can derived from the function names but only when not providing
bare function names).}

\item{fmt}{Formatting directives as one-sided R formulas using a leading \code{~}
(e.g., \verb{~ fmt_number(., decimals = 3, use_seps = FALSE}).}

\item{missing_text}{The text to be used in place of \code{NA} values in summary
cells with no data outputs.}

\item{formatter}{A formatter function name. These can be any of the \verb{fmt_*()}
functions available in the package (e.g., \code{\link[=fmt_number]{fmt_number()}}, \code{\link[=fmt_percent]{fmt_percent()}},
etc.), or a custom function using \code{\link[=fmt]{fmt()}}. The options of a formatter can
be accessed through \code{...}.}

\item{...}{Values passed to the \code{formatter} function, where the provided
values are to be in the form of named vectors. For example, when using the
default \code{formatter} function, \code{\link[=fmt_number]{fmt_number()}}, options such as \code{decimals},
\code{use_seps}, and \code{locale} can be used.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Add summary rows to one or more row groups by using the table data and any
suitable aggregation functions. You choose how to format the values in the
resulting summary cells by use of a \code{formatter} function (e.g, \code{fmt_number},
etc.) and any relevant options.
}
\details{
Should we need to obtain the summary data for external purposes, the
\code{\link[=extract_summary]{extract_summary()}} function can be used with a \code{gt_tbl} object where summary
rows were added via \code{summary_rows()}.
}
\section{Examples}{


Use \code{\link{sp500}} to create a \strong{gt} table with row groups. Create the summary
rows labeled \code{min}, \code{max}, and \code{avg} by row group (where each each row group
is a week number) with the \code{summary_rows()} function.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sp500 \%>\%
  dplyr::filter(date >= "2015-01-05" & date <="2015-01-16") \%>\%
  dplyr::arrange(date) \%>\%
  dplyr::mutate(week = paste0("W", strftime(date, format = "\%V"))) \%>\%
  dplyr::select(-adj_close, -volume) \%>\%
  gt(
    rowname_col = "date",
    groupname_col = "week"
  ) \%>\%
  summary_rows(
    fns = list(
      "min",
      "max",
      list(label = "avg", fn = "mean")
    ),
    formatter = fmt_number,
    use_seps = FALSE
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_summary_rows_1.png" alt="This image of a table was generated from the first code example in the `summary_rows()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

5-1
}

\seealso{
Other row addition/modification functions: 
\code{\link{grand_summary_rows}()},
\code{\link{row_group_order}()}
}
\concept{row addition/modification functions}
