% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{system_fonts}
\alias{system_fonts}
\title{Get a themed font stack that works well across systems}
\usage{
system_fonts(name)
}
\value{
A character vector of font names.
}
\description{
The vector of fonts given by \code{default_fonts()} should be used with a \strong{gt}
table that is rendered to HTML. We can specify additional fonts to use but
this default set should be placed after that to act as fallbacks. This is
useful when specifying \code{font} values in the \code{\link[=cell_text]{cell_text()}} function (itself
used in the \code{\link[=tab_style]{tab_style()}} function). If using \code{\link[=opt_table_font]{opt_table_font()}} (which also
has a \code{font} argument) we probably don't need to specify this vector of fonts
since it is handled by its \code{add} option (which is \code{TRUE} by default).
}
\section{Examples}{


Use \code{\link{exibble}} to create a \strong{gt} table.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  dplyr::select(char, time) |>
  gt() |>
  tab_style(
    style = cell_text(
      font = system_fonts(name = "didone")
    ),
    locations = cells_body(columns = time)
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_system_fonts_1.png" alt="This image of a table was generated from the first code example in the `system_fonts()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

8-29
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other helper functions: 
\code{\link{adjust_luminance}()},
\code{\link{cell_borders}()},
\code{\link{cell_fill}()},
\code{\link{cell_text}()},
\code{\link{cells_body}()},
\code{\link{cells_column_labels}()},
\code{\link{cells_column_spanners}()},
\code{\link{cells_footnotes}()},
\code{\link{cells_grand_summary}()},
\code{\link{cells_row_groups}()},
\code{\link{cells_source_notes}()},
\code{\link{cells_stub_grand_summary}()},
\code{\link{cells_stub_summary}()},
\code{\link{cells_stubhead}()},
\code{\link{cells_stub}()},
\code{\link{cells_summary}()},
\code{\link{cells_title}()},
\code{\link{currency}()},
\code{\link{default_fonts}()},
\code{\link{escape_latex}()},
\code{\link{google_font}()},
\code{\link{gt_latex_dependencies}()},
\code{\link{html}()},
\code{\link{md}()},
\code{\link{pct}()},
\code{\link{px}()},
\code{\link{random_id}()},
\code{\link{stub}()}
}
\concept{helper functions}
