% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_boxhead_panel}
\alias{tab_boxhead_panel}
\title{Arrange a boxhead into panels}
\usage{
tab_boxhead_panel(data, group, columns)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{group}{the name to assign to the spanner heading.}

\item{columns}{the columns to be components of the spanner heading.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Set a spanner with a name and mappings to columns extant in the table. This
creates a boxhead panel with spanner headings and column headings.
}
\examples{
# Create a table based on `rock` where
# there are column headings grouped
# under spanner headings
gt_tbl <-
  gt(data = rock) \%>\%
    tab_boxhead_panel(
      group = "perimeter",
      columns = c("peri", "shape"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_footnote}},
  \code{\link{tab_heading}}, \code{\link{tab_options}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}},
  \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
