% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{tab_boxhead_panel}
\alias{tab_boxhead_panel}
\title{Arrange a boxhead into panels}
\usage{
tab_boxhead_panel(html_tbl, ..., use_names = FALSE)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{...}{an optional series of named vectors
for specifying the mappings between column names
to spanner headings.}

\item{use_names}{an options to use the the
naming convention of \code{[spanner_heading].[column_heading]}
to generate spanner headings that correctly
enclose column headings.}
}
\value{
an object of class \code{html_table}.
}
\description{
If column names in the \code{tbl} supplied
use the naming convention of
\code{[spanner_heading].[column_heading]},
we can invoke \code{tab_boxhead_panel()}
with the \code{use_names} option set to
\code{TRUE}. This allows for parsing and
expansion of column headings into one or
more boxhead panels with spanner headings
and column headings.
}
\examples{
# Create a table based on `mtcars` where
# there are column headings grouped under
# spanner headings
tab_create(tbl = mtcars) \%>\%
  cols_remove(c("drat", "wt", "vs", "am")) \%>\%
  tab_boxhead_panel(
    "numbers" = c("cyl", "gear", "carb"),
    "measurements" = c(
      "hp", "qsec", "mpg", "disp")) \%>\%
  cols_reorder(columns = c(
    "mpg", "disp", "hp", "qsec",
    "cyl", "gear", "carb")) \%>\%
  theme_striped()
}
