% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_create_modify.R
\name{tab_caption}
\alias{tab_caption}
\title{Add or remove a table caption}
\usage{
tab_caption(data, caption)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{caption}{An optional table caption to use for cross-referencing
in R Markdown documents and \strong{bookdown} book projects.}
}
\value{
For \code{NULL} or character \code{caption} arguments, an object of class
\code{gt_tbl}.  For a missing \code{caption} argument, the current caption.
}
\description{
Add or remove a table caption
}
\details{
If \code{caption} is \code{NULL}, the caption is removed.  If \code{caption} is missing, the
current caption is returned.  Otherwise, caption of the current gt is created
or modified.
}
\seealso{
Other part creation/modification functions: 
\code{\link{tab_footnote}()},
\code{\link{tab_header}()},
\code{\link{tab_options}()},
\code{\link{tab_row_group}()},
\code{\link{tab_source_note}()},
\code{\link{tab_spanner_delim}()},
\code{\link{tab_spanner}()},
\code{\link{tab_stub_indent}()},
\code{\link{tab_stubhead}()},
\code{\link{tab_style}()}
}
\concept{part creation/modification functions}
