% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_create.R
\name{tab_create}
\alias{tab_create}
\title{Create gt table object}
\usage{
tab_create(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} object or a
tibble.}
}
\value{
an object of class \code{html_table}.
}
\description{
Create a gt table object so that we can perform
styling transformations before transforming
the entire object to a display table. Using this
function is the first step in that process.
}
\examples{
# Create an html table object using the
# iris dataset
html_table <- tab_create(tbl = airquality)

# The resulting object can be used
# in transformations
html_table_transformed <-
  html_table \%>\%
  theme_minimal() \%>\%
  cols_align_right() \%>\%
  fmt_number(
    columns = "Wind",
    decimals = 1)

# The object of this type can be
# displayed in the Viewer
html_table_transformed
}
