% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{tab_footnote}
\alias{tab_footnote}
\title{Add one or more footnotes}
\usage{
tab_footnote(html_tbl, glyph = "number", ..., .row = NULL, .column = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{glyph}{the type of glyph or the literal
glyph to associate with the targeted cells,
rows, or columns.}

\item{...}{a series of named vectors for
specifying the mappings between footnotes and
the targeted cells, rows, or columns.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a footnote with a glyph attached to the
targeted cells, rows, or columns.
}
