% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_footnote}
\alias{tab_footnote}
\title{Add a table footnote}
\usage{
tab_footnote(data, footnote, location)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{footnote}{text to be used in the footnote.}

\item{location}{the cell or set of cells to be associated with the footnote.
Supplying an object with the \code{\link{data_cells}()} helper function is
a useful way to specify the cell that is associated with the footnote.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a footnote with a glyph attached to the targeted cells, rows, or columns.
}
\examples{
# Add a footnote that is in
# reference to a single table cell
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_footnote(
      footnote = "Massive hp.",
      location = data_cells(
        row = "Maserati Bora",
        column = "hp"))
}
\seealso{
[data_cells()] as a useful helper function for targeting the cell
associated with the footnote.

Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_boxhead_panel}},
  \code{\link{tab_heading}}, \code{\link{tab_options}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}},
  \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
