% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_footnote.R
\name{tab_footnote}
\alias{tab_footnote}
\title{Add a table footnote}
\usage{
tab_footnote(data, footnote, locations)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{footnote}{the text to be used in the footnote. We can optionally use
the \code{\link{md}()} and \code{\link{html}()} functions to style the text
as Markdown or to retain HTML elements in the footnote text.}

\item{locations}{the cell or set of cells to be associated with the footnote.
Supplying any of the \code{cells_*()} helper functions is a useful way to
target the location cells that are associated with the footnote text. These
helper functions are: \code{\link{cells_title}()},
\code{\link{cells_boxhead}()}, \code{\link{cells_group}()},
\code{\link{cells_stub}()}, \code{\link{cells_data}()}, and
\code{\link{cells_summary}()}. Please see the help article
\link{location_cells} for more information on how these helper functions
can be used. Additionally, we can enclose several \code{cells_*()} calls
within a \code{list()} if we wish to link the footnote text to different
types of locations (e.g., cell data values, stub group headings, the table
title, etc.).}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a footnote with a glyph attached to the targeted cells, rows, or columns.
}
\examples{
# Add a footnote that is in
# reference to a single table cell
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_footnote(
      footnote = "Massive hp.",
      locations = cells_data(
        columns = "hp",
        rows = "Maserati Bora"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_header}},
  \code{\link{tab_options}}, \code{\link{tab_row_group}},
  \code{\link{tab_source_note}}, \code{\link{tab_spanner}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
