% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{tab_heading}
\alias{tab_heading}
\title{Add content to the heading part of the table}
\usage{
tab_heading(html_tbl, title, headnote = NULL, table_number = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{title}{text to be used in the table title.}

\item{headnote}{text to be used as the table's
headnote.}

\item{table_number}{a number to use as the
table number.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a title, headnote, and table number
information to the heading part of the table.
}
