% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_heading}
\alias{tab_heading}
\title{Add a table heading}
\usage{
tab_heading(data, title, headnote = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{title, headnote}{text to be used in the table title and, optionally, for
the table headnote (a line of text below the table title which is somewhat
smaller in size). We can optionally use the \code{\link{md}()} and
\code{\link{html}()} functions to style the text as Markdown or to retain
HTML elements in the text.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a title and optional headnote to the heading part of the table.
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a heading
# to describe the table
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_heading(
      title = md("Data listing from **mtcars**"),
      headnote = md("`mtcars` is an R dataset"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_options}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
