% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_options.R
\name{tab_options}
\alias{tab_options}
\title{Modify the table output options}
\usage{
tab_options(data, table.background.color = NULL, table.width = NULL,
  heading.background.color = NULL, heading.title.font.size = NULL,
  heading.title.background.color = NULL, heading.headnote.font.size = NULL,
  heading.headnote.background.color = NULL, heading.border.top = NULL,
  heading.border.bottom = NULL, stub_heading.background.color = NULL,
  stub_heading.border.top = NULL, stub_heading.border.bottom = NULL,
  stub_heading_field.background.color = NULL,
  stub_heading_field.border.top = NULL,
  stub_heading_field.border.bottom = NULL, field.border.top = NULL,
  field.border.bottom = NULL, row.padding = NULL, row.margin = NULL,
  row.striping.background.color = NULL, row.striping.include.stub = NULL,
  row.striping.include_field = NULL, summary_row.background.color = NULL,
  summary_row.padding = NULL, summary_row.text_transform = NULL,
  footnote.font.size = NULL, footnote.padding = NULL,
  sourcenote.font.size = NULL, sourcenote.padding = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()} function.}

\item{table.background.color}{the table's background color.}

\item{table.width}{the width of the table.}

\item{heading.background.color}{the background color of the table heading.}

\item{heading.title.font.size}{the font size of the title in the table
heading.}

\item{heading.title.background.color}{the background color of the title in
the table heading.}

\item{heading.headnote.font.size}{the font size of the headnote in the table
heading.}

\item{heading.headnote.background.color}{the background color of the headnote
in the table heading.}

\item{heading.border.top}{the parameters for the top border of the table
heading.}

\item{heading.border.bottom}{the parameters for the bottom border of the
table heading.}

\item{stub_heading.background.color}{the background color of the stub
heading.}

\item{stub_heading.border.top}{the parameters for the top border of the stub
heading.}

\item{stub_heading.border.bottom}{the parameters for the bottom border of the
stub heading.}

\item{stub_heading_field.background.color}{the background color of the stub
heading's field component.}

\item{stub_heading_field.border.top}{the parameters for the top border of the
stub heading's field component.}

\item{stub_heading_field.border.bottom}{the parameters for the bottom border
of the stub heading's field component.}

\item{field.border.top}{the parameters for the top border of the field.}

\item{field.border.bottom}{the parameters for the bottom border of the field.}

\item{row.padding}{the amount of padding in each row.}

\item{row.margin}{the size of the margin in each row.}

\item{row.striping.include_field}{an option for whether to include the field
when striping rows.}

\item{summary_row.background.color}{the background color of the summary rows.}

\item{summary_row.padding}{the amount of padding in each summary row.}

\item{summary_row.text_transform}{an option to apply text transformations to
the label text in each summary row.}

\item{footnote.font.size}{the font size of the footnotes in the footnote
section.}

\item{footnote.padding}{the amount of padding to apply to the footnote
section.}

\item{sourcenote.font.size}{the font size of the source notes in the source
note section.}

\item{sourcenote.padding}{the amount of padding to apply to the source note
section.}

\item{row.striping.color}{the color of the background for the striped rows.}

\item{row.striping.include_stub}{an option for whether to include the stub
when striping rows.}
}
\description{
Modify the options available in a table theme.
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}},
  \code{\link{tab_style}}
}
