% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_options.R
\name{tab_options}
\alias{tab_options}
\title{Modify the table output options}
\usage{
tab_options(data, table.font.size = NULL,
  table.background.color = NULL, table.width = NULL,
  table.border.top.style = NULL, table.border.top.width = NULL,
  table.border.top.color = NULL, heading.background.color = NULL,
  heading.title.font.size = NULL, heading.headnote.font.size = NULL,
  heading.border.bottom.style = NULL,
  heading.border.bottom.width = NULL,
  heading.border.bottom.color = NULL, boxhead.background.color = NULL,
  boxhead.font.size = NULL, boxhead.font.weight = NULL,
  stub_group.background.color = NULL, stub_group.font.size = NULL,
  stub_group.font.weight = NULL, stub_group.border.top.style = NULL,
  stub_group.border.top.width = NULL,
  stub_group.border.top.color = NULL,
  stub_group.border.bottom.style = NULL,
  stub_group.border.bottom.width = NULL,
  stub_group.border.bottom.color = NULL, field.border.top.style = NULL,
  field.border.top.width = NULL, field.border.top.color = NULL,
  field.border.bottom.style = NULL, field.border.bottom.width = NULL,
  field.border.bottom.color = NULL, row.padding = NULL,
  row.striping.background.color = NULL,
  row.striping.include.stub = NULL, row.striping.include_field = NULL,
  summary_row.background.color = NULL, summary_row.padding = NULL,
  summary_row.text_transform = NULL, footnote.font.size = NULL,
  footnote.padding = NULL, sourcenote.font.size = NULL,
  sourcenote.padding = NULL)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()} function.}

\item{table.font.size}{the table text font size. Can be specified as a
single-length character with units of pixels or as a percentage. If provided
as a single-length numeric vector, it is assumed that the value is given in
units of pixels.}

\item{table.background.color}{the table background color.}

\item{table.width}{the width of the table. Can be specified as a
single-length character with units of pixels or as a percentage. If provided
as a single-length numeric vector, it is assumed that the value is given in
units of pixels.}

\item{table.border.top.style}{the style of the table's top border.}

\item{table.border.top.width}{the width of the table's top border.}

\item{table.border.top.color}{the color of the table's top border.}

\item{heading.background.color}{the background color of the table heading.}

\item{heading.title.font.size}{the font size of the heading title.}

\item{heading.headnote.font.size}{the font size of the headnote.}

\item{heading.border.bottom.style}{the style of the heading's bottom border.}

\item{heading.border.bottom.width}{the width of the heading's bottom border.}

\item{heading.border.bottom.color}{the color of the heading's bottom border.}

\item{boxhead.background.color}{the background color of the boxhead.}

\item{boxhead.font.size}{the font size of the boxhead labels.}

\item{boxhead.font.weight}{the font weight of the boxhead labels.}

\item{stub_group.font.size}{the font size of the stub heading labels.}

\item{stub_group.font.weight}{the font weight of the stub heading labels.}

\item{stub_group.border.top.style}{the style of the stub heading's top border.}

\item{stub_group.border.top.width}{the width of the stub heading's top border.}

\item{stub_group.border.top.color}{the color of the stub heading's top border.}

\item{stub_group.border.bottom.style}{the style of the stub heading's bottom border.}

\item{stub_group.border.bottom.width}{the width of the stub heading's bottom border.}

\item{stub_group.border.bottom.color}{the color of the stub heading's bottom border.}

\item{field.border.top.style}{the style of the field's top border.}

\item{field.border.top.width}{the width of the field's top border.}

\item{field.border.top.color}{the color of the field's top border.}

\item{field.border.bottom.style}{the style of the field's bottom border.}

\item{field.border.bottom.width}{the width of the field's bottom border.}

\item{field.border.bottom.color}{the color of the field's bottom border.}

\item{row.padding}{the amount of padding in each row.}

\item{row.striping.include_field}{an option for whether to include the field
when striping rows.}

\item{summary_row.background.color}{the background color of the summary rows.}

\item{summary_row.padding}{the amount of padding in each summary row.}

\item{summary_row.text_transform}{an option to apply text transformations to
the label text in each summary row.}

\item{footnote.font.size}{the font size of the footnotes in the footnote
section.}

\item{footnote.padding}{the amount of padding to apply to the footnote
section.}

\item{sourcenote.font.size}{the font size of the source notes in the source
note section.}

\item{sourcenote.padding}{the amount of padding to apply to the source note
section.}

\item{field.background.color}{the background color of the field.}

\item{row.striping.color}{the color of the background for the striped rows.}

\item{row.striping.include_stub}{an option for whether to include the stub
when striping rows.}
}
\description{
Modify the options available in a table theme.
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}},
  \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
