% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_part.R
\name{tab_source_note}
\alias{tab_source_note}
\title{Add a source note citation}
\usage{
tab_source_note(html_tbl, source_note, lead_in = NULL, font = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{source_note}{text to be used in the
source note.}

\item{lead_in}{a lead in word or phrase
indicating the citation as the source of the
tabulated data.}

\item{font}{the name of the font to use for
the specified columns. This could be provided
as a vector of fonts where subsequent font names
provide fallbacks in the case that fonts are
not available.}
}
\value{
an object of class \code{html_table}.
}
\description{
Add a source note citation to the source note
part of the table.
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a source note
# to describe the source of the data
tab_create(tbl = mtcars) \%>\%
  theme_striped() \%>\%
  tab_source_note(
    source_note = "*Henderson and Velleman* (1981)."
    )
}
