% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_source_note}
\alias{tab_source_note}
\title{Add a source note citation}
\usage{
tab_source_note(data, source_note)
}
\arguments{
\item{data}{a table object that is created using the \code{gt()} function.}

\item{source_note}{text to be used in the source note.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a source note citation to the source note part of the table.
}
\examples{
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  tab_source_note(
    source_note = md("*Henderson and Velleman* (1981)."))
}
