% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_source_note}
\alias{tab_source_note}
\title{Add a source note citation}
\usage{
tab_source_note(data, source_note)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{source_note}{text to be used in the source note. We can optionally use
the \code{\link{md}()} and \code{\link{html}()} functions to style the text
as Markdown or to retain HTML elements in the text.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a source note citation to the source note part of the table.
}
\examples{
# Add a source note that provides
# a citation for the tabular data
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_source_note(
      source_note = md("*Henderson and Velleman* (1981)."))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_footnote}},
  \code{\link{tab_heading}}, \code{\link{tab_options}},
  \code{\link{tab_spanner}}, \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_label}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
