% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transforms.R
\name{tab_stub_block}
\alias{tab_stub_block}
\title{Arrange a table stub into blocks}
\usage{
tab_stub_block(html_tbl, ..., .default = NULL)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}

\item{...}{a series of named vectors for
specifying the mappings between row names or
row numbers to subhead names.}

\item{.default}{a subhead name to apply to
those rows not encapsulated \code{...}.}
}
\value{
an object of class \code{html_table}.
}
\description{
Arrange a table stub into blocks
}
\examples{
# Create an `mtcars` presentation table that
# groups certain rows together in stub blocks
mtcars_tbl <-
  tab_create(tbl = mtcars) \%>\%  # 1
  theme_striped() \%>\%
  tab_stub_block(
    "Mazda" = rownames_with("Mazda"),
    "Mercs" = rownames_with("Merc"),
    "Hornets" = rownames_with("Hornet"),
    "Toyotas" = rownames_with("Toyota"),
    "Supercars" = c(
      "Ferrari Dino","Maserati Bora",
      "Porsche 914-2"))
}
