% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_stubhead_caption}
\alias{tab_stubhead_caption}
\title{Add caption text to the stubhead}
\usage{
tab_stubhead_caption(data, caption)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()}
function.}

\item{caption}{the text to be used as the stubhead caption.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add a caption to the stubhead of a table. If a stub does not exist, no
  change will be made.
}
\examples{
# Create a table object using the
# `mtcars` dataset and add a caption
# to describe what is in the stub
gt_tbl <-
  gt(mtcars, rownames_to_stub = TRUE) \%>\%
    tab_stubhead_caption(
      caption = md("car *make* and *model*"))
}
\seealso{
Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_options}}, \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}}, \code{\link{tab_style}}
}
\concept{table-part creation/modification functions}
