% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_parts.R
\name{tab_style}
\alias{tab_style}
\title{Add custom styles to one or more cells}
\usage{
tab_style(data, style, location)
}
\arguments{
\item{data}{a table object that is created using the \code{\link{gt}()} function.}

\item{style}{a vector of styles to use. The \code{\link{apply_styles}()}
helper function can be used here to more easily generate valid styles.}

\item{location}{the cell or set of cells to be styled. Supplying an object
with the \code{data_cells()} helper function is a useful way to specify the
cell that is to be styled.}
}
\value{
an object of class \code{gt_tbl}.
}
\description{
Add custom styles to one or more cells
}
\examples{
# Add a style that is to be applied
# to a single table cell
gt(mtcars, rownames_to_stub = TRUE) \%>\%
  tab_style(
    style = apply_styles(bkgd_color = "steelblue"),
    location = data_cells(
      row = "Maserati Bora",
      column = vars(hp)))
}
\seealso{
[apply_styles()] as a helper for defining custom styles and
[data_cells()] as a useful helper function for targeting the cells to be
styled.

Other table-part creation/modification functions: \code{\link{gt_preview}},
  \code{\link{gt}}, \code{\link{tab_boxhead_panel}},
  \code{\link{tab_footnote}}, \code{\link{tab_heading}},
  \code{\link{tab_options}}, \code{\link{tab_source_note}},
  \code{\link{tab_stub_block}},
  \code{\link{tab_stubhead_caption}}
}
\concept{table-part creation/modification functions}
