% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{target_cell}
\alias{target_cell}
\title{Helper for targeting a single table cell}
\usage{
target_cell(row = NULL, column = NULL)
}
\arguments{
\item{row}{a single row to target.}

\item{column}{a single column to target.}
}
\value{
a list object of class \code{single_cell_target}.
}
\description{
Helper for targeting a single table cell
}
\seealso{
Other helper functions: \code{\link{col_labels}},
  \code{\link{columns_with}}, \code{\link{html}},
  \code{\link{md}}, \code{\link{not_in_group}},
  \code{\link{rownames_with}}, \code{\link{vars}}
}
