% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{test_image}
\alias{test_image}
\title{Generate a path to a test image}
\usage{
test_image(type = c("png", "svg"))
}
\arguments{
\item{type}{\emph{The image type}

\verb{singl-kw:[png|svg]} // \emph{default:} \code{"png"}

The type of image to produce here can either be \code{"png"} (the default) or
\code{"svg"}.}
}
\value{
A character vector with a single path to an image file.
}
\description{
Two test images are available within the \strong{gt} package. Both contain the
same imagery (sized at 200px by 200px) but one is a PNG file while the other
is an SVG file. This function is most useful when paired with \code{\link[=local_image]{local_image()}}
since we test various sizes of the test image within that function.
}
\section{Function ID}{

9-4
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other image addition functions: 
\code{\link{ggplot_image}()},
\code{\link{local_image}()},
\code{\link{web_image}()}
}
\concept{image addition functions}
