% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{text_replace}
\alias{text_replace}
\title{Perform highly targeted text replacement with a regex pattern}
\usage{
text_replace(data, pattern, replacement, locations = cells_body())
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{locations}{The cell or set of cells to be associated with the text
transformation. Only the \code{\link[=cells_body]{cells_body()}}, \code{\link[=cells_stub]{cells_stub()}},
\code{\link[=cells_column_labels]{cells_column_labels()}}, and \code{\link[=cells_row_groups]{cells_row_groups()}} helper functions can be
used here. We can enclose several of these calls within a \code{list()} if we
wish to make the transformation happen at different locations.}
}
\description{
Perform highly targeted text replacement with a regex pattern
}
\section{Function ID}{

4-2
}

\section{Function Introduced}{

\emph{In Development}
}

\seealso{
Other text transforming functions: 
\code{\link{text_case_match}()},
\code{\link{text_case_when}()},
\code{\link{text_transform}()}
}
\concept{text transforming functions}
