% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{text_transform}
\alias{text_transform}
\title{Perform text transformations with a custom function}
\usage{
text_transform(data, fn, locations = cells_body())
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{fn}{The function to use for text transformation.}

\item{locations}{The cell or set of cells to be associated with the text
transformation. Only the \code{\link[=cells_body]{cells_body()}}, \code{\link[=cells_stub]{cells_stub()}},
\code{\link[=cells_column_labels]{cells_column_labels()}}, and \code{\link[=cells_row_groups]{cells_row_groups()}} helper functions can be
used here. We can enclose several of these calls within a \code{list()} if we
wish to make the transformation happen at different locations.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Perform text transformations with a custom function
}
\section{Examples}{


Use \code{\link{exibble}} to create a \strong{gt} table. Transform the formatted text in the
\code{num} column using a function supplied to \code{text_transform()} (via the \code{fn}
argument). Note that the \code{x} in the \verb{fn = function (x)} part is an already
formatted vector of column values from the \code{num} column.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exibble |>
  dplyr::select(num, char, currency) |>
  dplyr::slice(1:4) |>
  gt() |>
  fmt_number(columns = num) |>
  fmt_currency(columns = currency) |>
  text_transform(
    fn = function(x) \{
      paste0(
        x, " (",
        dplyr::case_when(
          x > 20  ~ "large",
          x <= 20 ~ "small"),
        ")"
        )
    \},
    locations = cells_body(columns = num)
  )
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/rstudio/gt/master/images/man_text_transform_1.png" alt="This image of a table was generated from the first code example in the `text_transform()` help file." style="width:100\%;">
}}
}

\section{Function ID}{

4-1
}

\section{Function Introduced}{

\code{v0.2.0.5} (March 31, 2020)
}

\seealso{
Other text transforming functions: 
\code{\link{text_case_match}()},
\code{\link{text_case_when}()},
\code{\link{text_replace}()}
}
\concept{text transforming functions}
