% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_aesthetics.R
\name{theme_classical}
\alias{theme_classical}
\title{Use the 'classical' table theme}
\usage{
theme_classical(html_tbl)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{tab_create()} function.}
}
\value{
an object of class \code{html_table}.
}
\description{
Apply the 'classical' theme to the table.
Horizontal and vertical line styles
visually separate the columns and rows. This
adds reasonably thick horizontal and vertical
rules.
}
\examples{
# Create a table object using the
# `mtcars` dataset and use the
# 'classical' theme
tab_create(tbl = mtcars) \%>\%
  theme_classical()
}
