% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vars}
\alias{vars}
\title{Helper for targeting a series of row labels or columns}
\usage{
vars(...)
}
\arguments{
\item{...}{one or more column names or row labels, depending on the enclosing
function. This can be provided as bare column names or row labels.}
}
\description{
See \code{dplyr::\link[dplyr]{vars}} for details.
}
\keyword{internal}
