% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vec.R
\name{x_markdown}
\alias{x_markdown}
\title{Format a vector containing Markdown text}
\usage{
x_markdown(x, output = c("html", "latex", "rtf"))
}
\arguments{
\item{x}{A numeric vector.}

\item{output}{The output style of the resulting character vector. This can
either be \code{"html"} (the default), \code{"latex"}, or \code{"rtf"}.}
}
\value{
A character vector.
}
\description{
Any Markdown-formatted text in the input vector will be transformed to the
appropriate output type.
}
